% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_functions.R
\name{AIC.LDA}
\alias{AIC.LDA}
\title{Generalization of the AIC function to work on LDA topic models}
\usage{
\method{AIC}{LDA}(x, k = 2, correction = FALSE)
}
\arguments{
\item{x}{an LDA topic model}

\item{k}{penalty per df}

\item{correction}{whether or not to include the small sample size 
correction (thus generating AICc)}
}
\value{
Named (AIC or AICc) value.
}
\description{
Using the internal topicmodels calculations of likelihood and 
  df
}
\examples{
  data(rodents)
  lda_data <- dplyr::select(rodents, 
                            -c(newmoonnumber, newmoondate, nplots, ntraps))
  r_LDA <- LDATS::LDA(data = lda_data, ntopics = 2, nseeds = 2, 
                      ncores = 4)
  AIC(r_LDA[[1]])
  AIC(r_LDA[[2]])
}
