% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_on_LDA.R
\name{AIC.TS_fit}
\alias{AIC.TS_fit}
\title{Determine the AIC (deviance) value of a TS model}
\usage{
\method{AIC}{TS_fit}(object, ..., k = 2)
}
\arguments{
\item{object}{Class \code{TS_fit} object to be evaluated.}

\item{...}{Not used, simply included to maintain method compatability.}

\item{k}{Not used, simply included to maintain method compatability.}
}
\description{
Convenience function to extract the AIC (deviance) element of 
  \code{TS_fit}-class object.
}
