% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_model_selection.R
\name{aic_model}
\alias{aic_model}
\title{AIC model selection for LDA using VEM method}
\usage{
aic_model(dat, SEED, topic_min, topic_max)
}
\arguments{
\item{dat}{Table of integer data (species counts by period)}

\item{SEED}{set seed to keep LDA model runs consistent (default 2010)}

\item{topic_min}{lowest number of topics; must be >=2}

\item{topic_max}{highest number of topics}
}
\value{
data frame of number of topics (k) and aic value (aic)
}
\description{
Runs LDA using different numbers of topics and VEM method, calculates AIC values for comparison
}
\examples{
aic_values = aic_model(dat,2010,2,10)
}
