% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AICc}
\alias{AICc}
\title{Calculate AICc}
\usage{
AICc(object)
}
\arguments{
\item{object}{An object for which \code{\link{AIC}} and 
\code{\link{logLik}} have defined methods.}
}
\value{
\code{numeric} value of AICc.
}
\description{
Calculate the small sample size correction of
  \code{\link{AIC}} for the input object.
}
\examples{
  dat <- data.frame(y = rnorm(50), x = rnorm(50))
  mod <- lm(dat)
  AICc(mod)

}
