% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_checking.R
\name{argument_checking}
\alias{argument_checking}
\alias{check_class}
\alias{check_nonneg_integer}
\alias{check_nonneg_integer_matrix}
\alias{check_pos_integer}
\alias{check_control}
\alias{check_topics}
\alias{check_replicates}
\alias{check_nchangepoints}
\alias{check_document_term_table}
\alias{check_weights}
\alias{check_LDAs}
\alias{check_timename}
\alias{check_formulas}
\alias{check_document_covariate_table}
\title{Check that arguments are properly formatted for usage}
\usage{
check_class(object, eclass = "list")

check_nonneg_integer(object)

check_nonneg_integer_matrix(object)

check_pos_integer(object)

check_control(control, eclass = "list")

check_topics(topics)

check_replicates(replicates)

check_nchangepoints(nchangepoints)

check_document_term_table(document_term_table)

check_weights(weights)

check_LDAs(LDAs)

check_timename(LDAs, timename)

check_formulas(LDAs, formulas)

check_document_covariate_table(LDAs)
}
\arguments{
\item{object}{An object whose class should be checked against 
\code{eclass}.}

\item{eclass}{Expected class of \code{object} to be checked. If more
than one option is included, any are sufficient (\code{object} only
needs to be one \code{eclass}, not all).}

\item{control}{Control \code{list} to evaluate.}

\item{topics}{\code{vector} of the number of topics to evaluate for each 
model. Must be conformable to positive \code{integer} values.}

\item{replicates}{\code{integer} number of replicate starts to use for 
each value of \code{topics} in the LDAs. Must be conformable to  
positive \code{integer} values.}

\item{nchangepoints}{\code{integer}-conformable \code{vector} of the 
number of changepoints to evaluate (must be non-negative).}

\item{document_term_table}{Table of observation count data (rows: 
documents, columns: terms. May be a \code{matrix} or 
\code{data.frame} but must be conformable to a matrix of non-negative
\code{integers}.}

\item{weights}{\code{numeric} \code{vector} of the document weights to 
evaluate, or \code{TRUE} for triggering internal weighting by document
sizes.}

\item{LDAs}{\code{LDA_models} \code{list} of LDA models or singular LDA 
model (\code{LDA}) to evaluate.}

\item{timename}{Column name for the time variable to evaluate in the
\code{document_covariate_table} if provided.}

\item{formulas}{\code{vector} of the \code{\link[stats]{formula}}s 
to evaluate.}
}
\value{
an error message is thrown if the input is improper, otherwise 
  \code{NULL}.
}
\description{
Verify the class, structure, and values of inputted arguments
  to ensure proper LDATS modeling. \cr \cr
  \code{check_class} is a general class-verifier. \cr \cr
  \code{check_nonneg_integer} is a specified checking function for
    \code{numeric} values that must be integer-conformable and non-negative
    (0 is acceptable). \cr \cr
  \code{check_nonneg_integer_matrix} is a specified checking function for
    tables of \code{numeric} values that must be integer-conformable and 
    non-negative (0 is acceptable) and that must be conformable to 
    a \code{matrix}. \cr \cr
  \code{check_pos_integer} is a specified checking function for
    \code{numeric} values that must be integer-conformable and positive
    (0 is not acceptable). \cr \cr
  \code{check_control} verifies that a control \code{list} is a \code{list}
    \cr \cr
  \code{check_topics} ensures that the vector of numbers of topics is 
    positive integer-conformable. \cr \cr
  \code{check_replicates} ensures that the number of replicates is 
    positive integer-conformable. \cr \cr
  \code{check_nchangepoints} ensures that the of change points is 
    positive integer-conformable. \cr \cr
  \code{check_document_term_table} ensures that the table of document
    term counts is conformable to a \code{matrix} of positive integers. 
    \cr \cr
  \code{check_LDAs} verifies that the argument is either a
    \code{LDA} or \code{LDA_set} \code{list}. \cr \cr
  \code{check_document_covariate_table} check that the table of
   document-level covariates in the \code{LDAs} data is 
   conformable to a data frame and of the right size (correct number of 
   documents) for the document-topic output from the LDA models. \cr \cr
  \code{check_weights} ensures that the vector of document weights is 
   \code{numeric} and positive and inform the user if the average weight
   isn't 1.\cr \cr
  \code{check_timename} checks that the vector of time values is included 
    in the \code{document_covariate_table} and that it is either a 
    \code{integer}-conformable or a \code{Date}. 
    If it is a \code{Date}, the input is converted to an 
    \code{integer}, resulting in the timestep being 1 day, which is often
    not desired behavior. \cr \cr
  \code{check_formulas} verifies that the input contains only 
  \code{\link[stats]{formula}}s and that the response and predictor 
    variables are all included in \code{LDAs} data sets. \cr \cr
  \code{check_nchangepoints} checks that the \code{vector} of numbers of 
    changepoints is conformable to non-negative \code{integers}.\cr \cr
}
