% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_plots.R
\name{autocorr_plot}
\alias{autocorr_plot}
\title{Produce the autocorrelation panel for the TS diagnostic plot of
  a parameter}
\usage{
autocorr_plot(x)
}
\arguments{
\item{x}{Vector of parameter values drawn from the posterior distribution,
indexed to the iteration by the order of the vector.}
}
\description{
Produce a vanilla ACF plot using \code{acf} for the parameter
  of interest (rho or eta).
}
