% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_functions.R
\name{batch_LDA}
\alias{batch_LDA}
\title{Run a batch of LDA models in parallel}
\usage{
batch_LDA(data, ntopics, nseeds, method, sort, sortby, parallel,
  ncores = NULL)
}
\arguments{
\item{data}{Table of integer data (species counts by period)}

\item{ntopics}{set of topics to evaluate}

\item{nseeds}{number of seeds to use}

\item{method}{what LDA method to use (currently only support for "VEM")}

\item{sort}{Logical of whether or not to sort the output}

\item{sortby}{Which metric to sortby ("aic" or "aicc")}

\item{parallel}{Logical of whether or not to run the models in parallel}

\item{ncores}{Integer number of cores to use}
}
\value{
List of [1] list of models and [2] table model summaries
}
\description{
Runs each of the number of input topics for the number of seeds
}
\examples{
aic_values = batch_LDA(data, 2, 1, "VEM", TRUE, "aicc", FALSE)
}
