% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{BTS}
\alias{BTS}
\title{Bayesian Time Series analysis of a topic model classification}
\usage{
BTS(model, formula, data)
}
\arguments{
\item{model}{fitted LDA model}

\item{formula}{formula}

\item{data}{}
}
\description{
Bayesian Time Series analysis of a topic model classification
}
