% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA-distance.R
\name{calculate_LDA_distance}
\alias{calculate_LDA_distance}
\title{Calculates multiple LDA models using different seeds}
\usage{
calculate_LDA_distance(ldas, seeds)
}
\arguments{
\item{ldas}{list of results from LDA model}

\item{seeds}{vector of seeds for which LDA models will be fit and results compared}
}
\value{
list including: seed associated with highest log likelihood
                        mean distance between best model and all others
                        max distuance between best model and worst
}
\description{
Calculates multiple LDA models using different seeds
}
