% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{ccmat}
\alias{ccmat}
\title{Measure the cross-correlation among change points}
\usage{
ccmat(cps, lag = 0)
}
\arguments{
\item{cps}{change point estimates}

\item{lag}{lag to be used in the correlation estimation}
}
\value{
matrix of among-changepoint correlations
}
\description{
Measure the cross-correlation among change points
}
