% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepointmodel.r
\name{changepoint_model}
\alias{changepoint_model}
\title{Model to locate given number of change points in an LDA model output}
\usage{
changepoint_model(ldamodel, x, n_changepoints, N_temps = 6, maxit = 1e+05,
  penultimate_temp = 2^6, k = 0, weights)
}
\arguments{
\item{ldamodel}{output object from LDA model using VEM}

\item{x}{x}

\item{n_changepoints}{integer number of changepoints to look for}

\item{maxit}{maximum iterations}

\item{penultimate_temp}{penultimate_temp}

\item{k}{the exponent controlling the temperature sequence: 0 implies geometric sequence, 
1 implies squaring before exponentiating. Use larger values if the cooler chains aren't swapping enough.}
}
\value{

}
\description{
Model to locate given number of change points in an LDA model output
}
\author{
Dave Harris
}
