% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_TS.R
\name{check_chunks}
\alias{check_chunks}
\title{Verify the chunks of a multinomial time series model}
\usage{
check_chunks(data, changepoints)
}
\arguments{
\item{data}{Class \code{data.frame} object including the predictor and 
response variables.}

\item{changepoints}{Numeric vector indicating locations of the change 
points.}
}
\value{
Logical indicator of the check passing \code{TRUE} or failing
  \code{FALSE}.
}
\description{
Check to verify that a time series can be broken into a set 
  of chunks based on input changepoints.
}
