% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_control}
\alias{check_control}
\title{Check that a control list is proper}
\usage{
check_control(control, eclass = "TS_controls")
}
\arguments{
\item{control}{Control list to evaluate.}

\item{eclass}{Expected class of the list to be evaluated.}
}
\description{
Check that a list of controls is of the right class.
}
