% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_on_LDA.R
\name{check_formulas}
\alias{check_formulas}
\title{Verify that formulas vector is proper}
\usage{
check_formulas(formulas, document_covariate_table)
}
\arguments{
\item{formulas}{Vector of the formulas to evaluate.}

\item{document_covariate_table}{Document covariate table used to evaluate
the availability of the data required by the formula inputs.}
}
\value{
Nothing.
}
\description{
Verify that the vector of formulas is actually formatted
  as a vector formula objects and that the predictor variables are all 
  included in the document covariate table.
}
