% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{check_MV}
\alias{check_MV}
\title{Verify that MV matrix is proper}
\usage{
check_MV(MV)
}
\arguments{
\item{MV}{Matrix of observation count data (rows: documents (\code{M}), 
columns: terms (\code{V})).}
}
\value{
Nothing.
}
\description{
Verify that the matrix of observations is conformable to
  a matrix of integers
}
