% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_on_LDA.R
\name{check_nchangepoints}
\alias{check_nchangepoints}
\title{Verify that nchangepoints vector is proper}
\usage{
check_nchangepoints(nchangepoints)
}
\arguments{
\item{nchangepoints}{Vector of the number of changepoints to evaluate.}
}
\value{
Nothing.
}
\description{
Verify that the vector of numbers of changepoints is 
  conformable to integers greater than 1.
}
