% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_checking.R
\name{check_nreps}
\alias{check_nreps}
\title{Check that nseeds value or vector is proper}
\usage{
check_nreps(nreps)
}
\arguments{
\item{nseeds}{\code{integer} number of seeds (replicate starts) to use for 
each value of \code{topics} in the LDAs. Must be conformable to a  
positive \code{integer} value.}
}
\value{
an error message is thrown if the input is improper, otherwise 
  \code{NULL}.
}
\description{
Check that the vector of numbers of seeds is conformable to
  integers greater than 0.
}
\examples{
 check_nseeds(1)
 check_nseeds(2)

}
