% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_seeds}
\alias{check_seeds}
\title{Check that nseeds value or seeds vector is proper}
\usage{
check_seeds(nseeds)
}
\arguments{
\item{nseeds}{\code{integer} number of seeds (replicate starts) to use for
each value of \code{topics} in the LDAs. Must be conformable to a
positive \code{integer} value.}
}
\value{
an error message is thrown if the input is improper, otherwise
\code{NULL}.
}
\description{
Check that the vector of numbers of seeds is conformable to
integers greater than 0.
}
\examples{
 check_seeds(1)
 check_seeds(2)

}
