% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_on_LDA.R
\name{check_TS_on_LDA}
\alias{check_TS_on_LDA}
\title{Check all of the inputs to TS_on_LDA}
\usage{
check_TS_on_LDA(call)
}
\arguments{
\item{call}{Class \code{class} list, holding all of the arguments passed
to \code{\link{TS_on_LDA}}.}
}
\value{
Nothing.
}
\description{
Verify the inputs to (\code{\link{TS_on_LDA}}).
}
