% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_on_LDA.R
\name{check_weights}
\alias{check_weights}
\title{Verify that weights vector is proper}
\usage{
check_weights(weights)
}
\arguments{
\item{weights}{Vector of the document weights to evaluate.}
}
\value{
Nothing.
}
\description{
Verify that the vector of document weights is numeric
  and inform the user if weights are outside the optimal range: 
  \eqn{(0,1]}.
}
