% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_figure_scripts.R
\name{chpoint_histogram}
\alias{chpoint_histogram}
\title{Plot histogram of changepoint locations -- WIP}
\usage{
chpoint_histogram(results, year_continuous)
}
\arguments{
\item{results}{results object from changepoint_model function}

\item{year_continuous}{vector of dates/xaxis units}
}
\value{
ggplot object
}
\description{
Plot histogram of changepoint locations -- WIP
}
