% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_figure_scripts.R
\name{community_composition}
\alias{community_composition}
\title{Make table of species composition of topics}
\usage{
community_composition(ldamodel)
}
\arguments{
\item{ldamodel}{object of class LDA_VEM created by the function LDA in topicmodels package}
}
\value{
table of species composition of the topics in ldamodel, 3 decimal places
}
\description{
Make table of species composition of topics
}
\examples{
community_composition(ldamodel)
}
