% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{count_trips}
\alias{count_trips}
\title{Count trips of the ptMCMC particles}
\usage{
count_trips(ids)
}
\arguments{
\item{ids}{Matrix of identifiers of the particles in each chain for each
iteration of the ptMCMC algorithm.}
}
\value{
List of [1] vector of counts of trips and [2] vector of rates of 
  trips by temperature.
}
\description{
Count the full trips (from one extreme temperature chain to
  the other and back again) for each of the ptMCMC particles, as identified
  by their id on initialization.
}
