% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsampling_functions.R
\name{cp_hists_file}
\alias{cp_hists_file}
\title{Save out a multi-panel histogram of the change points for a dataset}
\usage{
cp_hists_file(dd = NULL, cpms = NULL, ntopics = NULL, niter = NULL,
  nn = NULL)
}
\arguments{
\item{dd}{data}

\item{cpms}{change point models}

\item{ntopics}{number of topics in the LDA used}

\item{niter}{number of iterations in the model}

\item{nn}{name}
}
\value{
Figure saved out
}
\description{
Save out a multi-panel histogram of the change points for a dataset
}
