% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodent_data_for_LDA.r
\name{create_rodent_table}
\alias{create_rodent_table}
\title{Create rodent species table}
\usage{
create_rodent_table(period_first = 1, period_last = 436,
  selected_plots = c(2, 4, 8, 11, 12, 14, 17, 22),
  selected_species = c("BA", "DM", "DO", "DS", "NA", "OL", "OT", "PB", "PE",
  "PF", "PH", "PI", "PL", "PM", "PP", "RF", "RM", "RO", "SF", "SH", "SO"))
}
\arguments{
\item{period_first}{first trapping period desired}

\item{period_last}{last trapping period desired}

\item{selected_plots}{vector of plot numbers to be included}

\item{selected_species}{vector of species codes to be included}
}
\value{
Table of species counts per period
}
\description{
Processes rodent capture data so it can be used for LDA analysis
}
\examples{
r_table = create_rodent_table())

}
