% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sim_data.R
\name{create_sim_data_2topic}
\alias{create_sim_data_2topic}
\title{create simulated data for demonstrating LDA analysis
   2 topics}
\usage{
create_sim_data_2topic(nspecies = 24, tsteps = 400)
}
\arguments{
\item{nspecies}{= number of species in all topic groups}

\item{tsteps}{= number of [monthly] time steps}
}
\value{
beta = matrix of species composition of the groups
   gamma = matrix of topic composition over time
           3 simulations of gamma: uniform, slow transition, and fast transition
}
\description{
create simulated data for demonstrating LDA analysis
   2 topics
}
