% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sim_data.R
\name{create_sim_data_2topic_nonuniform}
\alias{create_sim_data_2topic_nonuniform}
\title{variation: create simulated data for demonstrating LDA analysis
2 topics, nonuniform distribution of species in two community-types}
\usage{
create_sim_data_2topic_nonuniform(tsteps = 400)
}
\arguments{
\item{tsteps}{= number of [monthly] time steps}
}
\value{
beta = matrix of species composition of the groups
   gamma = matrix of topic composition over time
           3 simulations of gamma: uniform, slow transition, and fast transition
}
\description{
variation: create simulated data for demonstrating LDA analysis
2 topics, nonuniform distribution of species in two community-types
}
