% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sim_data.R
\name{create_sim_data_3topic}
\alias{create_sim_data_3topic}
\title{create simulated data for demonstrating LDA analysis
   3 topics}
\usage{
create_sim_data_3topic(nspecies = 24, tsteps = 400)
}
\arguments{
\item{nspecies}{= number of species in all topic groups}

\item{tsteps}{= number of [monthly] time steps}
}
\description{
create simulated data for demonstrating LDA analysis
   3 topics
}
