% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptMCMC.R
\name{diagnose_ptMCMC}
\alias{diagnose_ptMCMC}
\title{Summarize the ptMCMC diagnostics}
\usage{
diagnose_ptMCMC(rho_dist)
}
\arguments{
\item{rho_dist}{List of saved data objects from the ptMCMC estimation of
changepoint locations (unless \code{nchangepoints} is 0, then 
\code{NULL}).}
}
\value{
List of [1] step acceptance rates, [2] swap acceptance rates, [3]
  trip counts, and [4] trip rates.
}
\description{
Summarize the acceptance rates and trip metrics from 
  the saved output from the ptMCMC.
}
