% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDATS_functions.R
\name{doc_weights}
\alias{doc_weights}
\title{Calculate document weights}
\usage{
doc_weights(document_term_matrix)
}
\arguments{
\item{document_term_matrix}{matrix of documents (rows) by terms (columns)}
}
\value{
vector of weights, one for each document, with the largest sample
  receiving a weight of 1.0
}
\description{
Simple calculation based on the maximum (max value = 1)
}
