% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{document_weights}
\alias{document_weights}
\title{Calculate document weights on maximum number of words}
\usage{
document_weights(document_term_table)
}
\arguments{
\item{document_term_table}{Table of observation count data (rows: 
documents (\eqn{M}), columns: terms (\eqn{V})). May be a class 
\code{matrix} or \code{data.frame} but must be conformable to
a code of integers. This table is a document-level summary of the data 
noted as \eqn{w} (the word-level topic identity) in the math description.}
}
\value{
Vector of weights, one for each document, with the largest sample
  receiving a weight of 1.0.
}
\description{
Simple calculation of document weights based on the maximum 
  number of words in a document within the corpus (max value = 1).
}
