% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsampling_functions.R
\name{downsample}
\alias{downsample}
\title{Downsample a time series}
\usage{
downsample(data = NULL, frequency = NULL, span = NULL,
  selection_random = FALSE, selection_order = 1)
}
\arguments{
\item{data}{data frame with a column named "dates"}

\item{frequency}{number of samples to be retained within each year}

\item{span}{date range to include}

\item{selection_random}{logical indicating if the downsampling is random 
(TRUE) or ordered (FALSE)}

\item{selection_order}{}
}
\value{
Table of species counts per period
}
\description{
Downsample a time series
}
\author{
Juniper Simonis
}
