% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{est_changepts}
\alias{est_changepts}
\title{Use ptMCMC to estimate the distribution of changepoint locations}
\usage{
est_changepts(data, formula, nchangepoints, weights, control)
}
\arguments{
\item{data}{Class \code{data.frame} object including [1] the time variable
(indicated in \code{control}), [2] the predictor variables (required by
\code{formula}) and [3], the multinomial response variable (indicated
in \code{formula}). Note that the response variables should be formatted
as \code{<name>.1}, \code{<name>.2}, etc. (where \code{<name>} is the
actual name of the response variable, such as \code{gamma} for a standard
TS analysis of LDA output.}

\item{formula}{\code{formula} describing the continuous change. Any 
predictor variable included must also be a column in the
\code{data}.  Any (multinomial) response variable must also be a set of
columns in \code{data}.}

\item{nchangepoints}{Integer corresponding to the number of 
change points to include in the model. 0 is a valid input (corresponding
to no change points, so a singular time series model), and the current 
implementation can reasonably include up to 6 change points. The 
number of change points is used to dictate the segementation of the data  
for each continuous model and each LDA model.}

\item{weights}{Optional class \code{numeric} vector of weights for each 
document. Corresponds to the vector \strong{\eqn{v}} in the math 
description.}

\item{control}{Class \code{TS_controls} list, holding control parameters
for the Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls, generated by 
\code{\link{TS_controls_list}}.}
}
\value{
List of saved data objects from the ptMCMC estimation of
  changepoint locations (unless \code{nchangepoints} is 0, then 
  \code{NULL} is returned).
}
\description{
This is the main interface function for the Time Series 
  analyses.
}
