% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{est_regressors}
\alias{est_regressors}
\title{Estimate the distribution of regressors, unconditional on the
  changepoint locations}
\usage{
est_regressors(rho_dist, data, formula, weights, control)
}
\arguments{
\item{rho_dist}{List of saved data objects from the ptMCMC estimation of
changepoint locations (unless \code{nchangepoints} is 0, then 
\code{NULL}).}

\item{data}{Class \code{data.frame} object including [1] the time variable
(indicated in \code{control}), [2] the predictor variables (required by
\code{formula}) and [3], the multinomial response variable (indicated
in \code{formula}).}

\item{formula}{\code{formula} describing the continuous change. Any 
predictor variable included must also be a column in the
\code{data}.  Any (multinomial) response variable must also be a set of
columns in \code{data}.}

\item{weights}{Optional class \code{numeric} vector of weights for each 
document. Corresponds to the vector \strong{\eqn{v}} in the math 
description.}

\item{control}{Class \code{TS_controls} list, holding control parameters
for the Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls, generated by 
\code{\link{TS_controls_list}}.}
}
\value{
Matrix of draws (rows) from the marginal posteriors of the 
  coefficients across the segments (columns).
}
\description{
This function uses the estimated changepoint locations to 
  estimate the marginal distribution of the regression coefficients for
  the chunk-level models, unconditional on the changepoint locations.
}
