% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{find_changepoint_location}
\alias{find_changepoint_location}
\title{Find changepoint locations}
\usage{
find_changepoint_location(results)
}
\arguments{
\item{results}{results object output from changepoint_model}

\item{return}{vector of changepoint locations}
}
\description{
Find changepoint locations
}
\author{
Erica Christensen
}
