% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{fit_chunk_non_memoized}
\alias{fit_chunk_non_memoized}
\title{Fit a model to dates in (start,end] and return the log likelihood.}
\usage{
fit_chunk_non_memoized(ldamodel, x, start, end, make_plot = FALSE, weights,
  ...)
}
\arguments{
\item{ldamodel}{ldamodel}

\item{x}{x}

\item{start}{start}

\item{end}{end}

\item{make_plot}{make_plot}

\item{weights}{weights}
}
\description{
Fit a model to dates in (start,end] and return the log likelihood.
}
\examples{
NA
}
