% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_changepoint_functions.R
\name{fit_section}
\alias{fit_section}
\title{Fit a model to dates in (start,end] and return data frame for plotting}
\usage{
fit_section(ldamodel, x, start, end, weights, ...)
}
\arguments{
\item{ldamodel}{output of LDA()}

\item{x}{same x used in changepoint_model()}

\item{start}{value: start of the section to be fit}

\item{end}{value: end of the section to be fit}

\item{weights}{same weights used in changepoint_model()}
}
\description{
Fit a model to dates in (start,end] and return data frame for plotting
}
