% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_changepoint_functions.R
\name{get_ll_non_memoized}
\alias{get_ll_non_memoized}
\title{Get the log-likelihood associated with a set of breakpoints}
\usage{
get_ll_non_memoized(ldamodel, x, changepoints, make_plot = FALSE, weights,
  ...)
}
\arguments{
\item{ldamodel}{ldamodel}

\item{x}{x}

\item{changepoints}{changepoints}

\item{make_plot}{make_plot}

\item{weights}{weights}
}
\description{
Get the log-likelihood associated with a set of breakpoints
}
