% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepointmodel.r
\name{get_ll_non_memoized_plot}
\alias{get_ll_non_memoized_plot}
\title{adapted get_ll_non_memoized for plotting purposes}
\usage{
get_ll_non_memoized_plot(ldamodel, x, changepoints, weights, ...)
}
\arguments{
\item{ldamodel}{output object from LDA()}

\item{x}{same x used in changepoint_model()}

\item{changepoints}{vector of locations of changepoints found by changepoint_model()}

\item{weights}{same weights used in changepoint_model()}
}
\description{
adapted get_ll_non_memoized for plotting purposes
}
