% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA-distance.R
\name{Hellinger}
\alias{Hellinger}
\title{Compute hellinger distance}
\usage{
Hellinger(a, b)
}
\arguments{
\item{a, b}{numeric vectors}
}
\description{
Hellinger distance is Euclidean distance between square roots, divided by
sqrt(2)
}
