% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{identity_LDA}
\alias{identity_LDA}
\title{Identity Linguistic Decomposition Analysis}
\usage{
identity_LDA(LDA)
}
\arguments{
\item{LDA}{A prepared (via \code{\link{prep_LDA_models}} LDA model
\code{list}.}
}
\value{
\code{LDA} \code{list} with most components as placeholders.
}
\description{
This function acts as an "identity" model, wherein the 
  output is functionally the input. This allows for "single-topic" models
  that do not actually decompose the data to be included in the model set.
}
