% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{iftrue}
\alias{iftrue}
\title{Replace if TRUE}
\usage{
iftrue(x = TRUE, alt = NULL)
}
\arguments{
\item{x}{Focal input.}

\item{alt}{Alternative value.}
}
\value{
\code{x} if not \code{TRUE}, \code{alt} otherwise.
}
\description{
If the focal input is \code{TRUE}, replace it with 
  alternative.
}
