% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilr_TS.R
\name{ilr_TS_control}
\alias{ilr_TS_control}
\title{Create the controls list for the ILR Time Series model}
\usage{
ilr_TS_control(quiet = FALSE, ...)
}
\arguments{
\item{quiet}{\code{logical} indicator of whether the model should run 
quietly (if \code{FALSE}, a progress bar and notifications are printed).}

\item{...}{Not passed along to the output, rather included to allow for
automated removal of unneeded controls.}
}
\value{
\code{list}, with named elements corresponding to the arguments.
}
\description{
This function provides a simple creation and definition of a
  list used to control the time series model fit occurring within 
  \code{\link{ilr_TS}}.
}
