% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{initialize_cpts}
\alias{initialize_cpts}
\title{Initialize the changepoint matrix used in the ptMCMC algorithm}
\usage{
initialize_cpts(data, formula, nchangepoints, weights,
  control = TS_controls_list())
}
\arguments{
\item{data}{Class \code{data.frame} object including [1] the time variable
(indicated in \code{control}), [2] the predictor variables (required by
\code{formula}) and [3], the multinomial response variable (indicated
in \code{formula}). Note that the response variables should be formatted
as \code{<name>.1}, \code{<name>.2}, etc. (where \code{<name>} is the
actual name of the response variable, such as \code{gamma} for a standard
TS analysis of LDA output.}

\item{formula}{\code{formula} describing the continuous change. Any 
predictor variable included must also be a column in the
\code{data}.  Any (multinomial) response variable must also be a set of
columns in \code{data}.}

\item{nchangepoints}{Integer corresponding to the number of 
change points to include in the model. 0 is a valid input (corresponding
to no change points, so a singular time series model), and the current 
implementation can reasonably include up to 6 change points. The 
number of change points is used to dictate the segementation of the data  
for each continuous model and each LDA model.}

\item{weights}{Optional class \code{numeric} vector of weights for each 
document. Corresponds to the vector \strong{\eqn{v}} in the math 
description.}

\item{control}{Class \code{TS_controls} list, holding control parameters
for the Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls, generated by 
\code{\link{TS_controls_list}}.}
}
\value{
List of [1] matrix of change points (rows) for each temperature 
  (columns) and [2] vector of log-likelihood values for each of the chains.
}
\description{
Each of the chains is initialized using a draw from the 
  available times (i.e. assuming a uniform prior), the best fit (by 
  likelihood) draw is put in the focal chain with each subsequently worse
  fit placed into the subsequently hotter chain.
}
