% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{initialize_pbar}
\alias{initialize_pbar}
\alias{tick_pbar}
\title{Initialize and tick through the progress bar}
\usage{
initialize_pbar(control = TS_controls_list())

tick_pbar(pbar, control = TS_controls_list())
}
\arguments{
\item{control}{Class \code{TS_controls} list, holding control parameters
for the Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls, generated by 
\code{\link{TS_controls_list}}. Of use here is \code{quiet} which is a
a \code{logical} indicator of whether there should be information 
(i.e. the progress bar) printed during the run or not. Default is 
\code{TRUE}.}

\item{pbar}{The progress bar object returned from \code{initialize_pbar}.}
}
\value{
\code{initialize_pbar}: the initialized progress bar object.

\code{tick_pbar}: the ticked-forward \code{pbar}.
}
\description{
\code{initialize_pbar} creates and \code{tick_pbar} steps
  through the progress bar (if desired) in \code{\link{TS}}
}
