% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA}
\alias{LDA}
\alias{check_LDA_inputs}
\title{Run a set of Latent Dirichlet Allocation models}
\usage{
LDA(data, topics = 2:3, nreps = 2, control = list())

check_LDA_inputs(data, topics, nreps, control)
}
\arguments{
\item{topics}{Vector of the number of topics to evaluate for each model.
Must be conformable to \code{integer} values. REDUCED DIMENSIONALITY}

\item{nreps}{Number of replicate starts to use for each 
value of \code{topics}. Must be conformable to \code{integer} value.}

\item{control}{A \code{list} of parameters to control the running and 
selecting of LDA models. Values not input assume default values set 
by \code{\link{LDA_set_control}}. Values for running the LDAs replace 
defaults in (\code{LDAcontol}, see \code{\link[topicmodels]{LDA}} (but if
 \code{seed} is given, it will be overwritten; use \code{iseed} instead).}

\item{document_term_table}{Table of observation count data (rows: 
documents, columns: terms. May be a class \code{matrix} or 
\code{data.frame} but must be conformable to a matrix of integers,
as verified by \code{\link{check_document_term_table}}.}

\item{object}{A \code{LDA_VEM}-class object.}

\item{...}{Not used, simply included to maintain method compatibility.}
}
\value{
\code{LDA_set}: \code{list} (class: \code{LDA_set}) of LDA models 
  (class: \code{LDA_VEM}).
  \code{check_LDA_set_inputs}: an error message is thrown if any input is 
  improper, otherwise \code{NULL}.

Log likelihood of the model \code{logLik}, also with \code{df}
  (degrees of freedom) and \code{nobs} (number of observations) values.
}
\description{
For a given dataset consisting of counts of words across 
  multiple documents in a corpus, conduct multiple Latent Dirichlet 
  Allocation (LDA) models (using the Variational Expectation 
  Maximization (VEM) algorithm; Blei \emph{et al.} 2003) to account for [1]  
  uncertainty in the number of latent topics and [2] the impact of initial
  values in the estimation procedure. \cr \cr
  \code{LDA_set} is a list wrapper of \code{\link[topicmodels]{LDA}}
  in the \code{topicmodels} package (Grun and Hornik 2011). \cr \cr
  \code{check_LDA_set_inputs} checks that all of the inputs 
  are proper for \code{LDA_set} (that the table of observations is 
  conformable to a matrix of integers, the number of topics is an integer, 
  the number of seeds is an integer and the controls list is proper).

Imported but updated calculations from topicmodels package, as
  applied to Latent Dirichlet Allocation fit with Variational Expectation 
  Maximization via \code{\link[topicmodels]{LDA}}.
}
\details{
The number of degrees of freedom is 1 (for alpha) plus the number
  of entries in the document-topic matrix. The number of observations is 
  the number of entries in the document-term matrix.
}
\examples{
  data(rodents)
  lda_data <- rodents$document_term_table
  r_LDA <- LDA(lda_data, topics = 2, nseeds = 2)                         

  data(rodents)
  lda_data <- rodents$document_term_table
  r_LDA <- LDA_set(lda_data, topics = 2)   
  logLik(r_LDA[[1]])

}
\references{
Blei, D. M., A. Y. Ng, and M. I. Jordan. 2003. Latent Dirichlet
  Allocation. \emph{Journal of Machine Learning Research} 
  \strong{3}:993-1022.
  \href{http://jmlr.csail.mit.edu/papers/v3/blei03a.html}{link}.

  Grun B. and K. Hornik. 2011. topicmodels: An R Package for Fitting Topic
  Models. \emph{Journal of Statistical Software} \strong{40}:13.
  \href{https://www.jstatsoft.org/article/view/v040i13}{link}.

Buntine, W. 2002. Variational extensions to EM and multinomial PCA. 
  \emph{European Conference on Machine Learning, Lecture Notes in Computer 
  Science} \strong{2430}:23-34. \href{https://bit.ly/327sltH}{link}.

  Grun B. and K. Hornik. 2011. topicmodels: An R Package for Fitting Topic
  Models. \emph{Journal of Statistical Software} \strong{40}:13.
  \href{https://www.jstatsoft.org/article/view/v040i13}{link}.

  Hoffman, M. D., D. M. Blei, and F. Bach. 2010. Online learning for 
  latent Dirichlet allocation. \emph{Advances in Neural Information 
  Processing Systems} \strong{23}:856-864.
  \href{https://bit.ly/2LEr5sb}{link}.
}
