% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_analysis.R
\name{LDA_analysis_gibbs}
\alias{LDA_analysis_gibbs}
\title{Run through whole pipeline of LDA analysis, Gibbs method --- WIP}
\usage{
LDA_analysis_gibbs(dat, ngibbs, test_topics, dates)
}
\arguments{
\item{dat}{Table of integer data (species counts by time step)}

\item{ngibbs}{number of iterations for gibbs sampler -- must be greater than 200 (default 1000)}

\item{test_topics}{numbers of topics to test, of the form (topic_min,topic_max)}

\item{dates}{vector of dates that correspond to time steps from 'dat' -- for plotting purposes}
}
\description{
Run through whole pipeline of LDA analysis, Gibbs method --- WIP
}
\examples{
LDA_analysis_gibbs(dat,200,c(2,3),dates)
}
