% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_analysis.R
\name{LDA_analysis_VEM}
\alias{LDA_analysis_VEM}
\title{Run through whole pipeline of LDA analysis, VEM method}
\usage{
LDA_analysis_VEM(dat, SEED, test_topics)
}
\arguments{
\item{dat}{Table of integer data (species counts by time step)}

\item{SEED}{set seed to keep LDA model runs consistent (default 2010)}

\item{test_topics}{numbers of topics to test, of the form (topic_min, 
topic_max)}

\item{dates}{vector of dates that correspond to time steps from 'dat' -- 
for plotting purposes}

\item{n_chpoints}{number of change points the changepoint model should 
look for}

\item{maxit}{max iterations for changepoint model (more=slower)}
}
\description{
Run through whole pipeline of LDA analysis, VEM method
}
\examples{
LDA_analysis(dat, 2010, c(2, 3), dates, 2)
}
