% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA_control}
\alias{LDA_control}
\title{Create control list for set of LDA models}
\usage{
LDA_control(LDA_function = topicmodels_LDA, LDA_args = list(method =
  "VEM", seeded = TRUE), measurer_function = AIC,
  measurer_args = list(), selector_function = min,
  selector_args = list(), soften = TRUE, quiet = FALSE)
}
\arguments{
\item{LDA_function}{\code{function} used for the LDA model. Defaults to
classical LDA as coded via \code{topicmodels}' 
\code{\link[topicmodels]{LDA}}.}

\item{quiet}{\code{logical} indicator of whether the model should run 
quietly.}

\item{measurer, selector}{Function names for use in evaluation of the LDA
models. \code{measurer} is used to create a value for each model
and \code{selector} operates on the values to choose the model(s) to 
pass on.}

\item{args}{\code{list} of arguments to \code{LDA_function}.}

\item{...}{Arguments to be passed to \code{LDA_function} as a 
\code{control} input.}
}
\value{
\code{list} for controlling the LDA model fit.
}
\description{
This function provides a simple creation and definition of 
  the list used to control the set of LDA models, regardless of how they
  are implemented.
}
\examples{
  LDA_set_control()

}
