% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA_controls_list}
\alias{LDA_controls_list}
\title{Create control list for LDA model

  list used to control the set of LDA models. It is set up to be easy to 
  work with the existing control capacity of 
  \code{\link[topicmodels]{LDA}}.}
\usage{
LDA_controls_list(quiet = FALSE, measurer = AIC, selector = min, ...)
}
\arguments{
\item{quiet}{\code{logical} indicator of whether the model should run 
quietly.}

\item{measurer, selector}{Function names for use in evaluation of the LDA
models. \code{measurer} is used to create a value for each model
and \code{selector} operates on the values to choose the model(s) to 
pass on.}

\item{...}{Additional arguments to be passed to 
\code{\link[topicmodels]{LDA}} as a \code{control} input.}
}
\value{
Class \code{LDA_controls} list for controlling the LDA model fit.
}
\description{
Create control list for LDA model

  list used to control the set of LDA models. It is set up to be easy to 
  work with the existing control capacity of 
  \code{\link[topicmodels]{LDA}}.
}
