% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA_msg}
\alias{LDA_msg}
\title{Create the model-running-message for an LDA}
\usage{
LDA_msg(topics, rep, quiet = FALSE)
}
\arguments{
\item{LDA_topics}{\code{integer} value corresponding to the number of 
topics in the model.}

\item{LDA_rep}{\code{integer} value corresponding to the replicate number 
for the run of the model.}

\item{control}{Class \code{LDA_controls} list of control parameters to be
used in \code{LDA} (note that "seed" will be overwritten).}
}
\description{
Produce and print the message for a given LDA model.
}
\examples{
  LDA_msg(mod_topics = 4, mod_seeds = 2)

}
