% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_plots.R
\name{LDA_plot_bottom_panel}
\alias{LDA_plot_bottom_panel}
\title{Plot the topic proportions of an LDATS LDA model over time}
\usage{
LDA_plot_bottom_panel(x, xtime, xname, cols)
}
\arguments{
\item{x}{Object of class \code{LDA}.}

\item{xtime}{Optional x values used to plot the topic proportions according
to a specific time value (rather than simply the order of observations).}

\item{xname}{Optional name for the x values used in plotting the topic
proportions (otherwise defaults to "Document").}

\item{cols}{Colors to be used to plot the topics.}
}
\value{
Bottom (time series) panel of an LDATS LDA plot.
}
\description{
Create an LDATS LDA summary plot bottom panel showing the 
  topic proportions over time/documents.
}
