% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_plots.R
\name{LDA_plot_top_panel}
\alias{LDA_plot_top_panel}
\title{Plot the topic proportions of an LDATS LDA model for each of the
  words}
\usage{
LDA_plot_top_panel(x, cols)
}
\arguments{
\item{x}{Object of class \code{LDA}.}

\item{cols}{Colors to be used to plot the topics.}
}
\value{
Top (bar plot) panel of an LDATS LDA plot.
}
\description{
Create an LDATS LDA summary plot top panel showing the 
  topic proportions word-by-word.
}
