% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_functions.R
\name{LDA_select}
\alias{LDA_select}
\title{Selection of best LDA model(s) for use in time series}
\usage{
LDA_select(lda_models = NULL, LDA_eval = quote(AIC),
  LDA_selector = quote(min), ...)
}
\arguments{
\item{lda_models}{LDA model output}

\item{LDA_eval}{function name for evaluation of the LDA models}

\item{LDA_selector}{function name for selecting the LDA models}

\item{...}{additional arguments to be passed to subfunctions}
}
\value{
(currently) the output of the selected LDA model
}
