% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA_set}
\alias{LDA_set}
\title{Run a set of Latent Dirichlet Allocation models}
\usage{
LDA_set(MV, topics = 2, nseeds = 1, control = NULL)
}
\arguments{
\item{MV}{Matrix of observation count data (rows: documents (\code{M}), 
columns: terms (\code{V})).}

\item{topics}{Vector of the number of topics to evaluate.}

\item{nseeds}{Integer number of seeds (replicate starts) to use for each 
value of \code{ntopics}.}

\item{control}{Named list of control parameters to be used in 
\code{\link[topicmodels]{LDA}} (note that "seed" will be overwritten).}
}
\value{
List (class: "\code{LDA_list}") of LDA models (class: 
  "\code{LDA}").
}
\description{
For a given dataset (counts of words across several 
  documents), conduct multiple Latent Dirichlet Allocation (LDA) models
  (using the Variational Expectation Maximization (VEM) algorithm) to 
  account for [1] uncertainty in the number of latent topics and
  [2] the impact of intial values in the estimation procedure
}
\examples{
\dontrun{
  data(rodents)
  lda_data <- dplyr::select(rodents, -c(newmoon, date, plots, traps))
  r_LDA <- LDA_set(MV = lda_data, topics = 2, nseeds = 2)                         
}

}
