% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_TS.R
\name{LDA_TS}
\alias{LDA_TS}
\title{Run a set of Linguistic Decomposition Analysis models coupled to
  Bayesian Time Series models}
\usage{
LDA_TS(data, topics = 2, reps = 1, formulas = ~1,
  nchangepoints = 0, timename = "time", weights = TRUE,
  control = list())
}
\arguments{
\item{data}{A document term table, \code{list} of document term and 
covariate tables, a list of training and test sets of the two tables,
or a list of multiple replicate splits of training and test sets of
the two tables. \cr \cr
See \code{\link{conform_data}}.}

\item{topics}{Vector of the number of topics to evaluate for each model.
Must be conformable to \code{integer} values.}

\item{reps}{Number of replicate starts to use for each 
value of \code{topics}. Must be conformable to \code{integer} value.}

\item{nchangepoints}{\code{integer} corresponding to the number of 
change points to include in the model. 0 is a valid input (corresponding
to no change points, so a singular time series model), and the current 
implementation can reasonably include up to 6 change points. The 
number of change points is used to dictate the segmentation of the 
time series into chunks fit with separate models dictated by 
\code{formula}.}

\item{timename}{\code{character} element indicating the time variable
used in the time series. Defaults to \code{"time"}. The variable must be
integer-conformable or a \code{Date}. If the variable named
is a \code{Date}, the input is converted to an integer, resulting in the
timestep being 1 day, which is often not desired behavior.}

\item{weights}{Optional class \code{numeric} vector of weights for each 
document. Defaults to \code{NULL}, translating to an equal weight for
each document. When using \code{multinom_TS} in a standard LDATS 
analysis, it is advisable to weight the documents by their total size,
as the result of \code{\link[topicmodels]{LDA}} is a matrix of 
proportions, which does not account for size differences among documents.
For most models, a scaling of the weights (so that the average is 1) is
most appropriate, and this is accomplished using \code{document_weights}.}

\item{control}{A \code{list} of parameters to control the fitting of the
Time Series model. Values not input assume defaults set by 
\code{\link{TS_control}}.}

\item{formula}{\code{\link[stats]{formula}} defining the regression between
the change points. Any predictor variable included must also be a column 
in \code{data} and any (compositional) response variable must be a set 
of columns in \code{data}.}
}
\value{
NOTHING YET
}
\description{
\code{LDA_TS} is the main interface function for the LDATS application
    of Linguistic Decomposition Analysis and Time Series modeling
    (Christensen  \emph{et al.} 2018). \cr \cr
}
\details{
For a (potentiall subset) dataset consisting of counts of words 
  across multiple documents in a corpus, conduct multiple Linguistic 
  Decomposition Analysis (LDA) models (e.g., Latent Dirichlet Allocation
  using the Variational Expectation Maximization (VEM) algorithm; 
  Blei \emph{et al.} 2003) to account for [1]  
  uncertainty in the number of latent topics and [2] the impact of initial
  values in the estimation procedure. \cr \cr
  This model extends the approach of Western and Kleykamp (2004;
  see also Ruggieri 2013) to compositional response data using
  softmax regression (Ripley 1996, Venables and Ripley 2002, Bishop 2006) 
  or simplical geometry ()
  via a generalized linear modeling approach (McCullagh and Nelder 1989).
  The models can be fit using a range of methods, but available procedures
  are different flavors of parallel tempering Markov Chain Monte Carlo
  (ptMCMC) methods (Earl and Deem 2005).
}
\references{
Blei, D. M., A. Y. Ng, and M. I. Jordan. 2003. Latent Dirichlet
  Allocation. \emph{Journal of Machine Learning Research} 
  \strong{3}:993-1022.
  \href{http://jmlr.csail.mit.edu/papers/v3/blei03a.html}{link}.

  Grun B. and K. Hornik. 2011. topicmodels: An R Package for Fitting Topic
  Models. \emph{Journal of Statistical Software} \strong{40}:13.
  \href{https://www.jstatsoft.org/article/view/v040i13}{link}.
}
