% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_TS.R
\name{LDA_TS_control}
\alias{LDA_TS_control}
\title{Create the controls list for the complete Linguistic Decomposition
  Analysis and Time Series model}
\usage{
LDA_TS_control(LDA_model = topicmodels_LDA,
  LDA_model_args = list(method = "VEM", seeded = TRUE),
  LDA_measurer = AIC, LDA_measurer_args = list(),
  LDA_selector = which.min, LDA_selector_args = list(),
  TS_model = sequential_TS, TS_model_args = sequential_TS_control(),
  TS_response = multinom_TS, TS_response_args = multinom_TS_control(),
  TS_method = ldats_classic, TS_method_args = ldats_classic_control(),
  nsubsets = 1, subset_rule = NULL, summary_prob = 0.95,
  soften = TRUE, quiet = FALSE, TS_measurer = AIC,
  TS_measurer_args = list(), TS_selector = which.min,
  TS_selector_args = list(), ...)
}
\arguments{
\item{LDA_model}{Main LDA \code{function}.}

\item{LDA_model_args}{\code{list} of (named) arguments to be used in 
\code{LDA_model} via \code{\link{LDA_call}}.}

\item{LDA_measurer}{\code{function} used in evaluation of the LDA
models; \code{LDA_measurer} creates a value for each model.}

\item{LDA_measurer_args}{\code{list} of (named) arguments to be used in 
\code{LDA_measurer} via \code{\link{do.call}}.}

\item{LDA_selector}{\code{function} usde in evaluation of the LDA
models; \code{LDA_selector} operates on the values to choose the models.}

\item{LDA_selector_args}{\code{list} of (named) arguments to be used in 
\code{LDA_selector} via \code{\link{do.call}}.}

\item{TS_model}{Main TS \code{function}.}

\item{TS_model_args}{\code{list} of (named) arguments to be used in 
\code{TS_model}.}

\item{TS_response}{\code{function} used to model the compositional 
response.}

\item{TS_method}{\code{function} used to drive the sampler of the TS
models; \code{TS_method} defines and operates the computational 
procedure. \cr \cr
Current pre-built options include \code{\link{ldats_classic}}.}

\item{TS_method_args}{\code{list} of (named) arguments to be used in 
\code{TS_method} via \code{\link{do.call}}. 
\cr \cr
Could be managed via a \code{<method>_control} function like
\code{\link{ldats_classic_control}}.}

\item{nsubsets}{Number of data subsets.}

\item{subset_rule}{\code{function} used to subset the data.}

\item{summary_prob}{Probability used for summarizing the posterior 
distributions (via the highest posterior density interval, see
\code{\link[coda]{HPDinterval}}).}

\item{soften}{\code{logical} indicator of whether the model should error 
softly or if errors should trigger a full-stop to the pipeline.}

\item{quiet}{\code{logical} indicator of whether the model should run 
quietly (if \code{FALSE}, a progress bar and notifications are printed).}

\item{TS_measurer}{\code{function} used in evaluation of the TS
models; \code{measurer} creates a value for each model.}

\item{TS_measurer_args}{\code{list} of (named) arguments to be used in 
\code{TS_measurer} via \code{\link{do.call}}.}

\item{TS_selector}{\code{function} usde in evaluation of the TS
models; \code{TS_selector} operates on the values to choose the models.}

\item{TS_selector_args}{\code{list} of (named) arguments to be used in 
\code{TS_selector} via \code{\link{do.call}}.}

\item{...}{Not passed along to the output, rather included to allow for
automated removal of unneeded controls.}

\item{response_args}{\code{list} of (named) arguments to be used in 
\code{TS_response} via \code{\link{do.call}}. 
\cr \cr
Could be managed via a \code{<reponse>_TS_control} function like
\code{\link{multinom_TS_control}}.}
}
\value{
\code{list} of \code{list}s and single elements, with named 
  elements corresponding to the arguments.
}
\description{
Defines and creates a \code{list} used to control the 
  decomposition and time series model fitting occurring within 
  \code{\link{LDA_TS}}.
}
\examples{
  LDA_TS_control()

}
