% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_TS.R
\name{LDA_TS_controls_list}
\alias{LDA_TS_controls_list}
\title{Create the controls list for the LDATS model}
\usage{
LDA_TS_controls_list(TS_control = TS_controls_list(),
  LDA_control = LDA_controls_list(), quiet = FALSE)
}
\arguments{
\item{TS_control}{Class \code{TS_controls} list, holding control parameters
for the Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls, generated by 
\code{\link{TS_controls_list}}.}

\item{LDA_control}{Named list of control parameters to be used in 
\code{LDA} (note that "seed" will be overwritten).}

\item{quiet}{\code{logical} indicator of whether the model should run 
quietly.}
}
\value{
control Class \code{LDA_TS_controls} list that contains 
  \code{LDA_controls}, \code{TS_controls}, and the top-level \code{quiet}.
}
\description{
This function provides a simple creation and definition of the
  list used to control the LDATS model including the LDA and TS model 
  specifically.
}
