% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_TS_plots.R
\name{LDA_TS_summary_cols}
\alias{LDA_TS_summary_cols}
\title{Create the list of colors for the LDATS summary plot}
\usage{
LDA_TS_summary_cols(rho_cols = NULL, rho_option = "D",
  rho_alpha = 0.4, gamma_cols = NULL, gamma_option = "C",
  gamma_alpha = 0.8)
}
\arguments{
\item{rho_cols}{Colors to be used to plot the histograms of changepoints.}

\item{rho_option}{A character string indicating the colormap option to use
if `rho_cols == NULL`. Four options are available: "magma" (or "A"), 
"inferno" (or "B"), "plasma" (or "C"), "viridis" (or "D", the default
option) and "cividis" (or "E").}

\item{rho_alpha}{Numeric value [0,1] that indicates the transparency of the 
colors used. Supported only on some devices, see \code{rgb}.}

\item{gamma_cols}{Colors to be used to plot the LDA topic proportions,
time series of observed topic proportions, and time series of fitted 
topic proportions.}

\item{gamma_option}{A character string indicating the colormap option to 
use if `gamma_cols == NULL`. Four options are available: "magma" (or 
"A"), "inferno" (or "B", the default option), "plasma" (or "C"), 
"viridis" (or "D") and "cividis" (or "E")}

\item{gamma_alpha}{Numeric value [0,1] that indicates the transparency of 
the colors used. Supported only on some devices, see \code{rgb}.}
}
\value{
cols \code{list} of elements used to define the colors for the two
  panels of the summary plot, as generated simply using 
  \code{LDA_TS_summary_cols}. \code{cols} has two elements \code{LDA} and
  \code{TS}, each corresponding to its stage in the full model. 
  \code{LDA} contains entries \code{cols} and \code{options} (see
  \code{set_LDA_plot_colors} for details on usage. \code{TS} contains two
  entries, \code{rho} and \code{gamma}, each corresponding to the related 
  panel, and each containing default values for entries named \code{cols}, 
  \code{option}, and \code{alpha}. See \code{set_gamma_colors} and 
  \code{set_rho_hist_colors} for details on usage.
}
\description{
A default list generator function that produces the options
  for the colors controlling the panels of the LDATS summary plots, needed
  because the panels should be in different color schemes. See 
  \code{set_LDA_plot_colors}, \code{set_gamma_colors}, and 
  \code{set_rho_hist_colors} for specific details on usage.
}
