% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDATS.R
\docType{package}
\name{LDATS}
\alias{LDATS}
\alias{LDATS-package}
\title{Two-stage LDA and Time Series analyses}
\description{
Performs two-stage analysis of multivariate temporal data
  using a combination of Latent Dirichlet Allocation (Blei \emph{et al.}
  2003) and multinomial time series models (Western and Kleykamp 2004)
  following Christensen \emph{et al.} (2018).
}
\references{
Blei, D. M., A. Y. Ng, and M. I. Jordan. 2003. Latent Dirichlet
  Allocation. \emph{Journal of Machine Learning Research} 
  \strong{3}:993-1022.
  \href{http://jmlr.csail.mit.edu/papers/v3/blei03a.html}{link}.

  Christensen, E., D. J. Harris, and S. K. M. Ernest. 2018.
  Long-term community change through multiple rapid transitions in a 
  desert rodent community. \emph{Ecology} \strong{99}:1523-1529. 
  \href{https://doi.org/10.1002/ecy.2373}{link}.

  Western, B. and M. Kleykamp. 2004. A Bayesian change point model for 
  historical time series analysis. \emph{Political Analysis}
  \strong{12}:354-374.
  \href{https://doi.org/10.1093/pan/mph023}{link}.
}
\keyword{package}
