% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_methods.R
\name{ldats_classic_control}
\alias{ldats_classic_control}
\title{Create the controls list for the classic LDATS ptMCMC sampler}
\usage{
ldats_classic_control(ntemps = 6, penultimate_temp = 2^6,
  ultimate_temp = 1e+10, q = 0, nit = 10000, magnitude = 12,
  burnin = 0, thin_frac = 1, memoise = TRUE, quiet = FALSE)
}
\arguments{
\item{ntemps}{\code{integer} number of temperatures (chains) to use in the 
ptMCMC algorithm.}

\item{penultimate_temp}{Penultimate temperature in the ptMCMC sequence.}

\item{ultimate_temp}{Ultimate temperature in the ptMCMC sequence.}

\item{q}{Exponent controlling the ptMCMC temperature sequence from the 
focal chain (reference with temperature = 1) to the penultimate chain. 0
(default) implies a geometric sequence. 1 implies squaring before 
exponentiating.}

\item{nit}{\code{integer} number of iterations (steps) used in the ptMCMC
algorithm.}

\item{magnitude}{Average magnitude (defining a geometric distribution)
for the proposed step size in the ptMCMC algorithm.}

\item{burnin}{\code{integer} number of iterations to remove from the 
beginning of the ptMCMC algorithm.}

\item{thin_frac}{Fraction of iterations to retain, must be \eqn{(0, 1]},
and the default value of 1 represents no thinning.}

\item{memoise}{\code{logical} indicator of whether the response 
function should be memoised (via \code{\link[memoise]{memoise}}).}

\item{quiet}{\code{logical} indicator of whether the model should run 
quietly (if \code{FALSE}, a progress bar and notifications are printed).}
}
\value{
\code{list}, with named elements corresponding to the arguments.
}
\description{
This function provides a simple creation and definition of a
  list used to control time series model fitting via the
  \code{\link{ldats_classic}} sampler.
}
