% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list_depth}
\alias{list_depth}
\title{Determine the depth of a list}
\usage{
list_depth(xlist)
}
\arguments{
\item{xlist}{Focal input \code{list}.}
}
\value{
\code{integer} value of the depth of the list.
}
\description{
Evaluate an input for the depth of its nesting.
}
\details{
If \code{xlist = list()}, then technically the input value is a 
 list, but is empty (of length \code{0}), so depth is returned as \code{0}.
}
\examples{
 list_depth("a")
 list_depth(list())
 list_depth(list("a"))
 list_depth(list(list("a")))

}
