% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{logLik.LDA}
\alias{logLik.LDA}
\title{Determine the log likelihood of a Linguistic Decomposition Analysis
  model}
\usage{
\method{logLik}{LDA}(object, ...)
}
\arguments{
\item{object}{Class \code{LDA} object to be evaluated.}

\item{...}{Not used, simply included to maintain method compatibility.}
}
\value{
Log likelihood of the model \code{logLik}, also with \code{df}
  (degrees of freedom) and \code{nobs} (number of observations) values.
}
\description{
Convenience function to extract and format the log likelihood
  of a \code{LDA}-class object fit by \code{\link{LDA_call}}.
}
