% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{logLik.LDA_VEM}
\alias{logLik.LDA_VEM}
\title{Calculate the log likelihood of a VEM LDA model fit}
\usage{
\method{logLik}{LDA_VEM}(object, ...)
}
\arguments{
\item{object}{A \code{LDA_VEM}-class object.}

\item{...}{Not used, simply included to maintain method compatability.}
}
\value{
Log likelihood of the model, also with \code{df} (degrees of
   freedom) and \code{nobs} (number of observations) values.
}
\description{
Imported calculations from topicmodels, as applied to
  Latent Dirichlet Allocation fit with Variational Expectation Maximization
  via \code{\link[topicmodels]{LDA}}.
}
\references{
Grun B. and K. Hornik. 2011. topicmodels: An R Package for Fitting Topic
  Models. Journal of Statistical Software 40(13):1-30.
}
