% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_responses.R
\name{logLik.mlm}
\alias{logLik.mlm}
\title{Determine the log likelihood of a multivariate linear regression
  model}
\usage{
\method{logLik}{mlm}(object, ...)
}
\arguments{
\item{object}{multivariate linear regression model fit with 
 \code{\link{lm}}.
of class \code{mlm}}

\item{...}{not used.}

\item{object}{multivariate linear regression model fit using 
\code{\link[stats]{lm}} and of class \code{mlm}}

\item{...}{Not used, simply included to maintain method compatibility.}
}
\value{
log-lik at (unrestricted) maximum with df as attribute.

Log likelihood of the model \code{logLik}, also with \code{df}
  (degrees of freedom) and \code{nobs} (number of observations) values.
}
\description{
Convenience function to extract and format the log likelihood
  of a multivariate linear regression, such as fitted by the 
  \code{\link{simplex_TS}} models.
}
\details{
Log-likelihood of multivariate linear regression model

Adapted from the function contained in the old R-Forge Atools
  package (href{https://rdrr.io/rforge/Atools/src/R/logLik.mlm.R}{see}).
}
\examples{
y <- cbind(rnorm(10), rnorm(10)); x <- 1:10;
mod <- lm(y~x)
logLik(mod)
}
\author{
Andi Boeck
}
