% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{logLik.TS}
\alias{logLik.TS}
\title{Determine the log likelihood of a Time Series model}
\usage{
\method{logLik}{TS}(object, ...)
}
\arguments{
\item{object}{Class \code{TS} object to be evaluated.}

\item{...}{Not used, simply included to maintain method compatibility.}
}
\value{
Log likelihood of the model \code{logLik}, also with \code{df}
  (degrees of freedom) and \code{nobs} (number of observations) values.
}
\description{
Convenience function to extract and format the log likelihood
  of a \code{TS}-class object fit by \code{\link{sequential_TS}}.
}
