% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_fit.R
\name{logLik.TS_fit}
\alias{logLik.TS_fit}
\title{Log likelihood of a TS model (as a TS_fit-class list)}
\usage{
\method{logLik}{TS_fit}(object, ...)
}
\arguments{
\item{object}{A \code{TS_fit}-class object.}

\item{...}{Not used, simply included to maintain method compatibility.}
}
\value{
Log likelihood of the model, as class \code{logLik}, with 
  attributes \code{df} (degrees of freedom) and \code{nobs} (the number of
  weighted observations, accounting for size differences among documents).
}
\description{
Convenience function to simply extract the \code{logLik}
  element (and \code{df} and \code{nobs}) from a \code{TS_fit}
  object fit by a \code{<response>_TS} function.
}
