% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_TS.R
\name{logLik_multinom_TS_fit}
\alias{logLik_multinom_TS_fit}
\title{Log likelihood of a multinomial TS model}
\usage{
logLik_multinom_TS_fit(object, ...)
}
\arguments{
\item{object}{A \code{multinom_TS_fit}-class object.}

\item{...}{Not used, simply included to maintain method compatability.}
}
\value{
Log likelihood of the model.
}
\description{
Convenience function to simply extract the \code{logLik}
  element from a \code{multinom_TS_fit}-class object.
}
