% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{logLik_proposed_step}
\alias{logLik_proposed_step}
\title{Calculate the log-likelihood of a time series, given a set of 
  proposed changepoints.}
\usage{
logLik_proposed_step(prop_changepts, inputs)
}
\arguments{
\item{prop_changepts}{Matrix of proposed changepoints across chains.}

\item{inputs}{Class \code{ptMCMC_inputs} list, containing the static inputs
for use within the ptMCMC algorithm.}
}
\value{
Vector of log-likelihood values for each of the chains.
}
\description{
This function wraps around \code{TS_memo} to provide simpler
  interface within the ptMCMC algorithm.
}
