% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{logsumexp}
\alias{logsumexp}
\title{Calculate the log-sum-exponential (LSE) of a vector}
\usage{
logsumexp(x)
}
\arguments{
\item{x}{\code{numeric} vector}
}
\value{
The LSE.
}
\description{
Calculate the exponent of a vector (offset by the max), sum
the elements, calculate the log, remove the offset.
}
\examples{
  logsumexp(1:10)

}
