% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{messageq}
\alias{messageq}
\title{Optionally generate a message based on a logical input}
\usage{
messageq(msg = NULL, quiet = FALSE)
}
\arguments{
\item{msg}{\code{character} vector of the message(s) to generate or 
\code{NULL}. If more than one element is contained in \code{msg}, they
are concatenated with a newline between.}

\item{quiet}{\code{logical} indicator controlling if the message is
generated.}
}
\description{
Given the input to \code{quiet}, generate the message(s) 
  in \code{msg} or not.
}
\examples{
  messageq("hello")
  messageq("hello", TRUE)

}
