% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA-distance.R
\name{min_H}
\alias{min_H}
\title{Find the configuration with the minimum Hellinger distance between two
probability matrices like those from `ps`.}
\usage{
min_H(p1, p2, k)
}
\arguments{
\item{p1, p2}{numeric vectors: species composition}

\item{k}{number of topics}
}
\description{
Find the configuration with the minimum Hellinger distance between two
probability matrices like those from `ps`.
}
