% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{modalvalue}
\alias{modalvalue}
\title{Determine the mode of a distribution}
\usage{
modalvalue(x)
}
\arguments{
\item{x}{\code{numeric} vector.}
}
\value{
Numeric value of the mode.
}
\description{
Find the most common entry in a vector. Ties are not included,
  the first value encountered within the modal set if there are ties is
  deemed the mode.
}
