% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_functions.R
\name{MTS_prep}
\alias{MTS_prep}
\title{Prep the data inputs to the MTS function}
\usage{
MTS_prep(lda_models = NULL, document_covariate_matrix = NULL)
}
\arguments{
\item{lda_models}{selected LDA model(s)}

\item{document_covariate_matrix}{matrix of documents (rows) by covariates
(columns)}
}
\value{
data matrix ready for MTS analyses
}
\description{
prepares the output from an LDA to work with MTS
}
