% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_functions.R
\name{MTS_set}
\alias{MTS_set}
\title{Conduct a set of MTS analyses}
\usage{
MTS_set(prepped_data = NULL, formula, nchangepoints, document_term_matrix,
  ...)
}
\arguments{
\item{prepped_data}{data prepped for each of the MTS analyses}

\item{formula}{vector of formulas for the continuous change}

\item{nchangepoints}{vector of the number of change points to include in 
the model}

\item{document_term_matrix}{matrix of documents (rows) by terms (columns)}

\item{...}{additional arguments to be passed to subfunctions}
}
\value{

}
