% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_TS.R
\name{multinom_TS_control}
\alias{multinom_TS_control}
\title{Create the controls list for the multinomial Time Series model}
\usage{
multinom_TS_control(lambda = 0, quiet = FALSE, ...)
}
\arguments{
\item{lambda}{\code{numeric} "weight" decay term used to set the prior
on the regressors within each chunk-level model. Defaults to 0, 
corresponding to a fully vague prior.}

\item{quiet}{\code{logical} indicator of whether the model should run 
quietly (if \code{FALSE}, a progress bar and notifications are printed).}

\item{...}{Not passed along to the output, rather included to allow for
automated removal of unneeded controls.}
}
\value{
\code{list}, with named elements corresponding to the arguments.
}
\description{
This function provides a simple creation and definition of a
  list used to control the time series model fit occurring within 
  \code{\link{multinom_TS}}.
}
