% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector}
\usage{
normalize(x)
}
\arguments{
\item{x}{\code{numeric} vector.}
}
\value{
Normalized \code{x}.
}
\description{
Normalize a \code{numeric} vector to be on the scale of [0,1].
}
