% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsampling_functions.R
\name{ntopic_hist_file}
\alias{ntopic_hist_file}
\title{Save out a histogram of the topics for a dataset}
\usage{
ntopic_hist_file(dd = NULL, nn = NULL)
}
\arguments{
\item{dd}{data}

\item{nn}{name}
}
\value{
Figure saved out
}
\description{
Save out a histogram of the topics for a dataset
}
