% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_fit.R
\name{package_chunk_fits}
\alias{package_chunk_fits}
\title{Package the output of the chunk-level TS models into a TS_fit list}
\usage{
package_chunk_fits(chunks, fits)
}
\arguments{
\item{chunks}{Data frame of \code{start} and \code{end} times for each 
chunk (row).}

\item{fits}{List of chunk-level fits returned by \code{TS_chunk_memo},
the memoised version of \code{\link{multinom_TS_chunk}}.}
}
\value{
Object of class \code{TS_fit}, which is a list of [1]
  chunk-level model fits, [2] the total log likelihood combined across 
  all chunks, and [3] the chunk time data table.
}
\description{
Takes the list of fitted chunk-level models returned from
  a \code{<response>_TS_chunk} function and packages it as a 
  \code{TS_fit} object. This involves naming the model fits based 
  on the chunk time windows, combining the log likelihood values across the 
  chunks, and setting the class of the output object.
}
