% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{package_LDA}
\alias{package_LDA}
\alias{select_LDA}
\alias{measure_LDA}
\title{Measure, select, and package the output of a set of Linguistic
  Decomposition Analysis models}
\usage{
package_LDA(selected_LDAs, LDAs, control = list())

select_LDA(LDAs = list(), control = list())

measure_LDA(LDAs = list(), control = list())
}
\arguments{
\item{selected_LDAs}{\code{list} of selected LDA model \code{list}s.}

\item{LDAs}{\code{list} of LDA model \code{list}s.}

\item{control}{A \code{list} of parameters to control the fitting of the
LDA model. Values not input assume defaults set by 
\code{\link{LDA_control}}.}
}
\value{
\code{measure_LDA}: \code{vector} of values corresponding to the model
    evaluations.
  \code{select_LDA}: \code{list} of selected models' \code{list}s.
  \code{pacakage_LDA}: class \code{LDA_set} \code{list} of both selected 
    and all results from \code{\link{LDA_call}} applied for each model on
    each data input as well as the control \code{list} used to fit 
    the model.
}
\description{
\code{measure_LDA} determines the fit value used to select among the 
    models. \cr \cr
  \code{select_LDA} chooses the best model(s) of interest based on their
    measured values and the selector function. \cr \cr
  \code{package_LDA} sets the class and names the elements of the results
    \code{list} from \code{\link{LDA_call}} applied to the 
    combination of TS models requested for the data input(s).
}
