% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{package_LDA_set}
\alias{package_LDA_set}
\title{Package the output from LDA_set}
\usage{
package_LDA_set(mods, mod_topics, mod_seeds)
}
\arguments{
\item{mods}{Fitted models returned from \code{LDA}.}

\item{mod_topics}{Vector of \code{integer} values corresponding to the 
number of topics in each model.}

\item{mod_seeds}{Vector of \code{integer} values corresponding to the 
seed used for each model.}
}
\value{
List (class: \code{LDA_set}) of LDA models (class: \code{LDA}).
}
\description{
Name the elements (LDA models) and set the class 
  (\code{LDA_set}) of the models returned by the \code{LDA_set} function.
}
