% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_TS.R
\name{package_LDA_TS}
\alias{package_LDA_TS}
\title{Package LDA and TS model outputs}
\usage{
package_LDA_TS(LDAs, TSs, control)
}
\arguments{
\item{LDAs}{\code{LDA_set} \code{list} of selected and all LDAs from 
\code{\link{LDA}}.}

\item{TSs}{\code{TS_set} \code{list} of selected and all TSs from 
\code{\link{TS}}.}

\item{control}{A \code{list} of parameters to control the fitting of the
LDATS model. Values not input assume defaults set by 
\code{\link{LDA_TS_control}}.}
}
\value{
\code{LDA_TS} \code{list} with all fitted LDA and TS models and 
    selected models specifically as elements named
    \code{"LDA models"} (from \code{\link{LDA_set}}),
    \code{"Selected LDA models"} (from \code{\link{select_LDA}}), 
    \code{"TS models"} (from \code{\link{TS_on_LDA}}), and
    \code{"Selected TS models"} (from \code{\link{select_TS}}) elements.
}
\description{
Combine the results from each model component.
}
