% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{package_TS}
\alias{package_TS}
\alias{select_TS}
\alias{measure_TS}
\title{Measure, select, and package the output of a set of Time Series 
  models}
\usage{
package_TS(selected_TSs, TSs, control = list())

select_TS(TSs, control = list())

measure_TS(TSs, control = list())
}
\arguments{
\item{selected_TSs}{\code{list} of selected time series model \code{list}s.}

\item{TSs}{\code{list} of  time series model \code{list}s.}

\item{control}{A \code{list} of parameters to control the fitting of the
Time Series model. Values not input assume defaults set by 
\code{\link{TS_control}}.}
}
\value{
\code{measure_TS}: \code{vector} of values corresponding to the model
    evaluations.
  \code{select_TS}: \code{list} of selected models' \code{list}s.
  \code{pacakage_TS}: class \code{TS_set} \code{list} of both selected and
    all results from \code{\link{sequential_TS}} applied for each model on
    each LDA model input as well as the control \code{list} used to fit 
    the model.
}
\description{
\code{measure_TS} determines the fit value used to select among the 
    models. \cr \cr
  \code{select_TS} chooses the best model(s) of interest based on their
    measured values and the selector function. \cr \cr
  \code{package_TS} sets the class and names the elements of the results
    \code{list} from \code{\link{sequential_TS}} applied to the 
    combination of TS models requested for the LDA model(s) input.
}
