% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_plots.R
\name{plot.LDA_list}
\alias{plot.LDA_list}
\title{Plot a set of LDATS LDA models}
\usage{
\method{plot}{LDA_list}(x, ...)
}
\arguments{
\item{x}{An \code{LDA_list} object of LDA topic models.}

\item{...}{Additional arguments to be passed to subfunctions.}
}
\value{
Model-by-model plots.
}
\description{
Generalization of the plot function to work on a list of LDA 
  topic models (class \code{LDA_list}).
}
\examples{
\dontrun{
  data(rodents)
  lda_data <- select(rodents, -c(newmoon, date, plots, traps))
  r_LDA <- parLDA(lda_data, topics = 2, nseeds = 2)
  plot(r_LDA)
}
}
