% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_plots.R
\name{plot.LDA_set}
\alias{plot.LDA_set}
\title{Plot a set of LDATS LDA models}
\usage{
\method{plot}{LDA_set}(x, ..., selected = TRUE)
}
\arguments{
\item{x}{An \code{LDA_set} object of LDA topic models.}

\item{...}{Not used, retained for alignment with base function.}

\item{selected}{\code{logical} indicator of if only the selected LDAs
(the first element in \code{x}) should be plotted or if all the LDAs
(the second element in \code{x}) should be plotted.}
}
\value{
\code{NULL}.
}
\description{
Generalization of the \code{\link[graphics]{plot}} function to 
  work on a list of LDA topic models (class \code{LDA_set}).
}
