% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_plots.R
\name{plot.TS}
\alias{plot.TS}
\alias{TS_diagnostics_plot}
\alias{eta_diagnostics_plots}
\alias{rho_diagnostics_plots}
\alias{trace_plot}
\alias{ecdf_plot}
\alias{posterior_plot}
\alias{autocorr_plot}
\alias{set_TS_summary_plot_cols}
\alias{TS_summary_plot}
\alias{pred_gamma_TS_plot}
\alias{rho_lines}
\alias{rho_hist}
\alias{set_rho_hist_colors}
\alias{set_gamma_colors}
\title{Plot an LDATS Time Series model}
\usage{
\method{plot}{TS}(x, ..., plot_type = "summary", interactive = FALSE,
  cols = set_TS_summary_plot_cols(), bin_width = 1, xname = NULL,
  border = NA, selection = "median", LDATS = FALSE)

TS_diagnostics_plot(x, interactive = TRUE)

eta_diagnostics_plots(x, interactive)

rho_diagnostics_plots(x, interactive)

trace_plot(draw, ylab = "parameter value")

ecdf_plot(draw, xlab = "parameter value")

posterior_plot(draw, xlab = "parameter value")

autocorr_plot(draw)

set_TS_summary_plot_cols(rho_cols = NULL, rho_option = "D",
  rho_alpha = 0.4, gamma_cols = NULL, gamma_option = "C",
  gamma_alpha = 0.8)

TS_summary_plot(x, cols = set_TS_summary_plot_cols(), bin_width = 1,
  xname = NULL, border = NA, selection = "median", LDATS = FALSE)

pred_gamma_TS_plot(x, selection = "median", cols = set_gamma_colors(x),
  xname = NULL, together = FALSE, LDATS = FALSE)

rho_lines(spec_rhos)

rho_hist(x, cols = set_rho_hist_colors(x$rhos), bin_width = 1,
  xname = NULL, border = NA, together = FALSE, LDATS = FALSE)

set_rho_hist_colors(x = NULL, cols = NULL, option = "D", alpha = 1)

set_gamma_colors(x, cols = NULL, option = "D", alpha = 1)
}
\arguments{
\item{x}{A \code{TS_fit} object of a multinomial time series model fit by
\code{\link{TS}}.}

\item{...}{Additional arguments to be passed to subfunctions. Not currently
used, just retained for alignment with \code{\link[graphics]{plot}}.}

\item{plot_type}{"diagnostic" or "summary".}

\item{interactive}{\code{logical} input, should be \code{TRUE} unless
testing.}

\item{cols, rho_cols, gamma_cols}{In \code{plot.TS}, \code{cols} is a \code{list} of elements used to
  define the colors for the two panels of the summary plot, as generated 
  simply using \code{\link{set_TS_summary_plot_cols}}. 
  \code{cols} has two elements \code{rho} and \code{gamma}, each
  corresponding to the related panel, and each containing default values
  for entries named \code{cols}, \code{option}, and \code{alpha}. \cr
For \code{rho_cols} and \code{gamma_cols} always and for \code{cols} in 
  \code{set_rho_hist_colors}, \code{set_gamma_colors}, 
  \code{rho_hist}, and \code{pred_gamma_TS_plot}, colors to be used in 
  the specific plot. Any valid color values (\emph{e.g.}, see
  \code{\link[grDevices]{colors}}, \code{\link[grDevices]{rgb}}) can be 
  input as with a standard plot. The default (\code{NULL}) triggers use 
  of \code{\link[viridis]{viridis}} color options (see 
  \code{option},\code{rho_option},\code{gamma_option}).}

\item{bin_width}{Width of the bins used in the histograms of the summary 
time series plot, in units of the x-axis (the time variable used to fit 
the model).}

\item{xname}{Label for the x-axis in the summary time series plot. Defaults
to \code{NULL}, which results in usage of the \code{timename} element
of the control list (held in\code{control$TS_control$timename}). To have
no label printed, set \code{xname = ""}.}

\item{border}{Border for the histogram, default is \code{NA}.}

\item{selection}{Indicator of the change points to use in the time series
summary plot. Currently only defined for \code{"median"} and 
\code{"mode"}.}

\item{LDATS}{\code{logical} indicating if the plot is part of a larger 
LDATS plot output.}

\item{draw}{\code{vector} of parameter values drawn from the posterior 
distribution, indexed to the iteration by the order of the vector.}

\item{ylab}{\code{character} value used to label the y axis.}

\item{xlab}{\code{character} value used to label the x axis.}

\item{spec_rhos}{\code{numeric} vector indicating the locations along the
x axis where the specific change points being used are located.}

\item{option, rho_option, gamma_option}{A \code{character} string indicating
the color option from \code{\link[viridis]{viridis}} to use if
"cols == NULL". Four options are available: "magma" (or "A"), 
"inferno" (or "B"), "plasma" (or "C"), "viridis" (or "D", the default 
option) and "cividis" (or "E").}

\item{alpha, rho_alpha, gamma_alpha}{Numeric value [0,1] that indicates the 
transparency of the colors used. Supported only on some devices, see 
\code{\link[grDevices]{rgb}}.}
}
\value{
\code{plot.TS},\code{TS_diagnostics_plot},\code{eta_diagnostics_plots},
    \code{rho_diagnostics_plots},\code{trace_plot},\code{posterior_plot},
    \code{autocorr_plot},\code{ecdf_plot},\code{TS_summary_plot},
    \code{pred_gamma_TS_plot},\code{rho_hist},\code{rho_lines}:
    \code{NULL}. \cr \cr
  \code{set_rho_hist_cols},\code{set_gamma_colors}: \code{vector} of 
    \code{character} hex codes indicating colors to use.
  \code{set_TS_summary_plot_cols}: \code{list} of elements used to define
    the colors for the two panels. Contains two elements \code{rho} and 
    \code{gamma}, each corresponding to the related panel, and each 
    containing default values for entries named \code{cols}, 
    \code{option}, and \code{alpha}.
}
\description{
\code{plot.TS} is a generalization of the \code{\link[graphics]{plot}}
    function to work on fitted TS model objects (class \code{TS}) 
    returned from \code{\link{TS}}. \cr \cr
  \code{TS_diagnostics_plot} makes the 4-panel figures (showing trace 
    plots, posterior ECDF, posterior density, and iteration 
    autocorrelation) for each of the parameters (change point locations 
    and regressors) fitted within a compositional time series model (fit
    by \code{\link{TS}}). \cr \cr
  \code{eta_diagnostics_plots} creates the diagnostic plots
    for the regressors (etas) of a time series model. \cr \cr
  \code{rho_diagnostics_plots} creates the diagnostic plots
    for the change point locations (rho) of a time series model. \cr \cr
  \code{trace_plot} produces a trace plot for the parameter of interest 
    (rho or eta) as part of \code{\link{TS_diagnostics_plot}}. A 
    horizontal line is added to show the median of the posterior. \cr \cr
  \code{ecdf_plot} makes a vanilla ECDF (empirical cumulative distribution
    function) plot using \code{\link[stats]{ecdf}} for the parameter of 
    interest (rho or eta) as part of \code{\link{TS_diagnostics_plot}}.
    A horizontal line is added to show the median of the posterior. \cr \cr
  \code{autocorr_plot} produces a vanilla ACF plot using 
    \code{\link[stats]{acf}} for the parameter of interest (rho or eta)
    as part of \code{\link{TS_diagnostics_plot}}.\cr \cr
  \code{posterior_plot} makes a vanilla histogram plot using 
    \code{\link[graphics]{hist}} for the parameter of interest (rho or eta)
    as part of \code{\link{TS_diagnostics_plot}}. A vertical line is added 
    to show the median of the posterior. \cr \cr
  \code{TS_summary_plot} produces a two-panel figure of [1] the change 
    point distributions as histograms over time and [2] the time series of 
    the fitted topic proportions over time, based on a selected set of 
    change point locations. \cr \cr
  \code{pred_gamma_TS_plot} produces a time series of the 
    fitted topic proportions over time, based on a selected set of change 
    point locations. \cr \cr
  \code{rho_hist}: make a plot of the change point distributions as 
    histograms over time. \cr \cr
  \code{rho_lines} adds vertical lines to the plot of the time series of 
    fitted proportions associated with the change points of interest.
    \cr \cr
  \code{set_gamma_colors} creates the set of colors to be used in
    the time series of the fitted gamma (topic proportion) values. \cr \cr
  \code{set_rho_hist_colors} creates the set of colors to be used in
    the change point histogram. \cr \cr
  \code{set_TS_summary_plot_cols} acts as a default \code{list} 
    generator function that produces the options for the colors 
    controlling the panels of the TS summary plots, so needed
    because the panels should be in different color schemes. See 
    \code{\link{set_gamma_colors}} and \code{\link{set_rho_hist_colors}} 
    for specific details on usage.
}
