% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_plots.R
\name{plot.TS_fit}
\alias{plot.TS_fit}
\title{Plot a LDATS TS models}
\usage{
\method{plot}{TS_fit}(x, ..., plot_type = "diagnostic",
  cols = TS_summary_cols(), bin_width = 1, xlab = NULL,
  selection = "median")
}
\arguments{
\item{x}{A \code{TS_fit} object of a multinomial time series model fit by
\code{multinom_TS}.}

\item{...}{Additional arguments to be passed to subfunctions. Not currently
used, just retained for alignment with \code{plot}.}

\item{plot_type}{"diagnostic" or "summary".}

\item{cols}{\code{list} of elements used to define the colors for the two
panels of the summary plot, as generated simply using 
\code{TS_summary_cols}. \code{cols} has two elements \code{rho} and 
\code{gamma}, each corresponding to the related panel, and each 
containing default values for entries named \code{cols}, \code{option}, 
and \code{alpha}. See \code{set_gamma_colors} and 
\code{set_rho_hist_colors} for details on usage.}

\item{bin_width}{Width of the bins used in the histograms of the summary 
timeseries plot, in units of the x-axis (the time variable used to fit 
the model).}

\item{xlab}{Label for the x-axis in the summary time series plot.}

\item{selection}{Indicator of the changepoints to use in the timeseries
summary plot. Currently only defined for "median" and "mode".}
}
\value{
Nothing. Plots are generated in the active graphics device.
}
\description{
Generalization of the \code{plot} function to work on fitted
  TS model (class \code{TS_fit}).
}
