% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_plots.R
\name{plot.TS_set}
\alias{plot.TS_set}
\title{Plot a set of LDATS TS models}
\usage{
\method{plot}{TS_set}(x, ..., selected = TRUE)
}
\arguments{
\item{x}{An \code{TS_set} object of TS topic models.}

\item{...}{Not used, retained for alignment with base function.}

\item{selected}{\code{logical} indicator of if only the selected TSs
(the first element in \code{x}) should be plotted or if all the TSs
(the second element in \code{x}) should be plotted.}
}
\value{
\code{NULL}.
}
\description{
Generalization of the \code{\link[graphics]{plot}} function to 
  work on a list of TS topic models (class \code{TS_set}).
}
