% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_figure_scripts.R
\name{plot_community_composition}
\alias{plot_community_composition}
\title{Plot species composition of topics}
\usage{
plot_community_composition(composition, topic_order = 1:dim(composition)[1],
  colors = cbPalette)
}
\arguments{
\item{composition}{matrix of species composition of topics; as in output of community_composition()}

\item{topic_order}{order of topics -- for making this bar graph relate to the component community graph}
}
\value{
barplots of the n component communities
}
\description{
Plot species composition of topics
}
\examples{
plot_community_composition(community_composition(ldamodel))
}
