% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_figure_scripts.R
\name{plot_community_composition_gg}
\alias{plot_community_composition_gg}
\title{ggplot version of plot_community_composition}
\usage{
plot_community_composition_gg(composition, topic_order, ylim,
  colors = cbPalette)
}
\arguments{
\item{composition}{matrix of species composition of topics; as in output of community_composition()}

\item{topic_order}{order of topics -- for making this bar graph relate to the component community graph}

\item{ylim}{vector of limits for yaxis}
}
\value{
barplots of the n component communities
}
\description{
ggplot version of plot_community_composition
}
