% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_figure_scripts.R
\name{plot_component_communities}
\alias{plot_component_communities}
\title{Plot component communities}
\usage{
plot_component_communities(ldamodel, ntopics, xticks, ylab = "",
  topic_order = seq(ntopics), colors = cbPalette)
}
\arguments{
\item{ldamodel}{object of class LDA_VEM created by the function LDA in topicmodels package}

\item{ntopics}{number of topics used in ldamodel}

\item{xticks}{vector of dates for x-axis labels}

\item{ylab}{y axis label (optional)}

\item{topic_order}{order of topics (for color control)}
}
\value{
ggplot object
}
\description{
Plots timeseries of component communities (topics)
}
\examples{
plot_component_communities(ldamodel,ntopics,period_dates$date)
}
