% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_figure_scripts.R
\name{plot_component_communities_gibbs}
\alias{plot_component_communities_gibbs}
\title{Plot component communities}
\usage{
plot_component_communities_gibbs(results, ntopics, xticks)
}
\arguments{
\item{ntopics}{number of topics used in ldamodel}

\item{xticks}{vector of dates for x-axis labels}

\item{ldamodel}{model output of LDA model using Gibbs}
}
\value{
None
}
\description{
Plots timeseries of component communities (topics) from LDA using Gibbs sampler
}
\examples{
plot_component_communities_gibbs(ldamodel,ntopics,period_dates$date)
}
