% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_figure_scripts.R
\name{plot_component_communities_gibbs_credible}
\alias{plot_component_communities_gibbs_credible}
\title{Plot component communities including credible intervals --- WIP!
so far only works with 3 topics -- need to make more general}
\usage{
plot_component_communities_gibbs_credible(ldamodel, ntopics, xticks)
}
\arguments{
\item{ldamodel}{model output of LDA model using Gibbs}

\item{ntopics}{number of topics used in ldamodel}

\item{xticks}{vector of dates for x-axis labels}
}
\value{
None
}
\description{
Plots timeseries of component communities (topics) from LDA using Gibbs sampler
}
\examples{
plot_component_communities_gibbs_credible(ldamodel,ntopics,period_dates$date)
}
