% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_figure_scripts.R
\name{plot_component_communities_smooth}
\alias{plot_component_communities_smooth}
\title{Plot component communities -- smoothed}
\usage{
plot_component_communities_smooth(ldamodel, ntopics, xticks, smooth_factor)
}
\arguments{
\item{ldamodel}{object of class LDA_VEM created by the function LDA in topicmodels package}

\item{ntopics}{number of topics used in ldamodel}

\item{xticks}{vector of dates for x-axis labels}

\item{smooth_factor}{size of moving window average -- higher value is smoother}
}
\value{
None
}
\description{
Plots timeseries of component communities (topics)
Smooths using a simple moving window average
Plots raw data as dots, smoothed as lines
}
\examples{
plot_component_communities_smooth(ldamodel,ntopics,period_dates$date,5)
}
