% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_figure_scripts.R
\name{plot_gamma}
\alias{plot_gamma}
\title{Plot gamma}
\usage{
plot_gamma(gamma_frame, ntopics, ylab = "", colors = cbPalette)
}
\arguments{
\item{gamma_frame}{a data frame containing columns for date, relabund, and community}

\item{ntopics}{number of topics}

\item{ylab}{label for y axis (optional)}
}
\value{
a ggplot object
}
\description{
This plotting function plots the component communities over time
It's used by the plot_component_communities function, but can also take
any data frame as input as long as it's the right form -- used by the simulations
}
