% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{plot_sections}
\alias{plot_sections}
\title{plot all the fit sections}
\usage{
plot_sections(all_sections, x, changepoints)
}
\arguments{
\item{all_sections}{data frame containing plotting info for all sections}

\item{x}{same x as used in changepoint_model()}

\item{changepoints}{vector of changepoints}
}
\description{
plot all the fit sections
}
