% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_plots.R
\name{posterior_plot}
\alias{posterior_plot}
\title{Produce the posterior distribution histogram panel for the TS 
  diagnostic plot of a parameter}
\usage{
posterior_plot(x, xlab = "parameter value")
}
\arguments{
\item{x}{Vector of parameter values drawn from the posterior distribution,
indexed to the iteration by the order of the vector.}

\item{xlab}{\code{character} value used to label the x axis.}
}
\description{
Produce a vanilla histogram plot using \code{hist} for the 
  parameter of interest (rho or eta). Vertical line added to show the 
  median of the posterior.
}
