% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_plots.R
\name{pred_gamma_TS_plot}
\alias{pred_gamma_TS_plot}
\title{Create the plot of fitted topic proportions over time}
\usage{
pred_gamma_TS_plot(x, selection = "median", cols, xlab = NULL)
}
\arguments{
\item{x}{Object of class \code{TS_fit} produced by \code{\link{TS}}.}

\item{selection}{Indicator of the change points to use. Currently only
defined for "median" and "mode".}

\item{cols}{\code{list} of elements used to define the colors for the two
panels, as generated simply using \code{\link{set_TS_summary_plot_cols}}. 
Has two elements \code{rho} and \code{gamma}, each corresponding to the
related panel, and each containing default values for entries named
\code{cols}, \code{option}, and \code{alpha}. See
\code{\link{set_gamma_colors}} and \code{\link{set_rho_hist_colors}} for
details on usage.}

\item{xlab}{Label for the x-axis (time).}
}
\description{
Produces a time series of the fitted topic proportions 
  over time, based on a selected set of change point locations.
}
