% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{prep_changepts}
\alias{prep_changepts}
\title{Prepare the changepoints for the MTS algorithm}
\usage{
prep_changepts(data, formula, ntemps, nchangepoints, weights)
}
\arguments{
\item{data}{data frame including the predictor and response variables}

\item{formula}{formula for the continuous change}

\item{ntemps}{number of temperatures}

\item{nchangepoints}{number of change points to include in the model}

\item{weights}{weights}
}
\value{
list of [1] matrix of change points (rows) for each temperature 
  (columns) and [2] vector of logLiks
}
\description{
includes sorting by logLik
}
