% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldats_classic_method.R
\name{prep_cpts}
\alias{prep_cpts}
\alias{update_cpts}
\title{Initialize and update the change point matrix used in the LDATS
  classic ptMCMC algorithm}
\usage{
prep_cpts(TS, control = list())

update_cpts(cpts, swaps)
}
\arguments{
\item{TS}{\code{list} time series model object.}

\item{control}{A \code{list} of parameters to control the fitting of the
Time Series model. Values not input assume defaults set by 
\code{\link{TS_control}}.}

\item{cpts}{The existing matrix of change points.}

\item{swaps}{Chain configuration after among-temperature swaps.}
}
\value{
\code{list} of [1] \code{matrix} of change points (rows) for 
  each temperature (columns) and [2] \code{vector} of log-likelihood 
  values for each of the chains.
}
\description{
Each of the chains is initialized by \code{prep_cpts} using a 
  draw from the available times (i.e. assuming a uniform prior), the best 
  fit (by likelihood) draw is put in the focal chain with each subsequently 
  worse fit placed into the subsequently hotter chain. \code{update_cpts}
  updates the change points after every iteration in the ptMCMC algorithm.
}
