% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_on_LDA.R
\name{prep_formulas}
\alias{prep_formulas}
\title{Prepare the formulas by adding the response variable}
\usage{
prep_formulas(formulas, response)
}
\arguments{
\item{formulas}{Vector of the formulas to evaluate.}

\item{response}{Character element representing the response variable.}
}
\value{
Updated \code{formulas}.
}
\description{
Add the response variable to each of the formulas.
}
