% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{prep_LDA_control}
\alias{prep_LDA_control}
\title{Set the control inputs to include the seed}
\usage{
prep_LDA_control(seed, control = NULL)
}
\arguments{
\item{seed}{\code{number} of seeds (replicate starts) to use for the 
specific model.}

\item{control}{Named list of control parameters to be used in 
\code{\link[topicmodels]{LDA}} Note that is \code{control} has an 
element named \code{seed} it will be overwritten by the \code{seed} 
argument of \code{prep_LDA_control}.}
}
\value{
List (class: \code{list}) of controls to be used in the LDA.
}
\description{
Update the control list for the LDA model with the specific
  seed as indicated.
}
