% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldats_classic_method.R
\name{prep_proposal_dist}
\alias{prep_proposal_dist}
\title{Pre-calculate the change point proposal distribution for the ptMCMC 
  algorithm}
\usage{
prep_proposal_dist(TS, control = list())
}
\arguments{
\item{TS}{\code{list} time series model object.}

\item{control}{A \code{list} of parameters to control the fitting of the
time series model via the LDATS classic ptMCMC method. Values not input 
assume defaults set by \code{\link{ldats_classic_control}}.}
}
\value{
\code{list} of two \code{matrix} elements: [1] the size of the 
  proposed step for each iteration of each chain and [2] the identity of 
  the change point location to be shifted by the step for each iteration of
  each chain.
}
\description{
Calculate the proposal distribution in advance of actually
  running the ptMCMC algorithm in order to decrease computation time.
  The proposal distribution is a joint of three distributions:
  [1] a multinomial distribution selecting among the change points within
  the chain, [2] a binomial distribution selecting the direction of the 
  step of the change point (earlier or later in the time series), and 
  [3] a geometric distribution selecting the magnitude of the step.
}
