% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_tempering_functions.R
\name{prep_temps}
\alias{prep_temps}
\title{Prepare the temperatures}
\usage{
prep_temps(ntemps = 6, penultimate_temp = 2^6, ultimate_temp = 1e+10,
  k = 0, ...)
}
\arguments{
\item{ntemps}{number of temperatures}

\item{penultimate_temp}{penultimate temperature (second hottest chain)}

\item{ultimate_temp}{ultimate temperature (hottest chain)}

\item{k}{the exponent controlling the temperature sequence: 0 implies 
geometric sequence, 1 implies squaring before exponentiating. Use larger 
values if the cooler chains aren't swapping enough.}

\item{...}{additional arguments to be passed to subfunctions}
}
\value{
temperatures
}
\description{
based on general controls for parallel tempering
}
