% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_TS.R
\name{print.LDA_TS}
\alias{print.LDA_TS}
\title{Print the selected LDA and TS models of LDA_TS object}
\usage{
\method{print}{LDA_TS}(x, ...)
}
\arguments{
\item{x}{Class \code{LDA_TS} object to be printed.}

\item{...}{Not used, simply included to maintain method compatability.}
}
\description{
Convenience function to print only the selected elements of a 
  \code{LDA_TS}-class object returned by \code{\link{LDA_TS}}
}
