% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{print.TS}
\alias{print.TS}
\title{Print a Time Series model}
\usage{
\method{print}{TS}(x, ...)
}
\arguments{
\item{x}{Class \code{TS_fit} object to be printed.}

\item{...}{Not used, simply included to maintain method compatibility.}
}
\value{
The non-hidden parts of \code{x} are printed and returned 
  invisibly as a \code{list}.
}
\description{
Convenience function to print only the most important 
  components of a \code{TS}-class object fit by 
  \code{\link{sequential_TS}}.
}
