% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{proposal_dist}
\alias{proposal_dist}
\title{Calculate the proposal distribution for the time series model}
\usage{
proposal_dist(nit, ntemps, nchangepoints, magnitude)
}
\arguments{
\item{nit}{number of iterations used}

\item{ntemps}{number of temperatures used}

\item{nchangepoints}{number of change points to include in the model}

\item{magnitude}{the inverse of the probability used for the geometic 
distribution to generate the kick magnitudes}
}
\value{
list of two matrices: [1] kicks and [2] which_kicked
}
\description{
based on the provided magnitude
}
