% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptMCMC.R
\name{proposed_step_mods}
\alias{proposed_step_mods}
\title{Fit the chunk-level models to a time series, given a set of 
  proposed changepoints.}
\usage{
proposed_step_mods(prop_changepts, inputs)
}
\arguments{
\item{prop_changepts}{Matrix of proposed changepoints across chains.}

\item{inputs}{Class \code{ptMCMC_inputs} list, containing the static inputs
for use within the ptMCMC algorithm.}
}
\value{
List of models associated with the proposed step.
}
\description{
This function wraps around \code{TS_memo} to provide simpler
  interface within the ptMCMC algorithm.
}
