% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_tempering_functions.R
\name{remove_burnin}
\alias{remove_burnin}
\title{Remove the first defined number of burn-in samples}
\usage{
remove_burnin(object, burnin = 0)
}
\arguments{
\item{object}{Object to be trimmed}

\item{burnin}{number of samples to remove from the beginning}
}
\value{
object with burn-in samples removed
}
\description{
Flexible to the 2- or 3-D structure of objects to be trimmed
}
