% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_model_selection.R
\name{repeat_VEM}
\alias{repeat_VEM}
\title{repeat VEM a bunch of times with different seeds and calculate AICs to find distribution of "best" ntopics}
\usage{
repeat_VEM(dat, seeds, topic_min, topic_max)
}
\arguments{
\item{dat}{dat}

\item{seeds}{vector of seeds to use for analysis}

\item{topic_min}{lowest number of topics; must be >=2}

\item{topic_max}{highest number of topics}
}
\description{
repeat VEM a bunch of times with different seeds and calculate AICs to find distribution of "best" ntopics
}
\examples{
best_ntopic = repeat_VEM(dat,1:500,2,6)
}
