% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_plots.R
\name{rho_hist}
\alias{rho_hist}
\title{Create the changepoint histograms plot}
\usage{
rho_hist(x, cols, bin_width, xlab = NULL)
}
\arguments{
\item{x}{Object of class \code{TS_fit}.}

\item{cols}{Hex values of the colors to be used to plot the histograms of 
changepoints.}

\item{bin_width}{Width of the bins used in the histograms, in units of the
x-axis (the time variable used to fit the model).}

\item{xlab}{Label for the x-axis.}
}
\value{
Nothing. The plot is generated in the active graphics device.
}
\description{
Produces a plot of the changepoint distributions as histograms
  over time.
}
