% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rodents}
\alias{rodents}
\title{Portal rodent data}
\format{A tbl with rows corresponding to sampling events and columns to 
  covariates (newmoon number, date, number of plots trapped, number of 
  traps used) and the 21 species.}
\source{
\url{https://github.com/weecology/PortalData/Rodents}
}
\usage{
rodents
}
\description{
An example LDATS dataset, functionally that used in Christensen \emph{et 
  al.} (\emph{in press}). The data are counts of 21 rodent species across  
  436 sampling events, with the count being the total number observed 
  across 8 50 m x 50 m plots, each sampled using 49 live traps (Brown 1998, 
  Ernest \emph{et al.} 2016).
}
\references{
Brown, J. H. 1998. 
  The desert granivory experiments at Portal. 
  Pages 71-95 in W. J. Resetarits Jr. and J. Bernardo, \emph{editors}, 
  \emph{Experimental Ecology}. 
  Oxford University Press, New York, New York, USA.

  Christensen, E., D. J. Harris, and S. K. M. Ernest. \emph{In press}.
  Long-term community change through multiple rapid transitions in a 
  desert rodent community. 
  \href{https://www.biorxiv.org/content/early/2017/12/15/163931}{Preprint}.

  Ernest, S. K. M., \emph{et al}. 2016. 
  Long-term monitoring and experimental manipulation of a Chihuahuan desert
    ecosystem near Portal, Arizona ({1977-2013}).
  \emph{Ecology} \strong{97}:1082. {doi: 10.1890/15-2115.1}
}
\keyword{datasets}
