% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{sequential_TS}
\alias{sequential_TS}
\alias{package_sequential_TS}
\alias{est_changepoints}
\alias{est_regressors}
\title{Estimate a Time Series model sequentially}
\usage{
sequential_TS(TS, control = list())

package_sequential_TS(TS, rho_dist, eta_dist, control = list())

est_changepoints(TS, control = list())

est_regressors(rho_dist, TS, control = list())
}
\arguments{
\item{control}{A \code{list} of parameters to control the fitting of the
Time Series model. Values not input assume defaults set by 
\code{\link{TS_control}}.}
}
\value{
\code{est_changepoints}: \code{list} of saved data objects from the 
    estimation of change point locations, uunless \code{nchangepoints} 
    is 0, then \code{NULL}.
}
\description{
This set of functions estimates the Time Series model
  by sequential methods that first estimate the change point locations
  with full flexibility of the regressor models between change points,
  then estimate the regressors between the change points, conditional on
  their locations.

  \code{est_change points} estimates the change point location 
   distributions for multinomial Time Series analyses.
}
