% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_TS.R
\name{sequential_TS_control}
\alias{sequential_TS_control}
\title{Create the controls list for a sequential Time Series model}
\usage{
sequential_TS_control(method = "ldats_classic",
  method_args = ldats_classic_control(), summary_prob = 0.95,
  soften = TRUE, quiet = FALSE)
}
\arguments{
\item{method}{\code{function} used to drive the sampler of the TS
models; \code{method} defines and operates the computational procedure.
\cr \cr
Current pre-built options include \code{\link{ldats_classic}}.}

\item{method_args}{\code{list} of (named) arguments to be used in 
\code{method} via \code{\link{do.call}}. 
\cr \cr
Could be managed via a \code{<method>_control} function like
\code{\link{ldats_classic_control}}.}

\item{summary_prob}{Probability used for summarizing the posterior 
distributions (via the highest posterior density interval, see
\code{\link[coda]{HPDinterval}}).}

\item{soften}{\code{logical} indicator of whether the model should error 
softly or if errors should trigger a full-stop to the pipeline.}

\item{quiet}{\code{logical} indicator of whether the model should run 
quietly (if \code{FALSE}, a progress bar and notifications are printed).}

\item{response}{\code{character} element indicating the response variable 
used in the time series. \cr \cr
Must have a corresponding \code{<response>_TS} function.}
}
\value{
\code{list}, with named elements corresponding to the arguments.
}
\description{
This function provides a simple creation and definition of a
  list used to control the time series model fit occurring within 
  \code{\link{sequential_TS}}.
}
