% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_plots.R
\name{set_LDA_plot_colors}
\alias{set_LDA_plot_colors}
\title{Prepare the colors to be used}
\usage{
set_LDA_plot_colors(x, cols, option)
}
\arguments{
\item{x}{Object of class \code{LDA}}

\item{cols}{colors to be used to plot the topics}

\item{option}{A character string indicating the colormap option to use if 
`cols == NULL`. Four options are available: "magma" (or "A"), "inferno" 
(or "B"), "plasma" (or "C"), "viridis" (or "D", the default option) and 
"cividis" (or "E").}
}
\value{
Vector of colors to use.
}
\description{
Based on the inputs, create the set of colors to be used in
  the plots
}
