% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_LDA_data}
\alias{sim_LDA_data}
\title{Simulate LDA data from an LDA structure given parameters}
\usage{
sim_LDA_data(N, alpha, Beta, seed = NULL)
}
\arguments{
\item{N}{A vector of document sizes (total word counts). Must be integer
conformable. Is used to infer the total number of documents.}

\item{alpha}{Single positive numeric value for the Dirichlet distribution
parameter defining topics within documents.}

\item{Beta}{\code{matrix} of categorical distribution parameters defining
terms within topics. Dimension: (k) (number of topics) x (V)
(number of terms). Used to infer both (k) and (V). Must be
non-negative and sum to 1 within topics.}

\item{seed}{Input to \code{\link{set.seed}}.}
}
\value{
A document-by-term \code{matrix} of counts (dim: M x V).
}
\description{
For a given set of parameters \code{alpha} and \code{Beta} and
  document-specific total word counts, simulate a document-by-term matrix.
}
\details{
Additional structuring variables (the number of topics (k),
  the number of documents (M), and the number of terms (V)) are 
  inferred from input objects.
}
\examples{
  N <- c(10, 22, 15, 31)
  alpha <- 1.2
  Beta <- matrix(c(0.1, 0.1, 0.8, 0.2, 0.6, 0.2), 2, 3, byrow = TRUE)
  sim_LDA_data(N, alpha, Beta)

}
