% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_responses.R
\name{simplex_TS}
\alias{simplex_TS}
\alias{simplex_TS_chunk}
\alias{simplex_TS_control}
\title{Fit a simplex-based change point Time Series model}
\usage{
simplex_TS(data, formula, changepoints = NULL, timename = "time",
  weights = NULL, control = list())

simplex_TS_chunk(data, formula, chunk, timename = "time",
  weights = NULL, control = list())

simplex_TS_control(transformation = ilr, quiet = FALSE, ...)
}
\arguments{
\item{data}{\code{data.frame} including [1] the time variable (indicated 
in \code{timename}), [2] the predictor variables (required by
\code{formula}) and [3], the compositional response variable (indicated 
in \code{formula}). \cr \cr 
Note that the response variables should be formatted as a 
\code{data.frame} object named as indicated by the 
\code{response} entry in the \code{control} list, such as \code{gamma} 
for a standard TS analysis on LDA output. \cr \cr
See \code{Examples}.}

\item{formula}{\code{\link[stats]{formula}} defining the regression between
relationship the change points. Any 
predictor variable included must also be a column in 
\code{data} and any (compositional) response variable must be a set of
columns in \code{data}.}

\item{changepoints}{Numeric vector indicating locations of the change 
points. Must be conformable to \code{integer} values.}

\item{timename}{\code{character} element indicating the time variable
used in the time series. Defaults to \code{"time"}. The variable must be
integer-conformable or a \code{Date}. If the variable named
is a \code{Date}, the input is converted to an integer, resulting in the
timestep being 1 day, which is often not desired behavior.}

\item{weights}{Optional class \code{numeric} vector of weights for each 
document. Defaults to \code{NULL}, translating to an equal weight for
each document. When using \code{simplex_TS} in a standard LDATS 
analysis, it is advisable to weight the documents by their total size,
as the result of \code{\link{topicmodels_LDA}} is a matrix of 
proportions, which does not account for size differences among documents.
For most models, a scaling of the weights (so that the average is 1) is
most appropriate, and this is accomplished using 
\code{\link{document_weights}}.}

\item{control}{A \code{list} of parameters to control the fitting of the
Time Series model. Values not input assume defaults set by 
\code{\link{TS_control}}.}

\item{chunk}{Length-2 vector of times: [1] \code{start}, the start time 
for the chunk and [2] \code{end}, the end time for the chunk.}

\item{transformation}{Ratio \code{function} to use for the transformation 
to the simplex geometry. Options include \code{\link[compositions]{alr}},
\code{\link[compositions]{clr}}, and \code{\link[compositions]{ilr}}.}

\item{quiet}{\code{logical} indicator of whether the model should run 
quietly (if \code{FALSE}, a progress bar and notifications are printed).}

\item{...}{Not passed along to the output, rather included to allow for
automated removal of unneeded controls.}
}
\value{
\code{simplex_TS}: \code{TS_fit} \code{list} of [1] chunk-level model 
    fits (\code{"chunk models"}), [2] the total log likelihood across 
    all chunks (\code{"logLik"}), and [3] a \code{data.frame} of chunk 
    beginning and ending times (with columns \code{"start"} and 
    \code{"end"}). \cr \cr
  \code{simplex_TS_chunk}: fitted model object for the chunk, 
    of class \code{lm}. \cr \cr
  \code{simplex_TS_control}: \code{list}, with named elements 
    corresponding to response function controls.
}
\description{
\code{simplex_TS} fits a set of simplex regression models 
    (Aitchison 1986, Aitchison \emph{et al.} 2002) to a time 
    series of compositional data divided into multiple segments (a.k.a. 
    chunks) based on given locations for a set of change points, using
    e.g., the isometric log ratio (ILR) transformation
    (Egozcue \emph{et al.} 2003, Pawlowsky-Glahn 2003. \cr \cr
  \code{simplex_TS_chunk} fits a simplex regression model using, e.g., the
    ILR transformation to a defined chunk of time (a.k.a. segment)
    \code{[chunk$start, chunk$end]} within a time series. \cr \cr 
  \code{simplex_TS_control} defines and creates the control \code{list} for
    fitting.
}
\references{
Aitchison, J. 1986. \emph{The Statistical Analysis of Compositional
  Data}. Monographs on Statistics and Applied Probability. Chapman & Hall
  Ltd., London, UK.

  Aitchison, J, C. Barcelo-Vidal, J.J. Egozcue, and V. Pawlowsky-Glahn.
  2002. A consise guide to the algebraic geometric structure of the 
  simplex, the sample space for compositional data analysis, Terra Nostra, 
  Schriften der Alfred Wegener-Stiftung, 03/2003.

  Egozcue J.J., V. Pawlowsky-Glahn, G. Mateu-Figueras and C. Barcelo-Vidal.
  2003. Isometric logratio transformations for compositional data analysis. 
  \emph{Mathematical Geology}, \strong{35}:279-300.

  Pawlowsky-Glahn, V. 2003. Statistical modelling on coordinates. In: 
  Thio-Henestrosa, S. and J. A. Martin-Fernandez, Eds. 
  \emph{Proceedings of the 1st International Workshop on Compositional Data
   Analysis}, Universitat de Girona, ISBN 84-8458-111-X.
  \href{link}{http://ima.udg.es/Activitats/CoDaWork03}.
}
