% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_data}
\alias{simulate_data}
\alias{simulate_LDA_data}
\alias{simulate_TS_data}
\alias{simulate_LDA_TS_data}
\title{Simulate LDA_TS data from LDA and TS model structures and parameters}
\usage{
simulate_LDA_data(N, Beta, alpha = NULL, Theta = NULL, seed = NULL)

simulate_TS_data(X, Eta, rho, tD, err = 0, seed = NULL,
  invlink = softmax)

simulate_LDA_TS_data(N, Beta, X, Eta, rho, tD, err = 0, seed = NULL,
  invlink = softmax)
}
\arguments{
\item{N}{A vector of document sizes (total word counts). Must be integer
conformable. Is used to infer the total number of documents.}

\item{Beta}{\code{matrix} of categorical distribution parameters defining
terms within topics. Dimension: k x V (number of topics x number of 
terms). Used to infer both (k) and (V). Must be non-negative and sum to 
1 within topics.}

\item{alpha}{Single positive numeric value for the Dirichlet distribution
parameter defining topics within documents. To specifically define
document topic probabilities, use \code{Theta}.}

\item{Theta}{\code{matrix} of probabilities defining topics within 
documents. Dimension: M x k (documents x topics). Must be non-negative 
and sum to 1 within documents. To generally define document topic 
probabilities, use \code{alpha}.}

\item{seed}{Input to \code{\link{set.seed}}.}

\item{X}{\code{matrix} of covariates, dimension M (number of documents) x 
C (number of covariates, including the intercept) (a.k.a the design
matrix).}

\item{Eta}{\code{matrix} of regression parameters across the segments,
dimension: SC (number of segments x number of covariates, including the
intercept) x k (number of topics).}

\item{rho}{Vector of integer-conformable time locations of changepoints or 
\code{NULL} if no changepoints. Used to determine the number of 
segments. Must exist within the bounds of the times of the documents,
\code{tD}.}

\item{tD}{Vector of integer-conformable times of the documents. Must be
of length M (as determined by \code{X}).}

\item{err}{Additive error on the link-scale. Must be a non-negative 
\code{numeric} value. Default value of \code{0} indicates no error.}

\item{invlink}{\code{function} name for the inverse link function. 
Currently available are \code{\link{softmax}} and the inverses of 
the ILR, ALR, and CLR transforms (\code{\link[compositions]{ilrInv}},
\code{\link[compositions]{alrInv}}, and
\code{\link[compositions]{clrInv}}).}
}
\value{
\code{simulate_LDA}: A document-by-term \code{matrix} of counts
   (dim: M x V). \cr \cr
  \code{simulate_TS}:  document-by-topic \code{matrix} of 
   proportions (dim: M x k). \cr \cr
  \code{simulate_LDA_TS}: A document-by-term \code{matrix} of counts
   (dim: M x V).
}
\description{
\code{simulate_LDA_data} for a given set of parameters \code{alpha} and
    \code{Beta} and document-specific total word counts, simulate a 
    document-by-term matrix. \cr
    Additional structuring variables (the numbers of topics (k),
    documents (M), terms (V)) are inferred from input objects. \cr \cr
  \code{simulate_TS_data} for a given set of covariates \code{X};
    parameters \code{Eta}, \code{rho}, and \code{err}; and
    document-specific time stamps \code{tD}, simulate a document-by-topic
    matrix. Additional structuring variables (numbers of topics (k), 
    documents (M), segments (S), and covariates per segment (C)) are 
    inferred from input objects.
  \code{simulate_LDA_TS_data} for a given set of covariates \code{X}; 
    parameters  \code{Beta}, \code{Eta}, \code{rho}, and \code{err}; and 
    document-specific time stamps \code{tD} and lengths \code{N}),
    simulate a document-by-topic matrix. \cr
    Additional structuring variables (the numbers of topics (k), terms (V), 
    documents (M), segments (S), and covariates per segment (C))
    are inferred from input objects.
}
