% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{soft_call}
\alias{soft_call}
\title{Soften a flexible function call to errors}
\usage{
soft_call(what = function(x) {     invisible(NULL) },
  args = list(NULL), quote = FALSE, envir = parent.frame(),
  soften = FALSE)
}
\arguments{
\item{what}{\code{function} or a non-empty \code{character} string 
naming the \code{function} to be called. \cr
See \code{\link{do.call}}.}

\item{args}{\code{list} of arguments to the \code{what} call. The names
attribute of the \code{list} gives the argument names.\cr
See \code{\link{do.call}}.}

\item{quote}{\code{logical} value indicating whether to quote the 
arguments. \cr
See \code{\link{do.call}}.}

\item{envir}{\code{environment} within which to evaluate the call. This 
argument will be most useful if \code{what} is a \code{character} 
string and the arguments are symbols or \code{quote}d expressions.\cr
See \code{\link{do.call}}.}

\item{soften}{\code{logical} value indicating whether to soften any errors 
in the running of \code{what}. \cr}
}
\value{
Either the result of the (evaluated) call of \code{what} or a
  \code{list} of the error message that resulted (if \code{soften = TRUE}).
}
\description{
Wrapping around \code{\link{do.call}}, this function allows
  for a simple "softening" where errors are returned as text, rather
  than causing a break in the encompassing pipeline.
}
\details{
As this is a basic wrapper on \code{\link{do.call}}, the basic 
  rules of its usage still apply: \cr
  If \code{quote = FALSE} (the default), then the arguments are evaluated
    in the calling environment and not in \code{envir}. \cr
  If \code{quote = TRUE}, each argument is \code{\link{quote}}d 
    so that the effect of argument evaluation is to remove the quotes
    leaving the original arguments unevaluated when the call is built. \cr
  The behavior of some functions, such as \code{\link{substitute}}, will 
    not be the same for functions evaluated using \code{\link{do.call}}
    and thus \code{soft_call} as if they were evaluated from the 
    interpreter. The precise semantics are currently undefined and subject 
    to change.
}
