% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{softmax}
\alias{softmax}
\title{Caclulate the softmax of a vector or matrix of values}
\usage{
softmax(x)
}
\arguments{
\item{x}{\code{numeric} vector or matrix}
}
\value{
The softmax of \code{x}.
}
\description{
Calculate the softmax (normalized exponential) of a vector
 of values or a set of vectors stacked rowwise.
}
