% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{step_chains}
\alias{step_chains}
\alias{propose_step}
\alias{eval_step}
\alias{take_step}
\title{Conduct a within-chain step of the ptMCMC algorithm}
\usage{
step_chains(i, cpts, inputs)

propose_step(i, cpts, inputs)

eval_step(i, cpts, prop_step, inputs)

take_step(cpts, prop_step, accept_step)
}
\arguments{
\item{i}{Integer iteration index.}

\item{cpts}{Matrix of changepoint locations across chains.}

\item{inputs}{Class \code{ptMCMC_inputs} list, containing the static inputs
for use within the ptMCMC algorithm.}

\item{prop_step}{Proposed step output from \code{propose_step}.}

\item{accept_step}{\code{logical} indicator of acceptance of each chain's
proposed step.}
}
\value{
\code{step_chains}: the initialized progress bar object.

\code{propose_step}: List of changepoints and log-likelihood values
  for the proposal.

\code{propose_step}: \code{logical} vector indicating if each 
  chain's proposal was accepted.

\code{take_step}: list of changepoints, log-likelihoods, and
  logical indicators of acceptance for each chain.
}
\description{
This set of functions steps the chains forward one iteration 
  of the within-chain component of the ptMCMC algorithm. \code{step_chains}
  is the main function, comprised of a proposal (made by \code{prop_step}),
  an evaluation of that proposal (made by \code{eval_step}), and then an 
  update of the configuration (made by \code{take_step}).
}
