% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{summarize_cps}
\alias{summarize_cps}
\title{Summarize the change point estimations}
\usage{
summarize_cps(cps, prob = 0.95)
}
\arguments{
\item{cps}{change point estimates}

\item{prob}{probability used for the interval}
}
\value{
model summary table for the change point locations
}
\description{
Function for summarizing change point estimations
}
