% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{summarize_etas}
\alias{summarize_etas}
\alias{measure_eta_vcov}
\title{Summarize the eta distributions}
\usage{
summarize_etas(etas, control = TS_controls_list())

measure_eta_vcov(etas)
}
\arguments{
\item{etas}{Matrix of regressors (columns) across iterations of the 
ptMCMC (rows).}

\item{control}{Class \code{TS_controls} list, holding control parameters
for the Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls, generated by 
\code{\link{TS_controls_list}}.}
}
\value{
\code{summarize_etas}: table of summary statistics for chunk-level
  regressors.

\code{measure_eta_vcov}: variance-covariance matrix for chunk-level
  regressors.
}
\description{
\code{summarize_etas} calculates summary statistics for each
  of the chunk-level regressors and \code{measure_ets_vcov} generates
  the variance-covariance matrix for the regressors.
}
