% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{summarize_rhos}
\alias{summarize_rhos}
\alias{measure_rho_vcov}
\title{Summarize the rho distributions}
\usage{
summarize_rhos(rhos, control)

measure_rho_vcov(rhos)
}
\arguments{
\item{rhos}{Matrix of changepoint locations (columns) across iterations of 
the ptMCMC (rows) or \code{NULL} if no changepoints are in the model.}

\item{control}{Class \code{TS_controls} list, holding control parameters
for the Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls, generated by 
\code{\link{TS_controls_list}}.}
}
\value{
\code{summarize_rhos}: table of summary statistics for changepoint
  locations.

\code{measure_rho_vcov}: variance-covariance matrix for changepoint
  locations.
}
\description{
\code{summarize_rho} calculates summary statistics for each
  of the changepoint locations and \code{measure_rho_vcov} generates
  the variance-covariance matrix for the changepoint locations.
}
