% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_models.R
\name{summarize_sequential_TS}
\alias{summarize_sequential_TS}
\alias{summarize_etas}
\alias{measure_eta_vcov}
\alias{summarize_rhos}
\alias{measure_rho_vcov}
\title{Summarize the change point (rho) and regressor (eta) distributions 
  of a sequential time series model}
\usage{
summarize_etas(etas, TS)

measure_eta_vcov(etas)

summarize_rhos(rhos, TS)

measure_rho_vcov(rhos)
}
\arguments{
\item{etas}{\code{matrix} of regressors (columns) across iterations of the 
sampler (rows), as returned from \code{\link{est_regressors}}.}

\item{TS}{Time series model \code{list}.}

\item{rhos}{\code{matrix} of change point locations (columns) across
iterations of the sampler (rows) or \code{NULL} if no change points are 
in the model, as returned from \code{\link{est_change points}}.}
}
\value{
\code{summarize_etas}: table of summary statistics for chunk-level
    regressors including mean, median, mode, posterior interval, standard
    deviation, MCMC error, autocorrelation, and effective sample size for 
    each regressor. \cr \cr
  \code{measure_eta_vcov}: variance-covariance matrix for chunk-level
    regressors. \cr \cr
 \code{summarize_rhos}: table of summary statistics for change point
    locations including mean, median, mode, posterior interval, standard
    deviation, MCMC error, autocorrelation, and effective sample size for 
    each change point location. \cr \cr
  \code{measure_rho_vcov}: variance-covariance matrix for change 
    point locations.
}
\description{
\code{summarize_etas} calculates summary statistics for each
    of the chunk-level regressors. \cr \cr
  \code{measure_ets_vcov} generates the variance-covariance matrix for 
    the regressors. \cr \cr
 \code{summarize_rho} calculates summary statistics for each
    of the change point locations. \cr \cr
  \code{measure_rho_vcov} generates the variance-covariance matrix for the 
    change point locations.
}
