% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{summarize_TS}
\alias{summarize_TS}
\title{Summarize the Time Series model}
\usage{
summarize_TS(data, formula, weights, control, rho_dist, eta_dist)
}
\arguments{
\item{data}{Class \code{data.frame} object including [1] the time variable
(indicated in \code{control}), [2] the predictor variables (required by
\code{formula}) and [3], the multinomial response variable (indicated
in \code{formula}).}

\item{formula}{\code{formula} describing the continuous change. Any 
predictor variable included must also be a column in the
\code{data}.  Any (multinomial) response variable must also be a set of
columns in \code{data}.}

\item{weights}{Optional class \code{numeric} vector of weights for each 
document. Corresponds to the vector \strong{\eqn{v}} in the math 
description.}

\item{control}{Class \code{TS_controls} list, holding control parameters
for the Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls, generated by 
\code{\link{TS_controls_list}}.}

\item{rho_dist}{List of saved data objects from the ptMCMC estimation of
changepoint locations (unless \code{nchangepoints} is 0, then 
\code{NULL}).}

\item{eta_dist}{Matrix of draws (rows) from the marginal posteriors of the 
coefficients across the segments (columns).}
}
\value{
Set of results from TS.
}
\description{
Calculate relevant summaries and package the output as a
  \code{TS_fit}-class object.
}
