% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{swap_chains}
\alias{swap_chains}
\title{Conduct a set of among-chain swaps for the ptMCMC algorithm}
\usage{
swap_chains(steps, inputs, ids)
}
\arguments{
\item{steps}{Chain configuration after within-temperature steps.}

\item{inputs}{Class \code{ptMCMC_inputs} list, containing the static inputs
for use within the ptMCMC algorithm.}

\item{ids}{The vector of integer chain ids.}
}
\value{
List of updated changepoints, log-likelihoods, and chain ids, as 
  well as a vector of acceptance indicators for each swap.
}
\description{
This function handles the among-chain swapping based on 
  temperature and likelihood differential.
}
