% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_models.R
\name{topicmodels_LDA}
\alias{topicmodels_LDA}
\title{Latent Dirichlet Allocation Linguistic Decomposition Analysis
  as conducted via the topicmodels package}
\usage{
topicmodels_LDA(LDA, method = "VEM", seeded = TRUE, ...)
}
\arguments{
\item{LDA}{A prepared (via \code{\link{prep_LDA_models}} LDA model
\code{list}.}

\item{method}{Fitting routine used in 
\code{\link[topicmodels]{topicmodels::LDA}}. Currenlty, only
\code{"VEM"} and \code{"Gibbs"} are supported.}

\item{seeded}{\code{logical} indicator of if the LDA should be a seeded
replicate.}

\item{...}{Additional arguments to be passed to 
\code{\link[topicmodels]{topicmodels::LDA}} as a \code{control} input.}
}
\value{
\code{LDA} \code{list}.
}
\description{
Fit the standard LDATS LDA model (a true Latent Dirichlet
  Allocation) using \code{\link[topicmodels]{topicmodels::LDA}}
  (Grun and Hornik 2011). \cr 
  Default methodology is the Variational
  Expectation Maximization routine (VEM) as described by 
  Blei \emph{et al.} (2003) and implemented by Grun and Hornik (2011). \cr
  If the model is defined to only fit one topic, \code{\link{identity_LDA}}
  is used by default.
}
\references{
Blei, D. M., A. Y. Ng, and M. I. Jordan. 2003. Latent Dirichlet
  Allocation. \emph{Journal of Machine Learning Research} 
  \strong{3}:993-1022.
  \href{http://jmlr.csail.mit.edu/papers/v3/blei03a.html}{link}.

  Grun B. and K. Hornik. 2011. topicmodels: An R Package for Fitting Topic
  Models. \emph{Journal of Statistical Software} \strong{40}:13.
  \href{https://www.jstatsoft.org/article/view/v040i13}{link}.
}
