% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{TS}
\alias{TS}
\alias{prep_TS_models}
\title{Conduct a Bayesian compositional Time Series analysis}
\usage{
TS(LDAs, data, formulas = ~1, nchangepoints = 0, timename = "time",
  weights = NULL, control = list())

prep_TS_models(LDAs, data, formulas = ~1, nchangepoints = 0,
  timename = "time", weights = NULL, control = list())
}
\arguments{
\item{data}{\code{data.frame} including [1] the time variable (indicated 
in \code{timename}), [2] the predictor variables (required by
\code{formula}) and [3], the compositional response variable (indicated
in \code{formula}). \cr \cr
Note that the response variables should be formatted as a 
\code{data.frame} object named as indicated by the \code{response} entry
in the \code{control} list, such as \code{gamma} for a standard TS 
analysis on LDA output.}

\item{nchangepoints}{\code{integer} corresponding to the number of 
change points to include in the model. 0 is a valid input (corresponding
to no change points, so a singular time series model), and the current 
implementation can reasonably include up to 6 change points. The 
number of change points is used to dictate the segmentation of the 
time series into chunks fit with separate models dictated by 
\code{formula}.}

\item{timename}{\code{character} element indicating the time variable
used in the time series. Defaults to \code{"time"}. The variable must be
integer-conformable or a \code{Date}. If the variable named
is a \code{Date}, the input is converted to an integer, resulting in the
timestep being 1 day, which is often not desired behavior.}

\item{weights}{Optional class \code{numeric} vector of weights for each 
document. Defaults to \code{NULL}, translating to an equal weight for
each document. When using \code{multinom_TS} in a standard LDATS 
analysis, it is advisable to weight the documents by their total size,
as the result of \code{\link[topicmodels]{LDA}} is a matrix of 
proportions, which does not account for size differences among documents.
For most models, a scaling of the weights (so that the average is 1) is
most appropriate, and this is accomplished using \code{document_weights}.}

\item{control}{A \code{list} of parameters to control the fitting of the
Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls. Values not input assume defaults set by 
\code{\link{TS_control}}.}

\item{formula}{\code{\link[stats]{formula}} defining the regression between
the change points. Any predictor variable included must also be a column 
in \code{data} and any (compositional) response variable must be a set 
of columns in \code{data}.}

\item{control}{A \code{list} of parameters to control the fitting of the
Time Series model. Values not input assume defaults set by 
\code{\link{TS_control}}.}
}
\value{
\code{TS}: class \code{TS_set} \code{list} of both selected and all
    results from \code{\link{sequential_TS}} applied for each model on each
    LDA model input as well as the control \code{list} used to fit the 
    model. \cr \cr
  \code{prep_TS_models}: \code{list} of \code{list}s, each of which is a
    preliminary model object for a Time Series model fit.
}
\description{
\code{TS} is the main interface function for the LDATS application
    of Bayesian change point Time Series analyses (Christensen 
    \emph{et al.} 2018). \cr \cr
  \code{prep_TS_models} pre-prepares the TS model objects for simpler 
    use within the subfunctions.

 @details This model extends the approach of Western and Kleykamp (2004;
  see also Ruggieri 2013) to compositional response data using
  softmax regression (Ripley 1996, Venables and Ripley 2002, Bishop 2006) 
  or simplical geometry ()
  via a generalized linear modeling approach (McCullagh and Nelder 1989).
  The models can be fit using a range of methods, but available procedures
  are different flavors of parallel tempering Markov Chain Monte Carlo
  (ptMCMC) methods (Earl and Deem 2005)
}
