% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{TS_control}
\alias{TS_control}
\title{Create the controls list for the Time Series model}
\usage{
TS_control(model = sequential_TS, model_args = list(),
  response = "multinom", response_args = list(),
  method = "ldats_classic", method_args = ldats_classic_control(),
  summary_prob = 0.95, measurer = AIC, measurer_args = list(),
  selector = which.min, selector_args = list(), soften = TRUE,
  quiet = FALSE)
}
\arguments{
\item{response}{\code{character} element indicating the response variable 
used in the time series. \cr \cr
Must have a corresponding \code{<response>_TS} function.}

\item{method}{\code{function} used to drive the sampler of the TS
models; \code{method} defines and operates the computational procedure.
\cr \cr
Current pre-built options include \code{\link{ldats_classic}}.}

\item{method_args}{\code{list} of (named) arguments to be used in 
\code{method} via \code{\link{do.call}}. 
\cr \cr
Could be managed via a \code{<method>_control} function like
\code{\link{ldats_classic_control}}.}

\item{summary_prob}{Probability used for summarizing the posterior 
distributions (via the highest posterior density interval, see
\code{\link[coda]{HPDinterval}}).}

\item{measurer}{\code{function} used in evaluation of the TS
models; \code{measurer} creates a value for each model.}

\item{measurer_args}{\code{list} of (named) arguments to be used in 
\code{measurer} via \code{\link{do.call}}.}

\item{selector}{\code{function} usde in evaluation of the TS
models; \code{selector} operates on the values to choose the models.}

\item{selector_args}{\code{list} of (named) arguments to be used in 
\code{selector} via \code{\link{do.call}}.}

\item{soften}{\code{logical} indicator of whether the model should error 
softly or if errors should trigger a full-stop to the pipeline.}

\item{quiet}{\code{logical} indicator of whether the model should run 
quietly (if \code{FALSE}, a progress bar and notifications are printed).}
}
\value{
\code{list}, with named elements corresponding to the arguments.
}
\description{
This function provides a simple creation and definition of a
  list used to control the time series model fit occurring within 
  \code{\link{TS}}.
}
\examples{
  TS_control()

}
