% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{TS_controls_list}
\alias{TS_controls_list}
\title{Create the controls list for the Time Series model}
\usage{
TS_controls_list(memoise = TRUE, response = "gamma",
  timename = "newmoon")
}
\arguments{
\item{memoise}{Logical indicator of whether the multinomial functions 
should be memoised (via ). Memoisation 
happens to both \code\link{{multinom_TS}} and 
\code\link{{multinom_TS_chunk}}}

\item{response}{Character indicating the response variable used in the time
series.}

\item{timename}{Character indicating the time variable used in the time 
series.}
}
\value{
List of class \code{TS_controls}.
}
\description{
This function provides a simple creation and definition of a
  list used to control the time series model fit occurring within 
  \code{\link{TS}}.
}
