% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_on_LDA.R
\name{TS_on_LDA}
\alias{TS_on_LDA}
\title{Conduct a set of Time Series analyses on a set of LDA models}
\usage{
TS_on_LDA(LDA_models, document_covariate_table, formulas = ~1,
  nchangepoints = 0, weights = NULL, control = TS_controls_list())
}
\arguments{
\item{LDA_models}{List of LDA models (class \code{LDA_set}) or a singular
LDA model (class \code{LDA}).}

\item{document_covariate_table}{Document covariate table (rows:
documents (\code{M}), columns: time index and covariate options). 
Every model needs a covariate to describe the time value for each
document (in whatever relevant units), whose name in the table is input
via \code{timename}, that dictates the application of the changepoints. 
In addition, the table needs to include as columns all covariates named 
within the specific models described via the argument \code{formula} 
(if desired). Must be a conformable to a data table.}

\item{formulas}{Vector of \code{formula}(s) for the continuous change. Any 
predictor variable included in a formula must also be a column in the
\code{document_covariate_table}. Each element (formula) in the vector
is evaluated for each number of change points and each LDA model.}

\item{nchangepoints}{Vector of integers corresponding to the number of 
change points to include in the model. 0 is a valid input (corresponding
to no change points, so a singular time series model), and the current 
implementation can reasonably include up to 6 change points. Each element 
(number of change points) in the vector is used to dictate the
segementation of the data  for each continuous model and each LDA model.}

\item{weights}{Optional class \code{numeric} vector of weights for each 
document. Corresponds to the vector \strong{\eqn{v}} in the math 
description.}

\item{control}{Class \code{TS_controls} list, holding control parameters
for the Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls, generated by 
\code{\link{TS_controls_list}}.}
}
\value{
Class \code{TS_on_LDA} list of results from \code{\link{TS}} 
  applied for each model on each LDA model input.
}
\description{
This is a wrapper function that expands the main Time Series
  analyses function (\code{\link{TS}}) across the LDA models and the Time
  Series models, with respect to both continuous time formulas and the 
  number of discrete changepoints. This function allows direct passage of
  the control parameters for the parallel tempering MCMC through to the 
  main Time Series function, \code{\link{TS}}, via the 
  \code{ptMCMC_controls} argument.
}
