% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{update_list}
\alias{update_list}
\title{Update a list's elements}
\usage{
update_list(orig_list = list(), ...)
}
\arguments{
\item{orig_list}{\code{list} to be updated with \code{...}.}

\item{...}{Named elements to update in \code{orig_list}}
}
\value{
Updated \code{list}.
}
\description{
Update a list with new values for elements
}
\examples{
 orig_list <- list(a = 1, b = 3, c = 4)
 update_list(orig_list)
 update_list(orig_list, a = "a")
 update_list(orig_list, a = 10, b = NULL)

}
