% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_TS.R
\name{verify_changepoint_locations}
\alias{verify_changepoint_locations}
\title{Verify the change points of a multinomial time series model}
\usage{
verify_changepoint_locations(data, changepoints = NULL,
  timename = TS_controls_list()$timename)
}
\arguments{
\item{data}{Class \code{data.frame} object including the predictor and 
response variables.}

\item{changepoints}{Numeric vector indicating locations of the change 
points, noted as \ifelse{html}{
    \out{<i><b>&rho;</b><sub>m<sub>2</sub><sup>r</sup></sub></i>}
    }{\eqn{\boldsymbol{\rho}_{m^r_2}}} in the math description.}

\item{timename}{\code{character} name of the column in the 
\code{document_covariate_table} that contains the time index to use
for assignment of the changepoints (corresponding to the vector 
\ifelse{html}{\out{<b><i>t</i></b>}}{\eqn{\mathbf{t}}} 
in the mathematical description of the model).}
}
\value{
Logical indicator of the check passing \code{TRUE} or failing
  \code{FALSE}.
}
\description{
Verify that a time series can be broken into a set 
  of chunks based on input changepoints.
}
