% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_functions.R
\name{AIC.LDA}
\alias{AIC.LDA}
\title{Generalization of the AIC function to work on LDA topic models}
\usage{
\method{AIC}{LDA}(object, ..., k = 2)
}
\arguments{
\item{object}{an LDA topic model}

\item{...}{additional arguments to be passed to subfunctions}

\item{k}{penalty per df}
}
\value{
Named (AIC or AICc) value.
}
\description{
Using the internal topicmodels calculations of likelihood and 
  df
}
\examples{
  data(rodents)
  lda_data <- dplyr::select(rodents, -c(newmoon, date, plots, traps))
  r_LDA <- LDATS::LDA(data = lda_data, ntopics = 2, nseeds = 2)
  AIC(r_LDA[[1]])
  AIC(r_LDA[[2]])
}
