% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{AIC.LDA}
\alias{AIC.LDA}
\title{Determine the AIC of a Linguistic Decomposition Analysis
  model}
\usage{
\method{AIC}{LDA}(object, ..., k = 2)
}
\arguments{
\item{object}{Class \code{LDA} object to be evaluated.}

\item{...}{Not used, simply included to maintain method compatibility.}

\item{k}{Per-parameter numeric penalty.}
}
\value{
AIC of the model.
}
\description{
Convenience function to extract and format the AIC
  of a \code{LDA}-class object fit by \code{\link{LDA_call}}.
}
