% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{annual_hist}
\alias{annual_hist}
\title{Histogram showing percentage of MCMC samples that contained
a changepoint in a given year.}
\usage{
annual_hist(results, year_continuous)
}
\arguments{
\item{results}{results}

\item{year_continuous}{year_continuous}
}
\description{
Histogram showing percentage of MCMC samples that contained
a changepoint in a given year.
}
