% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_TS.R
\name{check_changepoints}
\alias{check_changepoints}
\title{Check that a set of change point locations is proper}
\usage{
check_changepoints(changepoints = NULL)
}
\arguments{
\item{changepoints}{Change point locations to evaluate.}
}
\description{
Check that the change point locations are \code{numeric}
  and conformable to \code{interger} values.
}
