% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_control}
\alias{check_control}
\title{Check that a control list is proper}
\usage{
check_control(control, eclass = "list")
}
\arguments{
\item{control}{Control list to evaluate.}

\item{eclass}{Expected class of the list to be evaluated.}
}
\value{
an error message is thrown if the input is improper, otherwise 
  \code{NULL}.
}
\description{
Check that a list of controls is of the right class.
}
\examples{
 check_control(list())

}
