% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_document_term_table}
\alias{check_document_term_table}
\title{Verify that document term table is proper}
\usage{
check_document_term_table(document_term_table)
}
\arguments{
\item{document_term_table}{Table of observation count data (rows: 
documents (\eqn{M}), columns: terms (\eqn{V})).}
}
\description{
Verify that the table of observations is conformable to
  a matrix of integers.
}
