% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{check_formula}
\alias{check_formula}
\title{Verify that a formula is proper}
\usage{
check_formula(formula, data)
}
\arguments{
\item{formula}{Formula to evaluate.}

\item{data}{Class \code{data.frame} object including [1] the time variable
(indicated in \code{control}), [2] the predictor variables (required by
\code{formula}) and [3], the multinomial response variable (indicated
in \code{control}).}
}
\value{
Nothing.
}
\description{
Verify that the formula is actually a formula and that the
  response and predictor variabless are all included in the data.
}
