% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{check_formula}
\alias{check_formula}
\title{Verify that formula vector is proper}
\usage{
check_formula(formula, document_covariate_table)
}
\arguments{
\item{formula}{Vector of the formulas to evaluate.}

\item{document_covariate_table}{Document covariate table used to evaluate
the availability of the data required by the formula inputs.}
}
\value{
Input \code{formula} properly formatted.
}
\description{
Verify that the vector of formulas is actually formatted
  as a vector formula objects and that the predictor variables are all 
  included in the document covariate table.
}
