% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{check_LDA_models}
\alias{check_LDA_models}
\title{Verify that LDA model input is proper}
\usage{
check_LDA_models(LDA_models)
}
\arguments{
\item{LDA_models}{List of LDA models or singular LDA model to evaluate.}
}
\value{
\code{LDA_models} as an \code{LDA_list}-class object.
}
\description{
Verify that the \code{LDA_models} input is, in fact, LDA 
  models or a singular LDA model. If there is only one model, convert it
  from a singular class \code{LDA} list to a length-1 class \code{LDA_list}
  class and return it.
}
