% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_on_LDA.R
\name{check_nchangepoints}
\alias{check_nchangepoints}
\title{Check that nchangepoints vector is proper}
\usage{
check_nchangepoints(nchangepoints)
}
\arguments{
\item{nchangepoints}{Vector of the number of changepoints to evaluate.}
}
\value{
An error message is thrown if \code{nchangepoints} is not proper,
else \code{NULL}.
}
\description{
Check that the vector of numbers of changepoints is
conformable to integers greater than 1.
}
\examples{
  check_nchangepoints(0)
  check_nchangepoints(2)

}
