% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{check_seeds}
\alias{check_seeds}
\title{Verify that nseeds value or seeds vector is proper}
\usage{
check_seeds(seeds)
}
\arguments{
\item{seeds}{Value of the number of random seeds to evaluate.}
}
\value{
Nothing.
}
\description{
Verify that the vector of numbers of seeds is conformable to
  integers greater than 1.
}
