% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_seeds}
\alias{check_seeds}
\title{Check that nseeds value or seeds vector is proper}
\usage{
check_seeds(seeds)
}
\arguments{
\item{nseeds}{\code{integer} number of seeds (replicate starts) to use for 
each value of \code{topics} in the LDAs. Must be conformable to a  
positive \code{integer} value.}
}
\description{
Check that the vector of numbers of seeds is conformable to
  integers greater than 0.
}
