% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{check_topics}
\alias{check_topics}
\title{Verify that topics vector is proper}
\usage{
check_topics(topics)
}
\arguments{
\item{topics}{Vector of the number of topics to evaluate.}
}
\value{
Nothing.
}
\description{
Verify that the vector of numbers of topics is conformable to
  integers greater than 1.
}
