% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_topics}
\alias{check_topics}
\title{Check that topics vector is proper}
\usage{
check_topics(topics)
}
\arguments{
\item{topics}{Vector of the number of topics to evaluate for each model.
Must be conformable to \code{integer} values.}
}
\description{
Check that the vector of numbers of topics is conformable to
  integers greater than 1.
}
