% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{count_trips}
\alias{count_trips}
\title{Number of times the particle went from hottest chain to the coldest one,
indicating good mixing.  Should probably be in the mid-hundreds or low
thousands if we want to be really confident about the results.}
\usage{
count_trips(results)
}
\arguments{
\item{results}{results}
}
\description{
Number of times the particle went from hottest chain to the coldest one,
indicating good mixing.  Should probably be in the mid-hundreds or low
thousands if we want to be really confident about the results.
}
