% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_methods.R
\name{count_trips}
\alias{count_trips}
\title{Count trips of the ptMCMC particles in a classic_ldats estimation}
\usage{
count_trips(ids)
}
\arguments{
\item{ids}{\code{matrix} of identifiers of the particles in each chain for 
each iteration of the ptMCMC algorithm (rows: chains, 
columns: iterations).}
}
\value{
\code{list} of [1] \code{vector} of within particle trip counts 
  (\code{$trip_counts}), and [2] \code{vector} of within-particle average 
  trip rates (\code{$trip_rates}).
}
\description{
Count the full trips (from one extreme temperature chain to
  the other and back again; Katzgraber \emph{et al.} 2006) for each of the
  ptMCMC particles, as identified by their id on initialization.
  \cr \cr
  This function was designed to work within \code{\link{TS}} and process
  the output of \code{\link{est_changepoints}} as a component of 
  \code{\link{diagnose_ptMCMC}}, but has been generalized
  and would work with any output from a ptMCMC as long as \code{ids}
  is formatted properly.
}
\references{
Katzgraber, H. G., S. Trebst, D. A. Huse. And M. Troyer. 2006. 
  Feedback-optimized parallel tempering Monte Carlo. \emph{Journal of 
  Statistical Mechanics: Theory and Experiment} \strong{3}:P03018
  \href{https://bit.ly/2LICGXh}{link}.
}
