% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_tempering_functions.R
\name{count_trips}
\alias{count_trips}
\title{Count trips}
\usage{
count_trips(ids)
}
\arguments{
\item{ids}{Matrix of identifiers of the initial temperature for particles}
}
\value{
list of counts and rates of trips by temperature
}
