% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_changepoint_functions.R
\name{cp_models}
\alias{cp_models}
\title{Run a suite of changepoint models for the same data set}
\usage{
cp_models(data = NULL, dates = NULL, ntopics = NULL, SEED = NULL,
  weights = NULL, maxit = NULL, maxcps = NULL)
}
\arguments{
\item{data}{data set to use}

\item{ntopics}{number of topics to use in the baseline LDA}

\item{SEED}{seed to use in the baseline LDA}

\item{weights}{weights to use through time for the data points}

\item{maxit}{maximum iterations}

\item{time}{time of each sample (row in data)}

\item{maxcp}{maximum number of change points to use in the models}
}
\value{
list of change point model results
}
\description{
Run a suite of changepoint models for the same data set
}
\author{
Juniper Simonis
}
