% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.r
\name{create_rodent_table}
\alias{create_rodent_table}
\title{Create rodent species table}
\usage{
create_rodent_table(periods = 1:436, plots = c(2, 4, 8, 11, 12, 14, 17, 22),
  species = c("BA", "DM", "DO", "DS", "NA", "OL", "OT", "PB", "PE", "PF",
  "PH", "PI", "PL", "PM", "PP", "RF", "RM", "RO", "SF", "SH", "SO"))
}
\arguments{
\item{periods}{vector of trapping periods to be included}

\item{plots}{vector of plot numbers to be included}

\item{species}{vector of species codes to be included}
}
\value{
Table of species counts per period
}
\description{
Processes rodent capture data so it can be used for LDA analysis
}
\examples{
create_rodent_table()

}
