% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_on_LDA.R
\name{expand_TS}
\alias{expand_TS}
\title{Expand the TS models needed across the factorial combination of
  LDA models, continuous formulas, and number of change points}
\usage{
expand_TS(LDA_models, formulas, nchangepoints)
}
\arguments{
\item{LDA_models}{\code{LDA_list}-class object of LDA models.}

\item{formulas}{Vector of the continuous formulas.}

\item{nchangepoints}{Vector of the number of changepoints.}
}
\value{
Expanded table of the three values: [1] the LDA model (indicated
  as a numeric element reference to the \code{LDA_list} object), [2] the 
  continuous formula, and [3] the number of changepoints.
}
\description{
Expand the completely crossed combination of model inputs: 
  LDA model results, continuous formulas
}
