% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_models.R
\name{identity_LDA}
\alias{identity_LDA}
\title{Identity Linguistic Decomposition Analysis}
\usage{
identity_LDA(LDA)
}
\arguments{
\item{LDA}{A prepared (via \code{\link{prep_LDA_models}} LDA model
\code{list}.}
}
\value{
\code{LDA} \code{list} with components (many of which are 
 placeholders):
 \describe{
   \item{alpha}{parameter estimate.}
   \item{beta}{parameter estimate.}
   \item{terms}{\code{character} \code{vector} of term names.}
   \item{document_topic_matrix}{estimated latent topic compositions.}
   \item{test_document_topic_matrox}{estimated latent topic compositions 
     of the test data (not presently available for usage).}
   \item{log_likelihood}{model log likelihood.}
   \item{data}{data object used to fit the LDA model.}
   \item{data_subset}{number of the data subset from the whole data set.}
   \item{topics}{\code{integer} number of topics in the model.}
   \item{replicat}{\code{integer} replicate number.}
   \item{control}{\code{list} of controls used to fit the model. See
     \code{\link{LDA_control}}.}
 }
}
\description{
This function acts as an "identity" model, wherein the 
  output is functionally the input. This allows for "single-topic" models
  that do not actually decompose the data to be included in the model set.
}
