% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_functions.R
\name{LDA}
\alias{LDA}
\title{A simple parallel wrapper to the topicmodels LDA function}
\usage{
LDA(data, ntopics = 2, nseeds = 1, ncores = 1, ...)
}
\arguments{
\item{data}{Table of integer data (species counts by period)}

\item{ntopics}{set of topics to evaluate}

\item{nseeds}{number of seeds to use}

\item{ncores}{Integer number of cores to use}

\item{...}{additional arguments to be passed to the LDA function}
}
\value{
List of LDA models
}
\description{
Runs each of the number of input topics for the number of 
  seeds
}
\examples{
  data(rodents)
  lda_data <- dplyr::select(rodents, 
                            -c(newmoonnumber, newmoondate, nplots, ntraps))
  r_LDA <- LDATS::LDA(data = lda_data, ntopics = 2:5, nseeds = 2, 
                      ncores = 8)
}
