% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA}
\alias{LDA}
\alias{check_LDA}
\alias{prepare_LDA}
\alias{run_LDA}
\alias{LDA_call}
\alias{LDA_msg}
\alias{package_LDA}
\alias{select_LDA}
\alias{measure_LDA}
\alias{LDA_control}
\title{Run a set of Linguistic Decomposition Analysis models}
\usage{
LDA(data, topics = 2, replicates = 1, control = list())

check_LDA(topics = 2, replicates = 1, control = list())

prepare_LDA(data, topics = 2, replicates = 1, control = list())

run_LDA(LDAs)

LDA_call(LDA)

LDA_msg(LDA)

package_LDA(LDAs)

select_LDA(LDAs)

measure_LDA(LDAs)

LDA_control(model = topicmodels_LDA, model_args = list(method = "VEM",
  seeded = TRUE), measurer = AIC, measurer_args = list(NULL),
  selector = which.min, selector_args = list(NULL), nsubsets = 1,
  subset_rule = NULL, soften = TRUE, quiet = FALSE, ...)
}
\arguments{
\item{data}{Any of the data structures allowable for LDATS analyses:
\code{matrix} or \code{data.frame} document term table, 
\code{list} of document term and covariate tables, a \code{list} of 
training and test sets of the two tables, or a \code{list} of multiple 
replicate splits of training and test sets of the two tables. \cr
See \code{\link{conform_data}}, which is used to ensure data structure
validity for the desired model.}

\item{topics}{Vector of the number of topics to evaluate for each model.
Must be conformable to \code{integer} values.}

\item{replicates}{Number of replicate starts to use for each 
value of \code{topics}. Must be conformable to \code{integer} value.}

\item{control}{A \code{list} of parameters to control the fitting of the
LDA model. Values not input assume defaults set by 
\code{\link{LDA_control}}.}

\item{LDA, LDAs}{model \code{list} (\code{LDA}) or a \code{list} of LDA 
model \code{list}s (\code{LDAs}).}

\item{model}{Main LDA \code{function}.}

\item{model_args}{\code{list} of (named) arguments to be used in 
\code{model} via \code{\link{LDA_call}}.}

\item{measurer}{\code{function} used in evaluation of the LDA
models; \code{measurer} creates a value for each model.}

\item{measurer_args}{\code{list} of (named) arguments to be used in 
\code{measurer} via \code{\link{do.call}}.}

\item{selector}{\code{function} usde in evaluation of the LDA
models; \code{selector} operates on the values to choose the models.}

\item{selector_args}{\code{list} of (named) arguments to be used in 
\code{selector} via \code{\link{do.call}}.}

\item{nsubsets}{Number of data subsets.}

\item{subset_rule}{\code{function} used to subset the data.}

\item{soften}{\code{logical} indicator of whether the model should error 
softly or if errors should trigger a full-stop to the pipeline.}

\item{quiet}{\code{logical} indicator of whether the model should run 
quietly (if \code{FALSE}, a progress bar and notifications are printed).}

\item{...}{Not passed along to the output, rather included to allow for
automated removal of unneeded controls.}
}
\value{
\code{LDA},\code{pacakage_LDA}: class \code{LDA_set} \code{list} of 
    both selected and all results from \code{\link{LDA_call}} applied for 
    each model on each data input(s) as well as the control \code{list} 
    used to fit the model. \cr \cr
  \code{prepare_LDA}: \code{list} of \code{list}s, each of which is a
    preliminary model object for an LDA model fit. \cr \cr
  \code{check_LDA}: an error message is thrown if any input is improper, 
    otherwise \code{NULL}.
  \code{LDA_control}: \code{list} of controls for the LDA model, with
    named elements corresponding to the arguments.
  \code{run_LDA}: \code{LDA_set} \code{list} of model results from all
    runs of a \code{<model>} function, such as 
    \code{\link{topicmodels_LDA}}. \cr \cr
  \code{LDA_call}: \code{LDA} \code{list} of model results from a single
    run of a \code{<model>} function, such as 
    \code{\link{topicmodels_LDA}}. \cr \cr
  \code{measure_LDA}: \code{vector} of values corresponding to the model
    evaluations. \cr \cr
  \code{select_LDA}: \code{list} of selected models' \code{list}s.
}
\description{
Conduct Linguistic Decomposition Analyses. \cr \cr
  \code{LDA} provides the main interface for Linguistic Decomposition 
    Analysis conducted within the LDATS application of (Christensen 
    \emph{et al.} 2018). \cr \cr
  \code{prepare_LDA} pre-prepares the LDA model objects for simpler 
    use within the subfunctions. \cr \cr 
  \code{check_LDA} ensures that the inputs are proper. 
    See \code{\link{check_topics}}, \code{\link{check_replicates}},
    and \code{\link{check_control}} for specifics. \cr \cr
  \code{LDA_control} defines and creates the control list used to fit 
    the LDA model. \cr \cr 
  \code{run_LDA} runs (via \code{\link{LDA_call}}) all LDA models
    as set up by \code{prep_LDA_models}. \cr \cr
  \code{LDA_call} runs (via \code{\link{do.call}}) a single LDA model
    as set up by \code{prep_LDA_models}. \cr \cr
  \code{LDA_msg} produces a model-running message if desired. \cr \cr
  \code{measure_LDA} determines the fit value used to select among the 
    models. \cr \cr
  \code{select_LDA} chooses the best model(s) of interest based on their
    measured values and the selector function. \cr \cr
  \code{package_LDA} sets the class and names the elements of the results
    \code{list} from \code{\link{LDA_call}} applied to the 
    combination of TS models requested for the data input(s).
}
\details{
For a (potentially subset) dataset consisting of counts of words 
  across multiple documents in a corpus, 
  \enumerate{
    \item Conduct multiple Linguistic Decomposition Analysis (LDA) models 
      (e.g., Latent Dirichlet Allocation using the Variational Expectation
      Maximization (VEM) algorithm; Blei \emph{et al.} 2003, Grun and
      Hornik 2011),
    \item Select from the LDA model results to pick those used in the Time
      Series (TS) models, and
    \item Package the results.
  }
}
\references{
Blei, D. M., A. Y. Ng, and M. I. Jordan. 2003. Latent Dirichlet
  Allocation. \emph{Journal of Machine Learning Research} 
  \strong{3}:993-1022.
  \href{http://jmlr.csail.mit.edu/papers/v3/blei03a.html}{link}.

  Christensen, E., D. J. Harris, and S. K. M. Ernest. 2018.
  Long-term community change through multiple rapid transitions in a 
  desert rodent community. \emph{Ecology} \strong{99}:1523-1529. 
  \href{https://doi.org/10.1002/ecy.2373}{link}.

  Grun B. and K. Hornik. 2011. topicmodels: An R Package for Fitting Topic
  Models. \emph{Journal of Statistical Software} \strong{40}:13.
  \href{https://www.jstatsoft.org/article/view/v040i13}{link}.
}
