% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA_control}
\alias{LDA_control}
\title{Create the controls list for the Time Series model}
\usage{
LDA_control(model = topicmodels_LDA, model_args = list(method = "VEM",
  seeded = TRUE), measurer = AIC, measurer_args = list(NULL),
  selector = which.min, selector_args = list(NULL), nsubsets = 1,
  subset_rule = NULL, soften = TRUE, quiet = FALSE, ...)
}
\arguments{
\item{model}{Main LDA \code{function}.}

\item{model_args}{\code{list} of (named) arguments to be used in 
\code{model} via \code{\link{LDA_call}}.}

\item{measurer}{\code{function} used in evaluation of the LDA
models; \code{measurer} creates a value for each model.}

\item{measurer_args}{\code{list} of (named) arguments to be used in 
\code{measurer} via \code{\link{do.call}}.}

\item{selector}{\code{function} usde in evaluation of the LDA
models; \code{selector} operates on the values to choose the models.}

\item{selector_args}{\code{list} of (named) arguments to be used in 
\code{selector} via \code{\link{do.call}}.}

\item{nsubsets}{Number of data subsets.}

\item{subset_rule}{\code{function} used to subset the data.}

\item{soften}{\code{logical} indicator of whether the model should error 
softly or if errors should trigger a full-stop to the pipeline.}

\item{quiet}{\code{logical} indicator of whether the model should run 
quietly (if \code{FALSE}, a progress bar and notifications are printed).}

\item{...}{Not passed along to the output, rather included to allow for
automated removal of unneeded controls.}
}
\value{
\code{list}, with named elements corresponding to the arguments.
}
\description{
This function provides a simple creation and definition of a
  list used to control the LDA model fit occurring within 
  \code{\link{LDA}}.
}
\examples{
  LDA_control()

}
