% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA_set}
\alias{LDA_set}
\alias{check_LDA_set_inputs}
\title{Run a set of Latent Dirichlet Allocation models}
\usage{
LDA_set(document_term_table, topics = 2, nseeds = 1,
  control = LDA_controls_list())

check_LDA_set_inputs(document_term_table, topics, nseeds, control)
}
\arguments{
\item{document_term_table}{Table of observation count data (rows: 
documents (\eqn{M}), columns: terms (\eqn{V})). May be a class 
\code{matrix} or \code{data.frame} but must be conformable to
a code of integers. This table is a document-level summary of the data 
noted as 
\ifelse{html}{\out{<b><i>w</i></b>}}{\eqn{\mathbf{w}}}
(the word-level topic identity) in the math description.}

\item{topics}{Vector of the number of topics to evaluate for each model
(\ifelse{html}{\out{<i>k<sub>m<sub>1</sub></sub></i>}}{\eqn{k_{m_1}}} 
topics for a model \ifelse{html}{\out{<i>m<sub>1</sub></i>}}{\eqn{m_1}}).}

\item{nseeds}{Integer number of seeds (replicate starts) to use for each 
value of \code{topics}.}

\item{control}{Class \code{LDA_controls} list of control parameters to be
used in \code{LDA} (note that \code{seed} will be overwritten).}
}
\value{
List (class: \code{LDA_set}) of LDA models (class: \code{LDA_VEM}).
}
\description{
For a given corpus \eqn{D} (dataset consisting of counts of 
  words across \eqn{M} documents), conduct multiple Latent Dirichlet 
  Allocation (LDA) models (using the Variational Expectation 
  Maximization (VEM) algorithm; Blei \emph{et al.} 2003) to account for [1]  
  uncertainty in the number of latent topics and [2] the impact of initial
  values in the estimation procedure. 

\code{LDA_set} is a list wrapper of \code{\link[topicmodels]{LDA}}
  in the \code{topicmodels} package (Grun and Hornik 2011).

\code{check_LDA_set_inputs} checks that all of the inputs 
  are proper for \code{LDA_set} (that the table of observations is 
  conformable to a matrix of integers, the number of topics is an integer, 
  the number of seeds is an integer and the controls list is proper).
}
\examples{
\dontrun{
  data(rodents)
  lda_data <- rodents$document_term_table
  r_LDA <- LDA_set(lda_data, topics = 2, nseeds = 2)                         
}

}
\references{
Blei, D. M., A. Y. Ng, and M. I. Jordan. 2003. Latent Dirichlet
  Allocation. \emph{Journal of Machine Learning Research} 
  \strong{3}:993-1022.
  \href{http://jmlr.csail.mit.edu/papers/v3/blei03a.html}{link}.

  Grun B. and K. Hornik. 2011. topicmodels: An R Package for Fitting Topic
  Models. \emph{Journal of Statistical Software} \strong{40}:13.
  \href{https://www.jstatsoft.org/article/view/v040i13}{link}.
}
