% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA_set}
\alias{LDA_set}
\title{Run a set of Latent Dirichlet Allocation models}
\usage{
LDA_set(document_term_table, topics = 2, nseeds = 1,
  control = LDA_controls_list())
}
\arguments{
\item{document_term_table}{Table of observation count data (rows: 
documents (\eqn{M}), columns: terms (\eqn{V})). May be a class 
\code{matrix} or \code{data.frame} but must be conformable to
a code of integers. This table is a document-level summary of the data 
noted as \eqn{w} (the word-level topic identity) in the math description.}

\item{topics}{Vector of the number of topics to evaluate.}

\item{nseeds}{Integer number of seeds (replicate starts) to use for each 
value of \code{topics}.}

\item{control}{Class \code{LDA_controls} list of control parameters to be
used in \code{LDA} (note that \code{seed} will be overwritten).}
}
\value{
List (class: \code{LDA_set}) of LDA models (class: \code{LDA}).
}
\description{
For a given dataset (counts of words across several 
  documents), conduct multiple Latent Dirichlet Allocation (LDA) models
  (using the Variational Expectation Maximization (VEM) algorithm; Blei  
  \emph{et al.}) to account for [1] uncertainty in the number of latent 
  topics and [2] the impact of intial values in the estimation procedure. 

This function is a wrapper of the \code{LDA} function
  in the \code{topicmodels} package (Grun and Hornik 2011).
}
\examples{
\dontrun{
  data(rodents)
  lda_data <- rodents$document_term_table
  r_LDA <- LDA_set(lda_data, topics = 2, nseeds = 2)                         
}

}
\references{
Blei, D. M., A. Y. Ng, and M. I. Jordan. 2003. Latent Dirichlet
  Allocation. \emph{Journal of Machine Learning Research} 
  \strong{3}:993-1022.
  \href{http://jmlr.csail.mit.edu/papers/v3/blei03a.html}{link}.

  Grun B. and K. Hornik. 2011. topicmodels: An R Package for Fitting Topic
  Models. \emph{Journal of Statistical Software} \strong{40}:13.
  \href{https://www.jstatsoft.org/article/view/v040i13}{link}.
}
