% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDATS_functions.R
\name{LDA_TS}
\alias{LDA_TS}
\title{Two-stage LDA-time series analysis}
\usage{
LDA_TS(document_term_matrix = NULL, document_covariate_matrix = NULL,
  formula = "1", nchangepoints = 1, ...)
}
\arguments{
\item{document_term_matrix}{matrix of documents (rows) by terms (columns)}

\item{document_covariate_matrix}{matrix of documents (rows) by covariates
(columns)}

\item{formula}{vector of formulas for the continuous change}

\item{nchangepoints}{vector of the number of change points to include in 
the model}

\item{...}{additional arguments to be passed to subfunctions}
}
\value{
a list of [1] the LDA model(s), [2] the selected LDA model(s), and
  [3] the time series model(s) on the selected LDA model(s)
}
\description{
Run LDA, select LDA, run MTS; return all outputs
}
\examples{
  \dontrun{
    data(rodents)
    lda_data <- dplyr::select(rodents, -c(newmoon, date, plots, traps))
    ts_data <- data.frame("time" = data.frame(rodents)[ , "newmoon"])
    r_LDATS <- LDATS::LDA_TS(lda_data, ts_data, formula = "time", 
                             ntopics = 2:5, nseeds = 2, ncores = 4, 
                             nit = 100) 
  }

}
