% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_TS.R, R/multinom_TS.R
\name{LDA_TS}
\alias{LDA_TS}
\alias{check_LDA_TS_inputs}
\alias{check_multinom_TS_inputs}
\title{Run a full set of Latent Dirichlet Allocations and Time 
  Series models}
\usage{
LDA_TS(document_term_table, document_covariate_table, topics = 2,
  nseeds = 1, formulas = ~1, nchangepoints = 0,
  weights = document_weights(document_term_table),
  control = LDA_TS_controls_list())

check_LDA_TS_inputs(document_term_table, document_covariate_table,
  topics = 2, nseeds = 1, formulas = ~1, nchangepoints = 0,
  weights = document_weights(document_term_table),
  control = LDA_TS_controls_list())

check_multinom_TS_inputs(data, formula, changepoints = NULL,
  weights = NULL, control = TS_controls_list())
}
\arguments{
\item{document_term_table}{Table of observation count data (rows: 
documents (\eqn{M}), columns: terms (\eqn{V})). May be a 
\code{matrix} or \code{data.frame} but must be conformable to a matrix
of integers (as verified by \code{\link{check_document_term_table}}). 
This table is a document-level summary of the data noted as
\strong{\eqn{w}} (the word-level topic identity) in the math description.}

\item{document_covariate_table}{Document covariate table (rows:
documents (\eqn{M}), columns: time index and covariate options). 
Every model needs a covariate to describe the time value for each
document (in whatever relevant units), whose name in the table is input
in \code{control} (the \code{timename} entry), that dictates the 
application of the change points. 
In addition, all covariates named within specific models in
\code{formula} must be included. Must be a conformable to a data table,
as verified by \code{\link{check_document_covariate_table}}.}

\item{topics}{Vector of the number of topics to evaluate in the LDAs. Must
be integer values.}

\item{nseeds}{\code{integer} number of seeds (replicate starts) to use for 
each value of \code{topics} in the LDAs.}

\item{formulas}{Vector of \code{formula}(s) for the continuous 
(non-changepoint) component of the time series models. Any 
predictor variable included in a formula must also be a column in the
\code{document_covariate_table}. Each element (formula) in the vector
is evaluated for each number of change points and each LDA model.}

\item{nchangepoints}{Vector of \code{integer}s corresponding to the number 
of change points to include in the time series models. 0 is a valid input 
corresponding to no change points (\emph{i.e.}, a singular time series
model), and the current implementation can reasonably include up to 6 
change points. Each element in the vector is the number of change points 
(\ifelse{html}{\out{<i>P<sub>m<sub>2</sub></sub></i>}}{\eqn{P_m_2}} for
model \ifelse{html}{\out{<i>m<sub>2</sub></i>}}{\eqn{m_2}} in the math
description) used to segement 
the data for each formula (entry in \code{formulas}) component of the 
TS model, for each selected LDA model.}

\item{weights}{Optional class \code{numeric} vector of weights for each 
document. Corresponds to the vector \strong{\eqn{v}} in the math 
description. Defaults to value calculated by 
\code{\link{document_weights}}, but can be set as \code{NULL}.}

\item{control}{Class \code{LDA_TS_controls} list (as generated by
\code{\link{LDA_TS_controls_list}}) that contains 
\code{LDA_controls} (\code{\link{LDA_controls_list}}), 
\code{TS_controls} (\code{\link{TS_controls_list}}), and the single 
top-level control \code{quiet}.}
}
\value{
\code{LDATS} returns a class \code{LDA_TS} list object including all 
  fitted LDA and TS models and selected models specifically as elements 
  \code{"LDA models"} (from \code{\link{LDA_set}}),
  \code{"Selected LDA model"} (from \code{\link{select_LDA}}), 
  \code{"TS models"} (from \code{\link{TS_on_LDA}}), and
  \code{"Selected TS model"} (from \code{\link{select_TS}}).
}
\description{
Conduct a complete LDATS analysis (Christensen 
  \emph{et al.} 2018), including running a suite of Latent Dirichlet
  Allocation (LDA) models (Blei \emph{et al.} 2003, Grun and Hornik 2011) 
  via \code{\link{LDA_set}}, selecting LDA model(s) via 
  \code{\link{select_LDA}}, running a complete set of Bayesian Time Series
  (TS) models (Western and Kleykamp 2004) via \code{\link{TS_on_LDA}} on
  the chosen LDA model(s), and selecting the best TS model via 
  \code{\link{select_TS}}.

\code{check_LDA_TS_inputs} verifies that the inputs to 
  \code{LDA_TS} are of proper classes for a full analysis.

\code{check_multinom_TS_inputs} verifies that the inputs to 
  \code{multinom_TS} are of proper classes for an analysis.
}
\examples{
\dontrun{
  data(rodents)
  document_term_table <- rodents$document_term_table
  document_covariate_table <- rodents$document_covariate_table
  
  mod <- LDA_TS(document_term_table, document_covariate_table,
                topics = 2, nseeds = 1, formulas = ~ 1, nchangepoints = 1,
                weights = document_weights(document_term_table), 
                control = LDA_TS_controls_list())
}

}
\references{
Blei, D. M., A. Y. Ng, and M. I. Jordan. 2003. Latent Dirichlet
  Allocation. \emph{Journal of Machine Learning Research} 
  \strong{3}:993-1022.
  \href{http://jmlr.csail.mit.edu/papers/v3/blei03a.html}{link}.

  Christensen, E., D. J. Harris, and S. K. M. Ernest. 2018.
  Long-term community change through multiple rapid transitions in a 
  desert rodent community. \emph{Ecology} \strong{99}:1523-1529. 
  \href{https://doi.org/10.1002/ecy.2373}{link}.

  Grun B. and K. Hornik. 2011. topicmodels: An R Package for Fitting Topic
  Models. \emph{Journal of Statistical Software} \strong{40}:13.
  \href{https://www.jstatsoft.org/article/view/v040i13}{link}.

  Western, B. and M. Kleykamp. 2004. A Bayesian change point model for 
  historical time series analysis. \emph{Political Analysis}
  \strong{12}:354-374.
  \href{https://doi.org/10.1093/pan/mph023}{link}.
}
