% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_TS.R
\name{LDA_TS_controls_list}
\alias{LDA_TS_controls_list}
\title{Create the controls list for the LDATS model}
\usage{
LDA_TS_controls_list(TS_control = TS_controls_list(),
  LDA_control = LDA_controls_list(), quiet = FALSE)
}
\arguments{
\item{TS_control}{Class \code{TS_controls} list, a named list generally 
created by \code{\link{TS_controls_list}} that holds control parameters
for the Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls.}

\item{LDA_control}{Named list of control parameters to be used in 
\code{\link{LDA_set}} and \code{\link[topicmodels]{LDA}},
as generally created by \code{\link{LDA_controls_list}}. Note that 
\code{seed} will be overwritten.}

\item{quiet}{\code{logical} indicator of whether the model should run 
quietly at the top level.}
}
\value{
control Class \code{LDA_TS_controls} list that contains 
  \code{LDA_controls}, \code{TS_controls}, and the top-level \code{quiet}.
}
\description{
Create and define a list of control options used to run the
  LDATS model, as implemented by \code{\link{LDA_TS}}.
}
