% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDATS_package.R
\docType{package}
\name{LDATS}
\alias{LDATS}
\alias{LDATS-package}
\title{Two-stage LDA-TimeSeries analyses}
\description{
Performs two-stage analysis of multivariate temporal data
  using a combination of Latent Dirichlet Allocation and multinomial 
  time series models.
}
\keyword{package}
