% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDATS.R
\name{LDATS}
\alias{LDATS}
\title{Two-stage LDA-time series analysis}
\usage{
LDATS(data = NULL, LDA_eval = quote(AIC), LDA_selector = quote(min), ...)
}
\arguments{
\item{data}{data set (currently just the data for the LDA)}

\item{LDA_eval}{function name for evaluation of the LDA models}

\item{LDA_selector}{function name for selecting the LDA models}

\item{...}{additional arguments to be passed to subfunctions}
}
\value{
(currently) the output of the selected LDA model
}
\description{
Two-stage LDA-time series analysis
}
