% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDATS_package.R
\docType{package}
\name{LDATS}
\alias{LDATS}
\alias{LDATS-package}
\title{Performs two-stage LDA-timeseries analyses}
\description{
This package is designed to analyze multivariate time series
  using a combination of Latent Dirichlet Allocation and multinomial 
  regression.
}
\keyword{package}
