% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldats_classic_method.R
\name{ldats_classic}
\alias{ldats_classic}
\title{Estimate changepoints using the LDATS classic ptMCMC method}
\usage{
ldats_classic(TS, control = list())
}
\arguments{
\item{TS}{\code{list} time series model object.}

\item{control}{A \code{list} of parameters to control the fitting of the
Values not input assume defaults set by \code{\link{TS_control}}.}
}
\value{
\code{list} of changepoint locations, log likelihoods, and model
 diagnostics.
}
\description{
Uses the LDATS classic ptMCMC method to fit a changepoint
 model.
}
