% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_methods.R
\name{ldats_classic}
\alias{ldats_classic}
\title{Estimate changepoints using the LDATS classic ptMCMC method}
\usage{
ldats_classic(TS, control = list())
}
\arguments{
\item{TS}{\code{list} time series model object.}

\item{control}{A \code{list} of parameters to control the fitting of the
time series model via the LDATS classic ptMCMC method. Values not input 
assume defaults set by \code{\link{ldats_classic_control}}.}
}
\value{
\code{list} of changepoint locations, log likelihoods, and model
 diagnostics.
}
\description{
Uses the LDATS classic parallel tempering Markov Chain Monte 
 Carlo (ptMCMC) methods (Earl and Deem 2005) to fit a changepoint
 model, following Christensen \emph{et al.} (2018).
}
\references{
Christensen, E., D. J. Harris, and S. K. M. Ernest. 2018.
  Long-term community change through multiple rapid transitions in a 
  desert rodent community. \emph{Ecology} \strong{99}:1523-1529. 
  \href{https://doi.org/10.1002/ecy.2373}{link}.

  Earl, D. J. and M. W. Deem. 2005. Parallel tempering: theory, 
  applications, and new perspectives. \emph{Physical Chemistry Chemical 
  Physics} \strong{7}: 3910-3916.
  \href{https://doi.org/10.1039/B509983H}{link}.
}
