% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_TS.R
\name{logLik.multinom_TS_fit}
\alias{logLik.multinom_TS_fit}
\title{Log likelihood of a multinomial TS model}
\usage{
logLik.multinom_TS_fit(x, ...)
}
\arguments{
\item{x}{A \code{multinom_TS_fit}-class object.}
}
\value{
Log likelihood of the model.
}
\description{
Convenience function to simply extract the \code{logLik}
  element from a \code{multinom_TS_fit}-class object.
}
