% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mirror_vcov}
\alias{mirror_vcov}
\title{Create a properly symmetrix variance covariance matrix}
\usage{
mirror_vcov(x)
}
\arguments{
\item{x}{Model object that has a defined method for \code{vcov}.}
}
\value{
Properly symmetric variance covariance matrix.
}
\description{
If the default matrix returned by \code{vcov} is not, in fact,
  symmetric (as occurs occasionally in the \code{multinom} function from
  \code{nnet}), make the matrix symmetric by mirroring the lower triangle.
}
