% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{MTS}
\alias{MTS}
\title{Multinomial Time Series analysis of a topic model classification}
\usage{
MTS(formula, data, nchangepoints, weights, nit = 10000, ts_memo, ...)
}
\arguments{
\item{formula}{formula for the continuous change}

\item{data}{data frame including the predictor and response variables}

\item{nchangepoints}{number of change points to include in the model}

\item{weights}{weights}

\item{nit}{number of iterations used}

\item{...}{additional arguments to be passed to subfunctions}
}
\value{

}
