% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_functions.R
\name{MTS_set}
\alias{MTS_set}
\title{Conduct a set of MTS analyses}
\usage{
MTS_set(data = NULL, formula, nchangepoints, weights = rep(1,
  nrow(data[[1]])), ...)
}
\arguments{
\item{formula}{vector of formulas for the continuous change}

\item{nchangepoints}{vector of the number of change points to include in 
the model}

\item{...}{additional arguments to be passed to subfunctions}

\item{prepped_data}{data prepped for each of the MTS analyses}

\item{vector}{of weights for each document}
}
\value{

}
