% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_TS.R
\name{multinom_chunk}
\alias{multinom_chunk}
\title{Fit a multinomial time series model chunk}
\usage{
multinom_chunk(data, formulaRHS, start_time, end_time, weights = NULL)
}
\arguments{
\item{data}{Class \code{data.frame} object including the predictor and 
response variables.}

\item{start_time}{Start time for the chunk.}

\item{end_time}{End time for the chunk.}

\item{weights}{Optional class \code{numeric} vector of weights for each 
document. Corresponds to the vector \strong{\eqn{v}} in the math 
description.}

\item{formula_RHS}{Right Hand Side of the continuous time formula as a 
character vector.}
}
\value{
Fitted model for the chunk.
}
\description{
Fit a multinomial regression model to a defined chunk of time
  \code{(start_time, end_time]} within a time series. The fit is conducted
  via \code{\link[nnet]{multinom}}
}
