% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{multinom_ts}
\alias{multinom_ts}
\title{Fit multinomial time series model}
\usage{
multinom_ts(formula_RHS, data, changepoints = NULL, weights, ...)
}
\arguments{
\item{formula_RHS}{Right Hand Side of the continuous time formula as a 
character vector}

\item{data}{data frame including the predictor and response variables}

\item{changepoints}{selections of the change points}

\item{weights}{weights}

\item{...}{other arguments to be passed to subfunctions}
}
\value{
list of chunk-level model fits and the total log likelihood 
  combined across all the chunks
}
\description{
Fit a multinomial regression model with covariate impacts 
  assuming no temporal autocorrelation
}
\examples{
NA
}
