% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_TS.R
\name{multinom_TS}
\alias{multinom_TS}
\title{Fit a multinomial change point Time Series model}
\usage{
multinom_TS(data, formula_RHS, changepoints = NULL, weights = NULL,
  control = TS_controls_list())
}
\arguments{
\item{data}{Class \code{data.frame} object including [1] the time variable
(indicated in \code{control}), [2] the predictor variables (required by
\code{formula}) and [3], the multinomial response variable (indicated
in \code{control}).}

\item{formula_RHS}{Right Hand Side of the continuous time formula as a 
character vector.}

\item{changepoints}{Numeric vector indicating locations of the change 
points.}

\item{weights}{Optional class \code{numeric} vector of weights for each 
document. Corresponds to the vector \strong{\eqn{v}} in the math 
description.}

\item{control}{Class \code{TS_controls} list, holding control parameters
for the Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls, generated by 
\code{\link{TS_controls_list}}.}
}
\value{
Object of class \code{multinom_TS_fit}, which is a list of [1]
  chunk-level model fits and [2] the total log likelihood combined across 
  all chunks.
}
\description{
Fit a set of multinomial regression models to a time series of
  of data divided into multiple chunks based on change points.
}
