% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_TS.R
\name{multinom_TS_chunk}
\alias{multinom_TS_chunk}
\title{Fit a multinomial Time Series model chunk}
\usage{
multinom_TS_chunk(data, formula, chunk, weights = NULL,
  control = TS_controls_list())
}
\arguments{
\item{data}{Class \code{data.frame} object including the predictor and 
response variables.}

\item{formula}{Formula as a class \code{formula} or class \code{character} 
object describing the chunk.}

\item{chunk}{Length-2 vector of times: [1] \code{start}, the start time 
for the chunk and [2] \code{end}, the end time for the chunk.}

\item{weights}{Optional class \code{numeric} vector of weights for each 
document. Corresponds to the vector \strong{\eqn{v}} in the math 
description.}

\item{control}{Class \code{TS_controls} list, holding control parameters
for the Time Series model, generated by \code{\link{TS_controls_list}}.}
}
\value{
Fitted model object for the chunk, of classes \code{multinom} and
  \code{nnet}.
}
\description{
Fit a multinomial regression model to a defined chunk of time
  \code{[chunk$start, chunk$_time]} within a time series. The fit is 
  conducted via \code{\link[nnet]{multinom}}.
}
\references{
Ripley, B. D. 1996. Pattern Recognition and Neural Networks. Cambridge.

  Venables, W. N. and B. D. Ripley. 2002. Modern Applied Statistics with S.
  Fourth edition. Springer.
}
