% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_TS.R
\name{package_LDA_TS}
\alias{package_LDA_TS}
\title{Package the output of LDA_TS}
\usage{
package_LDA_TS(LDAs, sel_LDA, TSs, sel_TSs)
}
\arguments{
\item{LDAs}{List (class: \code{LDA_set}) of LDA models (class: 
"\code{LDA}").}

\item{sel_LDA}{A reduced version of \code{LDAs} that only includes the 
selected LDA model(s). The returned object is still an object of
class \code{LDA_set}.}

\item{TSs}{Class \code{TS_on_LDA} list of results from \code{\link{TS}} 
applied for each model on each LDA model input.}

\item{sel_TSs}{A reduced version of \code{TSs} that only includes the 
selected TS model. The returned object is still an object of
class \code{TS_fit}.}
}
\value{
Class \code{LDA_TS} object including all fitted models and selected 
  models specifically.
}
\description{
Combine the objects, name them as elements of the list, and
  set the class of the list, for the return from \code{\link{LDA_TS}}.
}
