% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_functions.R
\name{plot.LDA}
\alias{plot.LDA}
\title{Plot an LDA model}
\usage{
\method{plot}{LDA}(x, ...)
}
\arguments{
\item{x}{an LDA topic model output}

\item{...}{additional arguments to be passed to subfunctions}
}
\value{
model plots
}
\description{
Function used to create an LDA summary plot, called using the 
  plot method
}
