% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_plots.R
\name{plot.LDA}
\alias{plot.LDA}
\title{Plot the results of an LDATS LDA model}
\usage{
\method{plot}{LDA}(x, ..., xtime = NULL, xname = NULL, cols = NULL,
  option = "D")
}
\arguments{
\item{x}{Object of class \code{LDA}}

\item{...}{Not used, retained for alignment with base function.}

\item{xtime}{Optional x values used to plot the topic proportions according
to a specific time value (rather than simply the order of observations).}

\item{xname}{Optional name for the x values used in plotting the topic
proportions (otherwise defaults to "Document").}

\item{cols}{colors to be used to plot the topics}

\item{option}{A character string indicating the colormap option to use if 
`cols == NULL`. Four options are available: "magma" (or "A"), "inferno" 
(or "B"), "plasma" (or "C"), "viridis" (or "D", the default option) and 
"cividis" (or "E").}
}
\value{
LDATS LDA model plots.
}
\description{
Create an LDATS LDA summary plot, with a top panel showing
  the topic proportions for each word and a bottom panel showing the topic
  proportions of each document/over time.
}
\examples{
\dontrun{
  data(rodents)
  lda_data <- dplyr::select(rodents, -c(newmoon, date, plots, traps))
  lda_models <- parLDA(lda_data, topics = 4, nseeds = 10)
  best_lda <- LDA_select(lda_models)
  plot(best_lda, option = "cividis")
}

}
