% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_plots.R
\name{plot.LDA}
\alias{plot.LDA}
\alias{LDA_plot_top_panel}
\alias{LDA_plot_bottom_panel}
\alias{set_LDA_plot_colors}
\title{Plot the results of an LDATS LDA model}
\usage{
\method{plot}{LDA}(x, ..., xtime = NULL, xname = NULL, cols = NULL,
  option = "C", alpha = 0.8, LDATS = FALSE)

LDA_plot_top_panel(x, cols = NULL, option = "C", alpha = 0.8,
  together = FALSE, LDATS = FALSE)

LDA_plot_bottom_panel(x, xtime = NULL, xname = NULL, cols = NULL,
  option = "C", alpha = 0.8, together = FALSE, LDATS = FALSE)

set_LDA_plot_colors(x, cols = NULL, option = "C", alpha = 0.8)
}
\arguments{
\item{x}{Object of class \code{LDA}.}

\item{...}{Not used, retained for alignment with base function.}

\item{xtime}{Optional x values used to plot the topic proportions according
to a specific time value (rather than simply the order of observations).}

\item{xname}{Optional name for the x values used in plotting the topic
proportions (otherwise defaults to "Document").}

\item{cols}{Colors to be used to plot the topics.
Any valid color values (\emph{e.g.}, see \code{\link[grDevices]{colors}},
\code{\link[grDevices]{rgb}}) can be input as with a standard plot. 
The default (\code{cols = NULL}) triggers use of 
\code{\link[viridis]{viridis}} color options (see \code{option}).}

\item{option}{A \code{character} string indicating the color option
from \code{\link[viridis]{viridis}} to use if `cols == NULL`. Four 
options are available: "magma" (or "A"), "inferno" (or "B"), "plasma" 
(or "C", the default option), "viridis" (or "D") and "cividis" (or "E").}

\item{alpha}{Numeric value [0,1] that indicates the transparency of the 
colors used. Supported only on some devices, see 
\code{\link[grDevices]{rgb}}.}

\item{LDATS}{\code{logical} indicating if the LDA plot is part of a larger 
LDATS plot output.}

\item{together}{\code{logical} indicating if the subplots are part of a 
larger LDA plot output.}
}
\value{
\code{plot.LDA},\code{LDA_plot_top_panel},\code{LDA_plot_bottom_panel}: 
    \code{NULL}. \cr \cr
  \code{set_LDA_plot_colors}: \code{vector} of \code{character} hex codes
    indicating colors to use.
}
\description{
\code{plot.LDA} creates an LDATS LDA summary plot, with a top panel 
    showing the topic proportions for each word and a bottom panel showing
    the topic proportions of each document/over time. \cr \cr
  \code{LDA_plot_top_panel} creates an LDATS LDA summary plot 
    top panel showing the topic proportions word-by-word. \cr \cr
  \code{LDA_plot_bottom_panel} creates an LDATS LDA summary plot
    bottom panel showing the topic proportions over time/documents. \cr \cr
  \code{set_LDA_plot_colors} creates the set of colors to be used in
    the LDA plots based on the variety of argument options.
}
