% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_functions.R
\name{plot.LDA}
\alias{plot.LDA}
\title{Function used to create an LDA summary plot, called using the plot 
  method}
\usage{
\method{plot}{LDA}(model, cols = NULL, ...)
}
\arguments{
\item{model}{an LDA topic model output}

\item{cols}{a vector of color codes, one for each topic}
}
\value{
model plots
}
