% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_functions.R
\name{plot.LDA_list}
\alias{plot.LDA_list}
\title{Generalization of the plot function to work on a list of LDA topic models}
\usage{
plot.LDA_list(model_list)
}
\arguments{
\item{model_list}{a list of LDA topic model outputs}
}
\value{
model-by-model plots
}
\description{
Generalization of the plot function to work on a list of LDA topic models
}
\examples{
NA
}
