% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_functions.R
\name{plot.LDA_list}
\alias{plot.LDA_list}
\title{Generalization of the plot function to work on a list of LDA topic 
  models}
\usage{
\method{plot}{LDA_list}(model_list, cols = NULL)
}
\arguments{
\item{model_list}{a list of LDA topic model outputs}

\item{cols}{a vector of color codes to use, one for each of the max topics}
}
\value{
model-by-model plots
}
\examples{
  data(rodents)
  lda_data <- dplyr::select(rodents, -c(newmoon, date, plots, traps))
  r_LDA <- LDATS::LDA(data = lda_data, ntopics = 2, nseeds = 2)
  plot(r_LDA)
}
