% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_functions.R
\name{plot.LDA_list}
\alias{plot.LDA_list}
\title{Plot a set of LDA models}
\usage{
\method{plot}{LDA_list}(x, ...)
}
\arguments{
\item{x}{a list of LDA topic model outputs}

\item{...}{additional arguments to be passed to subfunctions}
}
\value{
model-by-model plots
}
\description{
Generalization of the plot function to work on a list of LDA 
  topic models
}
\examples{
\dontrun{
  data(rodents)
  lda_data <- select(rodents, -c(newmoon, date, plots, traps))
  r_LDA <- parLDA(data = lda_data, ntopics = 2, nseeds = 2)
  plot(r_LDA)
}
}
