% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_TS_plots.R
\name{plot.LDA_TS}
\alias{plot.LDA_TS}
\title{Plot a LDATS full model}
\usage{
\method{plot}{LDA_TS}(x, ..., control = LDA_TS_controls_list(),
  cols = set_LDA_TS_plot_cols(), bin_width = 1, xlab = NULL,
  selection = "median", interactive = TRUE)
}
\arguments{
\item{x}{A \code{LDA_TS} object of a full LDATS model fit by
\code{LDA_TS}.}

\item{...}{Additional arguments to be passed to subfunctions. Not currently
used, just retained for alignment with \code{plot}.}

\item{control}{Class \code{LDA_TS_controls} list that contains 
\code{LDA_controls}, \code{TS_controls}, and the top-level \code{quiet}.}

\item{cols}{\code{list} of elements used to define the colors for the two
panels of the summary plot, as generated simply using 
\code{LDA_TS_summary_cols}. \code{cols} has two elements \code{LDA} and
\code{TS}, each corresponding to its stage in the full model. 
\code{LDA} contains entries \code{cols} and \code{options} (see
\code{set_LDA_plot_colors} for details on usage. \code{TS} contains two
entries, \code{rho} and \code{gamma}, each corresponding to the related 
panel, and each containing default values for entries named \code{cols}, 
\code{option}, and \code{alpha}. See \code{set_gamma_colors} and 
\code{set_rho_hist_colors} for details on usage.}

\item{bin_width}{Width of the bins used in the histograms of the summary 
timeseries plot, in units of the x-axis (the time variable used to fit 
the model).}

\item{xlab}{Label for the x-axis in the summary time series plot.}

\item{selection}{Indicator of the changepoints to use in the timeseries
summary plot. Currently only defined for "median" and "mode".}

\item{interactive}{\code{logical} indicator, should be set to \code{TRUE}
currently (except for testing).}
}
\value{
Nothing. Plots are generated in the active graphics device.
}
\description{
Generalization of the \code{plot} function to work on fitted
  LDA_TS model (class \code{LDA_TS}).
}
\examples{
\dontrun{
  data(rodents)
  lda_data <- rodents$document_term_table
  document_term_table <- rodents$document_term_table
  document_covariate_table <- rodents$document_covariate_table
  
  mod <- LDA_TS(document_term_table, document_covariate_table,
                topics = 2, nseeds = 1, formulas = ~ 1, nchangepoints = 1,
                weights = document_weights(document_term_table), 
                control = LDA_TS_controls_list())
  plot(mod)
}

}
