% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_TS_plots.R
\name{plot.LDA_TS}
\alias{plot.LDA_TS}
\title{Plot the key results from a full LDATS analysis}
\usage{
\method{plot}{LDA_TS}(x, ..., control = LDA_TS_controls_list(),
  cols = set_LDA_TS_plot_cols(), bin_width = 1, xlab = NULL,
  selection = "median")
}
\arguments{
\item{x}{A \code{LDA_TS} object of a full LDATS model fit by
\code{\link{LDA_TS}}.}

\item{...}{Additional arguments to be passed to subfunctions. Not currently
used, just retained for alignment with \code{plot}.}

\item{control}{Class \code{LDA_TS_controls} list (as generated by
\code{\link{LDA_TS_controls_list}} that contains 
\code{LDA_controls} (\code{\link{LDA_controls_list}}), 
\code{TS_controls} (\code{\link{TS_controls_list}}), and the single 
top-level control \code{quiet}.}

\item{cols}{\code{list} of elements used to define the colors for the two
panels of the summary plot, as generated simply using 
\code{\link{set_LDA_TS_plot_cols}}. \code{cols} has two elements: 
\code{LDA} and \code{TS}, each corresponding the set of plots for
its stage in the full model. \code{LDA} contains entries \code{cols}
and \code{options} (see \code{\link{set_LDA_plot_colors}}). \code{TS} 
contains two entries, \code{rho} and \code{gamma}, each corresponding
to the related panel, and each containing default values for entries
named \code{cols}, \code{option}, and \code{alpha} (see 
\code{\link{set_TS_summary_plot_cols}}, \code{\link{set_gamma_colors}},  
and \code{\link{set_rho_hist_colors}}).}

\item{bin_width}{Width of the bins used in the histograms of the summary 
timeseries plot, in units of the time variable used to fit the model
(the x-axis).}

\item{xlab}{Label for the x-axis in the summary time series plot. Defaults
to \code{NULL}, which results in usage of the \code{timename} element 
of the control list (held in\code{control$TS_control$timename}). To have
no label printed, set \code{xlab = ""}.}

\item{selection}{Indicator of the change points to use in the timeseries
summary plot. Currently only defined for \code{"median"} and 
\code{"mode"}.}
}
\description{
Generalization of the \code{\link[graphics]{plot}} function to
  work on fitted LDA_TS model objects (class \code{LDA_TS}) returned by 
  \code{\link{LDA_TS}}).
}
\examples{
\dontrun{
  data(rodents)
  document_term_table <- rodents$document_term_table
  document_covariate_table <- rodents$document_covariate_table
  
  mod <- LDA_TS(document_term_table, document_covariate_table,
                topics = 2, nseeds = 1, formulas = ~ 1, nchangepoints = 1,
                weights = document_weights(document_term_table), 
                control = LDA_TS_controls_list())
  plot(mod, binwidth = 5, xlab = "New moon")
}

}
