% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_plots.R
\name{plot.LDA_VEM}
\alias{plot.LDA_VEM}
\alias{LDA_plot_top_panel}
\alias{LDA_plot_bottom_panel}
\title{Plot the results of an LDATS LDA model}
\usage{
\method{plot}{LDA_VEM}(x, ..., xtime = NULL, xname = NULL,
  cols = NULL, option = "E", alpha = 0.8)

LDA_plot_top_panel(x, cols, option, alpha)

LDA_plot_bottom_panel(x, xtime, xname, cols, option, alpha)
}
\arguments{
\item{x}{Object of class \code{LDA_VEM}.}

\item{...}{Not used, retained for alignment with base function.}

\item{xtime}{Optional x values used to plot the topic proportions according
to a specific time value (rather than simply the order of observations).}

\item{xname}{Optional name for the x values used in plotting the topic
proportions (otherwise defaults to "Document").}

\item{cols}{Colors to be used to plot the topics, or \code{NULL} to use
\code{\link[viridis]{viridis}} options.}

\item{option}{A \code{character} string indicating the color option
from \code{\link[viridis]{viridis}} to use if `cols == NULL`. Four 
options are available: "magma" (or "A"), "inferno" (or "B"), "plasma" 
(or "C"), "viridis" (or "D", the default option) and "cividis" (or "E").}

\item{alpha}{Numeric value [0,1] that indicates the transparency of the 
colors used. Supported only on some devices, see 
\code{\link[grDevices]{rgb}}.}
}
\description{
Create an LDATS LDA summary plot, with a top panel showing
  the topic proportions for each word and a bottom panel showing the topic
  proportions of each document/over time. The plot function is defined for
  class \code{LDA_VEM} specifically (see \code{\link[topicmodels]{LDA}}).

\code{LDA_plot_top_panel}: create an LDATS LDA summary plot 
  top panel showing the topic proportions word-by-word.

\code{LDA_plot_bottom_panel}: create an LDATS LDA summary plot
  bottom panel showing the topic proportions over time/documents.
}
\examples{
\dontrun{
  data(rodents)
  lda_data <- rodents$document_term_table
  r_LDA <- LDA_set(lda_data, topics = 4, nseeds = 10) 
  best_lda <- select_LDA(r_LDA)
  plot(best_lda, option = "cividis")
}

}
