% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_plots.R
\name{pred_gamma_TS_plot}
\alias{pred_gamma_TS_plot}
\title{Create the plot of fitted topic proportions over time}
\usage{
pred_gamma_TS_plot(x, selection = "median", cols, xlab)
}
\arguments{
\item{x}{Object of class \code{TS_fit}.}

\item{selection}{Indicator of the changepoints to use. Currently only
defined for \code{"median"}.}

\item{cols}{Hex values of the colors to be used to plot the time series.}

\item{xlab}{Label for the x-axis.}
}
\value{
Nothing. The plot is generated in the active graphics device.
}
\description{
Produces a time series of the fitted topic proportions 
  (gamma values) over time, based on a selected set of changepoint 
  locations.
}
