% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{prep_cpts}
\alias{prep_cpts}
\alias{update_cpts}
\title{Initialize and update the changepoint matrix used in the ptMCMC
  algorithm}
\usage{
prep_cpts(data, formula, nchangepoints, weights,
  control = TS_controls_list())

update_cpts(cpts, swaps)
}
\arguments{
\item{data}{Class \code{data.frame} object including [1] the time variable
(indicated in \code{control}), [2] the predictor variables (required by
\code{formula}) and [3], the multinomial response variable (indicated
in \code{formula}).}

\item{formula}{\code{formula} describing the continuous change. Any 
predictor variable included must also be a column in the
\code{data}.  Any (multinomial) response variable must also be a set of
columns in \code{data}.}

\item{nchangepoints}{Integer corresponding to the number of 
change points to include in the model. 0 is a valid input (corresponding
to no change points, so a singular time series model), and the current 
implementation can reasonably include up to 6 change points. The 
number of change points is used to dictate the segementation of the data  
for each continuous model and each LDA model.}

\item{weights}{Optional class \code{numeric} vector of weights for each 
document. Corresponds to the vector \strong{\eqn{v}} in the math 
description.}

\item{control}{Class \code{TS_controls} list, holding control parameters
for the Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls, generated by 
\code{\link{TS_controls_list}}.}

\item{cpts}{The existing matrix of changepoints.}

\item{swaps}{Chain configuration after among-temperature swaps.}
}
\value{
List of [1] matrix of change points (rows) for each temperature 
  (columns) and [2] vector of log-likelihood values for each of the chains.
}
\description{
Each of the chains is initialized by \code{prep_cpts} using a 
  draw from the available times (i.e. assuming a uniform prior), the best 
  fit (by likelihood) draw is put in the focal chain with each subsequently 
  worse fit placed into the subsequently hotter chain. \code{update_cpts}
  updates the changepoints after every iteration in the ptMCMC algorithm.
}
