% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldats_classic_method.R
\name{prep_ids}
\alias{prep_ids}
\alias{update_ids}
\title{Initialize and update the chain ids throughout the LDATS classic
  ptMCMC algorithm}
\usage{
prep_ids(TS, control = list())

update_ids(ids, swaps)
}
\arguments{
\item{TS}{\code{list} time series model object.}

\item{control}{A \code{list} of parameters to control the fitting of the
Values not input assume defaults set by \code{\link{TS_control}}.}

\item{ids}{The existing vector of chain ids.}

\item{swaps}{Chain configuration after among-temperature swaps.}
}
\value{
The vector of chain ids.
}
\description{
\code{prep_ids} creates and \code{update_ids} updates 
  the active vector of identities (ids) for each of the chains in the 
  ptMCMC algorithm. These ids are used to track trips of the particles
  among chains.
  \cr \cr
  These functions were designed to work within \code{\link{TS}} and 
  specifically \code{\link{est_changepoints}}, but have been generalized
  and would work within any general ptMCMC as long as \code{control},
  \code{ids}, and \code{swaps} are formatted properly.
}
