% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{prep_LDA_control}
\alias{prep_LDA_control}
\title{Set the control inputs to include the seed}
\usage{
prep_LDA_control(seed, control = list())
}
\arguments{
\item{seed}{\code{integer} used to set the seed of the specific model.}

\item{control}{Named list of control parameters to be used in 
\code{\link[topicmodels]{LDA}} Note that if \code{control} has an 
element named \code{seed} it will be overwritten by the \code{seed} 
argument of \code{prep_LDA_control}.}
}
\value{
List (class: \code{list}) of controls to be used in the LDA.
}
\description{
Update the control list for the LDA model with the specific
  seed as indicated. And remove controls not used within the LDA itself.
}
\examples{
  prep_LDA_control(seed = 1) 

}
