% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{prep_LDA_control}
\alias{prep_LDA_control}
\title{Set the control inputs to include the seed}
\usage{
prep_LDA_control(seed, control = NULL)
}
\arguments{
\item{control}{Named list of control parameters to be used in 
\code{\link[topicmodels]{LDA}} (note that "seed" will be overwritten).}

\item{nseeds}{Integer number of seeds (replicate starts) to use for the 
specific model.}
}
\value{
List (class: "\code{LDAcontrol}") of LDA controls to be used.
}
\description{
Update the control liste for the LDA model with the specific
  seed as indicated.
}
