% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{prep_pbar}
\alias{prep_pbar}
\alias{update_pbar}
\title{Initialize and tick through the progress bar}
\usage{
prep_pbar(control = list(), type = "rho", nr = NULL)

update_pbar(pbar, control = list())
}
\arguments{
\item{control}{A \code{list} of parameters to control the fitting of the
iterative model.}

\item{type}{\code{character} value of possible types of progress bars.
Currently available options are "rho" (for change point locations) and
"eta" (for time series regressors).}

\item{nr}{\code{integer} number of unique realizations, needed when
\code{type} = "eta".}

\item{pbar}{The progress bar object returned from \code{prep_pbar}.}
}
\value{
\code{prep_pbar}: the initialized progress bar object. \cr \cr
  \code{update_pbar}: the ticked-forward \code{pbar}.
}
\description{
\code{prep_pbar} creates and \code{update_pbar} steps
  through the progress bars (if desired) in, e.g., \code{\link{TS}}
}
