% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldats_classic_method.R
\name{prep_ptMCMC_inputs}
\alias{prep_ptMCMC_inputs}
\title{Prepare the inputs for the ptMCMC algorithm estimation of 
  change points}
\usage{
prep_ptMCMC_inputs(TS, control = list())
}
\arguments{
\item{TS}{\code{list} time series model object.}

\item{control}{A \code{list} of parameters to control the fitting of the
time series model via the LDATS classic ptMCMC method. Values not input 
assume defaults set by \code{\link{ldats_classic_control}}.}
}
\value{
\code{list} containing the static 
  inputs for use within the ptMCMC algorithm for estimating change points.
}
\description{
Package the static inputs (controls and data structures) used
  by the ptMCMC algorithm in the context of estimating change points. 
  \cr \cr
  This function was designed to work within \code{\link{TS}} and 
  specifically \code{\link{est_changepoints}}. It is still hardcoded to do
  so, but has the capacity to be generalized to work with any estimation
  via ptMCMC with additional coding work.
}
