% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{prep_saves}
\alias{prep_saves}
\alias{update_saves}
\alias{process_saves}
\title{Prepare and update the data structures to save the ptMCMC output}
\usage{
prep_saves(nchangepoints, control = TS_controls_list())

update_saves(i, saves, steps, swaps)

process_saves(saves, control)
}
\arguments{
\item{nchangepoints}{Integer corresponding to the number of 
change points to include in the model. 0 is a valid input (corresponding
to no change points, so a singular time series model), and the current 
implementation can reasonably include up to 6 change points. The 
number of change points is used to dictate the segementation of the data  
for each continuous model and each LDA model.}

\item{control}{Class \code{TS_controls} list, holding control parameters
for the Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls, generated by 
\code{\link{TS_controls_list}}.}

\item{i}{Integer iteration index.}

\item{saves}{The existing list of saved data objects.}

\item{steps}{Chain configuration after within-temperature steps.}

\item{swaps}{Chain configuration after among-temperature swaps.}
}
\value{
List of saved ptMCMC objects: changepoints, log-likelihoods, 
  chain ids, step acceptances, and swap acceptances.
}
\description{
\code{prep_saves} creates the data structure used to save the 
  output from each iteration of the ptMCMC algorithm, which is added via
  \code{update_saves}. Once the ptMCMC is complete, the saved data objects
  are then processed (burnin iterations are dropped and the remaining
  iterations are thinned) via \code{process_saves}.
}
