% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_methods.R
\name{prep_saves}
\alias{prep_saves}
\alias{update_saves}
\alias{process_saves}
\title{Prepare and update the data structures to save the LDATS classic 
  ptMCMC output}
\usage{
prep_saves(TS)

update_saves(i, saves, steps, swaps)

process_saves(saves, TS)
}
\arguments{
\item{TS}{\code{list} time series model object.}

\item{i}{\code{integer} iteration index.}

\item{saves}{The existing list of saved data objects.}

\item{steps}{Chain configuration after within-temperature steps.}

\item{swaps}{Chain configuration after among-temperature swaps.}
}
\value{
\code{list} of ptMCMC objects: change points (\code{$cpts}), 
  log-likelihoods (\code{$lls}), chain ids (\code{$ids}), step acceptances
  (\code{$step_accepts}), and swap acceptances (\code{$swap_accepts}).
}
\description{
\code{prep_saves} creates the data structure used to save the 
  output from each iteration of the ptMCMC algorithm, which is added via
  \code{update_saves}. Once the ptMCMC is complete, the saved data objects
  are then processed (burn-in iterations are dropped and the remaining
  iterations are thinned) via \code{process_saves}.
  \cr \cr
  This set of functions was designed to work within \code{\link{TS}} and 
  specifically \code{\link{est_changepoints}}. They are still hardcoded to
  do so, but have the capacity to be generalized to work with any
  estimation via ptMCMC with additional coding work.
}
