% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptMCMC.R
\name{prep_temp_sequence}
\alias{prep_temp_sequence}
\title{Prepare the ptMCMC temperature sequence}
\usage{
prep_temp_sequence(control = TS_controls_list())
}
\arguments{
\item{control}{Class \code{TS_controls} list, holding control parameters
for the Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls, generated by 
\code{\link{TS_controls_list}}.}
}
\value{
Vector of temperatures.
}
\description{
Create the series of temperatures used in the ptMCMC 
  algorithm.
  \cr \cr
  This function was designed to work within \code{\link{TS}} and 
  \code{\link{est_changepoints}} specifically, but has been generalized
  and would work with any ptMCMC model as long as \code{control}
  includes the relevant control parameters (and provided that the 
  \code{\link{check_control}} function and its use here are generalized).
}
