% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_methods.R
\name{prep_temp_sequence}
\alias{prep_temp_sequence}
\title{Prepare the ptMCMC temperature sequence for the LDATS classic 
 method}
\usage{
prep_temp_sequence(TS)
}
\arguments{
\item{TS}{\code{list} time series model object.}
}
\value{
\code{vector} of temperatures.
}
\description{
Create the series of temperatures used in the ptMCMC 
  algorithm.
  \cr \cr
  This function was designed to work within \code{\link{TS}} and 
  \code{\link{est_changepoints}} specifically, but has been generalized
  and would work with any ptMCMC model as long as \code{control}
  includes the relevant control parameters.
}
\examples{
  prep_temp_sequence()

}
