% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptMCMC.R
\name{prep_temp_sequence}
\alias{prep_temp_sequence}
\title{Prepare the ptMCMC temperature sequence}
\usage{
prep_temp_sequence(control = list())
}
\arguments{
\item{control}{A \code{list} of parameters to control the fitting of the
Time Series model including the parallel tempering Markov Chain
Monte Carlo (ptMCMC) controls. Values not input assume defaults set by
\code{\link{TS_control}}.}
}
\value{
\code{vector} of temperatures.
}
\description{
Create the series of temperatures used in the ptMCMC
algorithm.
\cr \cr
This function was designed to work within \code{\link{TS}} and
\code{\link{est_changepoints}} specifically, but has been generalized
and would work with any ptMCMC model as long as \code{control}
includes the relevant control parameters (and provided that the
\code{\link{check_control}} function and its use here are generalized).
}
\examples{
  prep_temp_sequence()

}
