% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{prep_temps}
\alias{prep_temps}
\title{Prepare the temperatures for the MTS algorithm}
\usage{
prep_temps(ntemps = 6, penultimate_temp = 2^6, k = 0, ...)
}
\arguments{
\item{ntemps}{number of temperatures}

\item{penultimate_temp}{penultimate temperature}

\item{k}{the exponent controlling the temperature sequence: 0 implies 
geometric sequence, 1 implies squaring before exponentiating. Use larger 
values if the cooler chains aren't swapping enough.}
}
\value{
temperatures
}
