% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{print.TS_fit}
\alias{print.TS_fit}
\title{Print a Time Series model fit}
\usage{
\method{print}{TS_fit}(x, ...)
}
\arguments{
\item{x}{Class \code{TS_fit} object to be printed.}

\item{...}{Not used, simply included to maintain method compatibility.}
}
\description{
Convenience function to print only the most important 
  components of a \code{TS_fit}-class object fit by 
  \code{\link{TS}}.
}
