% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_on_LDA.R
\name{print.TS_on_LDA}
\alias{print.TS_on_LDA}
\title{Print a set of Time Series models fit to LDAs}
\usage{
\method{print}{TS_on_LDA}(x, ...)
}
\arguments{
\item{x}{Class \code{TS_on_LDA} object to be printed.}

\item{...}{Not used, simply included to maintain method compatability.}
}
\description{
Convenience function to print only the names of a 
  \code{TS_on_LDA}-class object generated by \code{\link{TS_on_LDA}}.
}
