% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_on_LDA.R
\name{print_model_run_message}
\alias{print_model_run_message}
\title{Print the message to the console about which combination of TS
  and LDA is being run}
\usage{
print_model_run_message(models, i, LDA_models, control)
}
\arguments{
\item{models}{\code{data.frame} object returned from 
\code{\link{expand_TS}} that contains the combinations of LDA models, 
and formulas and nchangepoints used in the TS models.}

\item{i}{\code{integer} index of the row to use from \code{models}.}

\item{LDA_models}{List of LDA models (class \code{LDA_set}) or a singular
LDA model (class \code{LDA}).}

\item{control}{Class \code{TS_controls} list, holding control parameters
for the Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls, generated by 
\code{\link{TS_controls_list}}. Of particular importance here is 
the \code{logical}-class element named \code{quiet}.}
}
\value{
Nothing (message is printed, not returned).
}
\description{
If desired, print a message at the beginning of every model
  combination stating the TS model and the LDA model being evaluated.
}
