% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_plots.R
\name{rho_lines}
\alias{rho_lines}
\title{Add changepoint location lines to the time series plot}
\usage{
rho_lines(spec_rhos)
}
\arguments{
\item{spec_rhos}{\code{numeric} vector indicating the locations along the
x axis where the specific changepoints being used are located.}
}
\value{
Nothing. Lines are added to the active plot object.
}
\description{
Adds vertical lines to the fitted gamma time series plot 
  associated with the changepoints of interest.
}
