% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{select_LDA}
\alias{select_LDA}
\title{Select the best LDA model(s) for use in time series}
\usage{
select_LDA(lda_models = NULL, measurer = AIC, selector = min)
}
\arguments{
\item{lda_models}{An object of class \code{LDA_list} produced by
\code{LDA_set}.}

\item{measurer, }{selector Function names for use in evaluation of the LDA
models. \code{measurer} is used to create a value for each model
and \code{selector} operates on the values to choose the model(s) to 
pass on.}
}
\value{
A reduced version of \code{lda_models} that only includes the 
  selected LDA model(s). The returned object is still an object of
  class \code{LDA_list}.

\dontrun{
  data(rodents)
  lda_data <- dplyr::select(rodents, -c(newmoon, date, plots, traps))
  r_LDA <- LDA_set(MV = lda_data, topics = 2, nseeds = 2)  
  select_LDA(r_LDA)                       
}
}
\description{
Select the best model(s) of interest from an
  \code{LDA_list} object, based on a set of user-provided functions. The
  functions default to choosing the model with the lowest AIC value.
}
