% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{select_LDA}
\alias{select_LDA}
\title{Select the best LDA model(s) for use in time series}
\usage{
select_LDA(LDA_models = NULL, measurer = AIC, selector = min)
}
\arguments{
\item{LDA_models}{An object of class \code{LDA_set} produced by
\code{LDA_set}.}

\item{measurer, selector}{Function names for use in evaluation of the LDA
models. \code{measurer} is used to create a value for each model
and \code{selector} operates on the values to choose the model(s) to 
pass on.}
}
\value{
A reduced version of \code{LDA_models} that only includes the 
  selected LDA model(s). The returned object is still an object of
  class \code{LDA_set}.
}
\description{
Select the best model(s) of interest from an
  \code{LDA_set} object, based on a set of user-provided functions. The
  functions default to choosing the model with the lowest AIC value.
}
\examples{
\dontrun{
  data(rodents)
  lda_data <- rodents$document_term_table
  r_LDA <- LDA_set(lda_data, topics = 2, nseeds = 2)  
  select_LDA(r_LDA)                       
}

}
