% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_on_LDA.R
\name{select_TS}
\alias{select_TS}
\title{Select the best Time Series model}
\usage{
select_TS(TS_models, measurer = AIC, selector = min)
}
\arguments{
\item{TS_models}{An object of class \code{TS_on_LDA} produced by
\code{TS_on_LDA}.}

\item{measurer, selector}{Function names for use in evaluation of the TS
models. \code{measurer} is used to create a value for each model
and \code{selector} operates on the values to choose the model.}
}
\value{
A reduced version of \code{TS_models} that only includes the 
  selected TS model. The returned object is still an object of
  class \code{TS_fit}.
}
\description{
Select the best model of interest from an
  \code{TS_on_LDA} object, based on a set of user-provided functions. The
  functions default to choosing the model with the lowest AIC value.
}
