% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_models.R
\name{sequential_TS}
\alias{sequential_TS}
\alias{package_sequential_TS}
\alias{est_changepoints}
\alias{est_regressors}
\alias{sequential_TS_control}
\alias{summarize_etas}
\alias{measure_eta_vcov}
\alias{summarize_rhos}
\alias{measure_rho_vcov}
\title{Estimate a Time Series model sequentially}
\usage{
sequential_TS(TS, control = list())

package_sequential_TS(TS, rho_dist, eta_dist)

est_changepoints(TS)

est_regressors(rho_dist, TS)

sequential_TS_control(method = ldats_classic,
  method_args = ldats_classic_control(), summary_prob = 0.95,
  soften = TRUE, quiet = FALSE, ...)

summarize_etas(etas, TS)

measure_eta_vcov(etas)

summarize_rhos(rhos, TS)

measure_rho_vcov(rhos)
}
\arguments{
\item{TS}{Time series model \code{list}.}

\item{control}{A \code{list} of parameters to control the fitting of the
Time Series model. Values not input assume defaults set by 
\code{\link{sequential_TS_control}}.}

\item{rho_dist}{\code{list} of saved data objects from the estimation of
change point locations (unless \code{nchangepoints} is 0, then 
\code{NULL}) returned from \code{\link{est_changepoints}}.}

\item{eta_dist}{\code{matrix} of draws (rows) from the marginal posteriors
of the coefficients across the segments (columns), as estimated by
\code{\link{est_regressors}}.}

\item{method}{\code{function} used to drive the sampler of the TS
models; \code{method} defines and operates the computational procedure.
\cr \cr
Current pre-built options include \code{\link{ldats_classic}}.}

\item{method_args}{\code{list} of (named) arguments to be used in 
\code{method} via \code{\link{do.call}}. 
\cr \cr
Could be managed via a \code{<method>_control} function like
\code{\link{ldats_classic_control}}.}

\item{summary_prob}{Probability used for summarizing the posterior 
distributions (via the highest posterior density interval, see
\code{\link[coda]{HPDinterval}}).}

\item{soften}{\code{logical} indicator of whether the model should error 
softly or if errors should trigger a full-stop to the pipeline.}

\item{quiet}{\code{logical} indicator of whether the model should run 
quietly (if \code{FALSE}, a progress bar and notifications are printed).}

\item{...}{Not passed along to the output, rather included to allow for
automated removal of unneeded controls.}

\item{etas}{\code{matrix} of regressors (columns) across iterations of the 
sampler (rows), as returned from \code{\link{est_regressors}}.}

\item{rhos}{\code{matrix} of change point locations (columns) across
iterations of the sampler (rows) or \code{NULL} if no change points are 
in the model, as returned from \code{\link{est_change points}}.}

\item{response}{\code{character} element indicating the response variable 
used in the time series. \cr \cr
Must have a corresponding \code{<response>_TS} function.}
}
\value{
\code{sequential_TS} and \code{package_sequential_TS}: 
    \code{TS}-class list containing the following elements, many of
     which are hidden for \code{print}ing, but are accessible:
    \describe{
      \item{data}{\code{data} input to the function.}
      \item{formula}{\code{\link[stats]{formula}} input to the function.}
      \item{nchangepoints}{\code{nchangepoints} input to the function.}
      \item{weights}{\code{weights} input to the function.}
      \item{timename}{\code{timename} input to the function.}
      \item{control}{\code{control} input to the function.}
      \item{lls}{Iteration-by-iteration 
                 \link[=logLik.multinom_TS_fit]{logLik} values for the
                  full time series fit by \code{\link{multinom_TS}}.}
      \item{rhos}{Iteration-by-iteration change point estimates from
                  \code{\link{est_changepoints}} and diagnostics.}
      \item{focal_rhos}{Simplified object of just the change point 
                        locations of interest.}
      \item{etas}{Iteration-by-iteration marginal regressor estimates from
                  \code{\link{est_regressors}}, which have been 
                  unconditioned with respect to change point locations.}
      \item{rho_summary}{Summary table describing \code{rhos} (the change
                         point locations), see 
                         \code{\link{summarize_rhos}}.}
      \item{rho_vcov}{Variance-covariance matrix for the estimates of
                     \code{rhos} (the change point locations), see 
                     \code{\link{measure_rho_vcov}}.}
      \item{eta_summary}{Summary table describing \code{ets} (the 
                         regressors), see 
                         \code{\link{summarize_etas}}.}
      \item{eta_vcov}{Variance-covariance matrix for the estimates of
                     \code{etas} (the regressors), see 
                     \code{\link{measure_eta_vcov}}.}
      \item{logLik}{Across-iteration average of log-likelihoods 
                   (\code{lls}).}
      \item{nparams}{Total number of parameters in the full model,
                     including the change point locations and regressors.}
    } \cr \cr
  \code{sequential_TS_control}: \code{list} of named control elements for 
    sequential model fitting. 
  \code{est_changepoints}: \code{list} of saved data objects from the 
    estimation of change point locations, uunless \code{nchangepoints} 
    is 0, then \code{NULL}. \cr \cr
  \code{est_regressors}: \code{matrix} of draws (rows) from the marginal 
    posteriors of the coefficients across the segments (columns). \cr \cr
  \code{summarize_etas}: table of summary statistics for chunk-level
    regressors including mean, median, mode, posterior interval, standard
    deviation, MCMC error, autocorrelation, and effective sample size for 
    each regressor. \cr \cr
  \code{measure_eta_vcov}: variance-covariance matrix for chunk-level
    regressors. \cr \cr
 \code{summarize_rhos}: table of summary statistics for change point
    locations including mean, median, mode, posterior interval, standard
    deviation, MCMC error, autocorrelation, and effective sample size for 
    each change point location. \cr \cr
  \code{measure_rho_vcov}: variance-covariance matrix for change 
    point locations.
}
\description{
This set of functions estimates the Time Series model
  by sequential methods that first estimate the change point locations
  with full flexibility of the regressor models between change points,
  then estimate the regressors between the change points, intially 
  conditional on their locations, but with marginal estimation to produce
  regressor values unconditional on change point locations. \cr \cr
  \code{sequential_TS} combines each stage of the model estimation and 
    packages the model results in a consistent output. \cr \cr
  \code{sequential_TS_control} defines and creates the control \code{list}
    used to fit the sequential Time Series model.
  \code{est_changepoints} estimates the change point location 
    distributions for multinomial Time Series analyses. \cr \cr
  \code{est_regressors} uses the marginal posterior distributions of
    the change point locations (estimated by 
    \code{\link{est_changepoints}}) in combination with the conditional 
    (on the change point locations) posterior distributions of the 
    regressors (estimated by a \code{<response>_TS} function) to 
    estimate the marginal posterior distribution of the regressors, 
    unconditional on the change point locations. \cr \cr
  \code{package_sequential_TS} calculates relevant summaries for the run of
    a sequenial Time Series model within \code{\link{sequential_TS}} and 
    packages the output as a \code{TS}-class object. \cr \cr
  \code{summarize_etas} calculates summary statistics for each
    of the chunk-level regressors. \cr \cr
  \code{measure_ets_vcov} generates the variance-covariance matrix for 
    the regressors. \cr \cr
 \code{summarize_rho} calculates summary statistics for each
    of the change point locations. \cr \cr
  \code{measure_rho_vcov} generates the variance-covariance matrix for the 
    change point locations.
}
\details{
The general approach follows that of Western and Kleykamp
  (2004), although we note some important differences. Our regression
  models are fit independently for each chunk (segment of time), and 
  therefore the variance-covariance matrix for the full model 
  has \code{0} entries for covariances between regressors in different
  chunks of the time series. \cr \cr
  Further differences are model-specific. For example, the original softmax 
  multinomial regression model used here is a standard (non-hierarchical) 
  model (Ripley 1996, Venables and Ripley 2002, Bishop 2006), mean that
  there is no error term in the regression (as there is in the normal
  model used by Western and Kleykamp 2004), and so the posterior 
  distribution used here is a multivariate normal, as opposed to a
  multivariate t, as used by Western and Kleykamp (2004).
}
\references{
Bishop, C. M. 2006. \emph{Pattern Recognition and Machine Learning}. 
   Springer, New York, NY, USA.

  Ripley, B. D. 1996. \emph{Pattern Recognition and Neural Networks}. 
  Cambridge University Press, Cambridge, UK.

  Venables, W. N. and B. D. Ripley. 2002. \emph{Modern and Applied
  Statistics with S}. Fourth Edition. Springer, New York, NY, USA.

  Western, B. and M. Kleykamp. 2004. A Bayesian change point model for 
  historical time series analysis. \emph{Political Analysis}
  \strong{12}:354-374.
  \href{https://doi.org/10.1093/pan/mph023}{link}.
}
