% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_plots.R
\name{set_gamma_colors}
\alias{set_gamma_colors}
\title{Prepare the colors to be used in the gamma time series}
\usage{
set_gamma_colors(x, cols = NULL, option = "D", alpha = 1)
}
\arguments{
\item{x}{Object of class \code{TS_fit}.}

\item{cols}{Colors to be used to plot the time series of fitted topic 
proportions (gammas).}

\item{option}{A character string indicating the colormap option to use if 
`cols == NULL`. Four options are available: "magma" (or "A"), "inferno" 
(or "B"), "plasma" (or "C"), "viridis" (or "D", the default option) and 
"cividis" (or "E").}

\item{alpha}{Numeric value [0,1] that indicates the transparency of the 
colors used. Supported only on some devices, see \code{rgb}.}
}
\value{
Vector of \code{character} hex codes indicating colors to use.
}
\description{
Based on the inputs, create the set of colors to be used in
  the timeseries of the fitted gamma (topic proportion) values.
}
