% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{TS}
\alias{TS}
\alias{prep_TS_models}
\title{Conduct a Bayesian compositional Time Series analysis}
\usage{
TS(LDAs, data, formulas = ~1, nchangepoints = 0, timename = "time",
  weights = NULL, control = list())

prep_TS_models(LDAs, data, formulas = ~1, nchangepoints = 0,
  timename = "time", weights = NULL, control = list())
}
\arguments{
\item{data}{\code{data.frame} including [1] the time variable (indicated 
in \code{timename}), [2] the predictor variables (required by
\code{formula}) and [3], the compositional response variable (indicated
in \code{formula}). \cr \cr
Note that the response variables should be formatted as a 
\code{data.frame} object named as indicated by the \code{response} entry
in the \code{control} list, such as \code{gamma} for a standard TS 
analysis on LDA output.}

\item{formulas}{Vector of \code{\link[stats]{formula}}(s) defining the 
regression between the change points. Any predictor variable included 
must also be a column in \code{data} and any (compositional) response 
variable must be a set of columns in \code{data}. \cr
Each element (formula) in the vector is evaluated for each number of 
change points and each LDA model.}

\item{nchangepoints}{\code{integer}-conformable vector corresponding to the 
number of change points to include in the models. 0 is valid (corresponds
to no change points, so a singular time series model) and the current 
implementation can reasonably include up to 6 change points. The 
number of change points is used to dictate the segmentation of the 
time series into chunks fit with separate models dictated by 
\code{formula}. \cr
Each element in the vector is the number of change points 
used to segment the data for each formula (entry in \code{formulas}) 
component of the TS model, for each selected LDA model.}

\item{timename}{\code{character} element indicating the time variable
used in the time series. Defaults to \code{"time"}. The variable must be
integer-conformable or a \code{Date}. If the variable named
is a \code{Date}, the input is converted to an integer, resulting in the
timestep being 1 day, which is often not desired behavior.}

\item{weights}{Optional class \code{numeric} vector of weights for each 
document. Defaults to \code{NULL}, translating to an equal weight for
each document. When using \code{\link{sequential_TS}} in a standard LDATS 
analysis, it is advisable to weight the documents by their total size,
as the result of, e.g., \code{\link[topicmodels]{LDA}} is a matrix of 
proportions, which does not account for size differences among documents.
For most models, a scaling of the weights (so that the average is 1) is
most appropriate, and this is accomplished using \code{document_weights}.}

\item{control}{A \code{list} of parameters to control the fitting of the
Time Series model. Values not input assume defaults set by 
\code{\link{TS_control}}.}
}
\value{
\code{TS}: class \code{TS_set} \code{list} of both selected and all
    results from \code{\link{sequential_TS}} applied for each model on each
    LDA model input as well as the control \code{list} used to fit the 
    model. \cr \cr
  \code{prep_TS_models}: \code{list} of \code{list}s, each of which is a
    preliminary model object for a Time Series model fit.
}
\description{
Analyze compositional Time Series models. \cr \cr
  \code{TS} is the main interface function for the LDATS application
    of Bayesian change point Time Series analyses (Christensen 
    \emph{et al.} 2018). \cr \cr
  \code{prep_TS_models} pre-prepares the TS model objects for simpler 
    use within the subfunctions.
}
\details{
For a (potentially subset) dataset consisting of proportions of
  topics across multiple documents in a corpus 
  \enumerate{
    \item Conduct multiple compositional Bayesian TS models 
      (e.g., changepoint softmax regression; Ripley 1996, Venables 
      and Ripley 2002, Western and Kleykamp 2004, Bishop 2006, Ruggieri 
      2013) via a generalized linear modeling approach (McCullagh and 
      Nelder 1989) and using parallel tempering Markov Chain Monte Carlo
     (ptMCMC) methods (Earl and Deem 2005),
    \item Select from the TS model results to pick those used to summarize
      the whole model, and
    \item Package the results.
  }
}
