% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_plots.R
\name{TS_diagnostics_plot}
\alias{TS_diagnostics_plot}
\alias{eta_diagnostics_plots}
\alias{rho_diagnostics_plots}
\title{Plot the diagnostics of the parameters fit in a TS model}
\usage{
TS_diagnostics_plot(x)

eta_diagnostics_plots(x)

rho_diagnostics_plots(x)
}
\arguments{
\item{x}{Object of class \code{TS_fit} to have its diagnostics plotted.}
}
\value{
Nothing. Plots sent to active graphics device.
}
\description{
Plot 4-panel figures (showing trace plots, posterior ECDF, 
  posterior density, and iteration autocorrelation) for each of the 
  parameters (changepoint locations and regressors) fitted within a 
  multinomial time series model (fit by \code{multinom_TS})
}
