% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_plots.R
\name{TS_summary_plot}
\alias{TS_summary_plot}
\title{Create the summary plot for a TS fit to an LDA model}
\usage{
TS_summary_plot(x, bin_width, xlab, selection = "median",
  cols = TS_summary_cols(), split_need = TRUE)
}
\arguments{
\item{x}{Object of class \code{TS_fit}.}

\item{bin_width}{Width of the bins used in the histograms, in units of the
x-axis (the time variable used to fit the model).}

\item{xlab}{Label for the x-axis.}

\item{selection}{Indicator of the changepoints to use. Currently only
defined for \code{"median"}.}

\item{cols}{\code{list} of elements used to define the colors for the two
panels, as generated simply using \code{TS_summary_cols}. \code{cols}
has two elements \code{rho} and \code{gamma}, each corresponding to the
related panel, and each containing default values for entries named
\code{cols}, \code{option}, and \code{alpha}. See
\code{set_gamma_colors} and \code{set_rho_hist_colors} for
details on usage.}
}
\value{
Nothing. The plot is generated in the active graphics device.
}
\description{
Produces a two-panel figure of [1] the changepoint 
  distributions as histograms over time and [2] the time series of the 
  fitted topic proportions (gamma values) over time, based on a selected 
  set of changepoint locations.
}
